svsm::acpi::tables

Function load_fw_cpu_info

pub fn load_fw_cpu_info(
    fw_cfg: &FwCfg<'_>,
) -> Result<Vec<ACPICPUInfo>, SvsmError>

Loads ACPI CPU information by parsing the ACPI tables provided by the firmware It processes the Multiple APIC Description Table (MADT) to extract information about each CPU’s APIC ID and enabled status.

Arguments

fw_cfg: A reference to the Firmware Configuration (FwCfg) interface for accessing ACPI tables.

Returns

A Result containing a vector of ACPICPUInfo structs representing CPU information. If successful, the vector contains information about each detected CPU; otherwise, an error is returned.

Errors

This function returns an error if there are issues with reading or parsing ACPI tables, or if the required ACPI tables are not found.

Example

use svsm::acpi::tables::load_fw_cpu_info;
use svsm::fw_cfg::FwCfg;
use svsm::io::IOPort;

#[derive(Debug)]
struct MyIo;

impl IOPort for MyIo {
    // your implementation
}

let io = MyIo;
let fw_cfg = FwCfg::new(&io);
match load_fw_cpu_info(&fw_cfg) {
    Ok(cpu_info) => {
        for info in cpu_info {
            // You can print id (info.apic_id) and whether it is enabled (info.enabled)
        }
    }
    Err(err) => {
        // Print error
    }
}