svsm::async

Function task_busypoll_to_completion

pub fn task_busypoll_to_completion<O, T: FnMut(&mut Context<'_>) -> Poll<O>>(
    task: T,
) -> O

Busypoll an asynchronous task to completion.

The task to poll on is represented as a closure, task. Note that this scheme enables convenient polling on standard Rust Futures, as well as on Future-like types taking additional arguments for their poll().

Arguments:

task -> The task to poll to completion.