svsm

Module attest

Structs

AttestationDriver

The attestation driver that communicates with the proxy via some communication channel (serial port, virtio-vsock, etc…).

Enums

AttestationError

Possible errors when attesting TEE evidence.

Functions

evidence 🔒

Hash negotiation parameters and fetch TEE evidence.

hash 🔒

Hash the negotiation parameters from the attestation server for inclusion in the attestation evidence.

sc_key_generate 🔒

Generate a key used to establish a secure channel between the confidential guest and attestation server.