Struct svsm::cpu::apic::LocalApic

pub struct LocalApic {
    irr: [u32; 8],
    allowed_irr: [u32; 8],
    isr_stack_index: usize,
    isr_stack: [u8; 16],
    tmr: [u32; 8],
    host_tmr: [u32; 8],
    update_required: bool,
    interrupt_delivered: bool,
    interrupt_queued: bool,
    lazy_eoi_pending: bool,
    nmi_pending: bool,
}

Fields
irr: [u32; 8]allowed_irr: [u32; 8]isr_stack_index: usizeisr_stack: [u8; 16]tmr: [u32; 8]host_tmr: [u32; 8]update_required: boolinterrupt_delivered: boolinterrupt_queued: boollazy_eoi_pending: boolnmi_pending: bool
Implementations

impl LocalApic

pub const fn new() -> Self

fn scan_irr(&self) -> u8

fn remove_vector_register(register: &mut [u32; 8], irq: u8)

fn insert_vector_register(register: &mut [u32; 8], irq: u8)

fn test_vector_register(register: &[u32; 8], irq: u8) -> bool

fn rewind_pending_interrupt(&mut self, irq: u8)

pub fn check_delivered_interrupts<T: GuestCpuState>( &mut self, cpu_state: &mut T, caa_addr: Option<VirtAddr>, )

fn get_ppr_with_tpr(&self, tpr: u8) -> u8

fn get_ppr<T: GuestCpuState>(&self, cpu_state: &T) -> u8

fn clear_guest_eoi_pending( caa_addr: Option<VirtAddr>, ) -> Option<GuestPtr<SvsmCaa>>

fn deliver_interrupt_immediately<T: GuestCpuState>( &self, irq: u8, cpu_state: &mut T, ) -> bool

Attempts to deliver the specified IRQ into the specified guest CPU so that it will be immediately observed upon guest entry. Returns true if the interrupt request was delivered, or false if the guest cannot immediately receive an interrupt.

fn consume_pending_ipis(&mut self, cpu_shared: &PerCpuShared)

pub fn present_interrupts<T: GuestCpuState>( &mut self, cpu_shared: &PerCpuShared, cpu_state: &mut T, caa_addr: Option<VirtAddr>, )

fn perform_host_eoi(vector: u8)

fn perform_eoi(&mut self)

fn get_isr(&self, index: usize) -> u32

fn post_interrupt(&mut self, irq: u8, level_sensitive: bool)

fn post_icr_interrupt(&mut self, icr: ApicIcr)

fn post_ipi_one_target(cpu: &PerCpuShared, icr: ApicIcr)

fn send_logical_ipi(&mut self, icr: ApicIcr) -> bool

Sends an IPI using the APIC logical destination mode. Returns true if the host needs to be notified.

fn logical_destination_match(destination: u32, apic_id: u32) -> bool

Returns true if the specified APIC ID matches the given logical destination.

fn send_physical_ipi(&mut self, icr: ApicIcr) -> bool

Send an IPI using the APIC physical destination mode. Returns true if the host needs to be notified.

fn send_ipi(&mut self, icr: ApicIcr)

Sends an IPI using the specified ICR.

pub fn read_register<T: GuestCpuState>( &mut self, cpu_shared: &PerCpuShared, cpu_state: &mut T, caa_addr: Option<VirtAddr>, register: u64, ) -> Result<u64, SvsmError>

Reads an APIC register, returning its value, or an error if an invalid register is requested.

fn handle_icr_write(&mut self, value: u64) -> Result<(), SvsmError>

pub fn write_register<T: GuestCpuState>( &mut self, cpu_state: &mut T, caa_addr: Option<VirtAddr>, register: u64, value: u64, ) -> Result<(), SvsmError>

Writes a value to the specified APIC register. Returns an error if an invalid register or value is specified.

pub fn configure_vector(&mut self, vector: u8, allowed: bool)

fn signal_one_host_interrupt( &mut self, vector: u8, level_sensitive: bool, ) -> bool

fn signal_several_interrupts(&mut self, group: usize, bits: u32)

fn consume_host_interrupts(&mut self)

fn handoff_to_host(&mut self)

pub fn disable_apic_emulation<T: GuestCpuState>( &mut self, cpu_state: &mut T, caa_addr: Option<VirtAddr>, )