svsm::cpu::features

Struct CpuFeat

struct CpuFeat {
    val: ImmutAfterInitCell<u32>,
    leaf: u32,
    subleaf: u32,
    reg: CpuidReg,
    shift: u8,
    bitsize: u8,
    expected: u32,
}

A discoverable CPU feature.

At the moment this type is used for CPUID detection, but it can be expanded to use other sources of information.

Fields
val: ImmutAfterInitCell<u32>

Raw value

leaf: u32

CPUID leaf

subleaf: u32

CPUID subleaf

reg: CpuidReg

CPUID output register

shift: u8

Bitshift to apply to raw value

bitsize: u8

Bitmask to apply to raw value

expected: u32

Expected value after shift + mask

Implementations

impl CpuFeat

const fn new_bit(leaf: u32, reg: CpuidReg, bit: u8) -> Self

Create a new feature, indicated by a single bit in the specified register in the given CPUID leaf.

const fn new_u32(leaf: u32, reg: CpuidReg, expected: u32) -> Self

Create a new feature, indicated by the full value of a register in the given CPUID leaf.

const fn new( leaf: u32, reg: CpuidReg, shift: u8, bitsize: u8, expected: u32, ) -> Self

Create a new CPU feature, detected by querying the given CPUID leaf, and applying a bitshift and bitmask on the specified output register to compare it to the given expected value.

const fn with_subfn(self, subleaf: u32) -> Self

Create a copy of the given CPU feature, but with the specified subleaf.

const fn get_reg(&self, cpuid: &CpuidResult) -> u32

Get the CPUID register that corresponds to this feature

const fn mask(&self) -> u32

fn get_or_init(&self) -> u32

fn get(&self) -> u32

Gets the raw value of this feature, lazily querying CPUID if the value is not cached from a previous query

fn enabled(&self) -> bool

Checks whether this feature is available by comparing it to its expected value, lazily querying CPUID if the value is not cached from a previous query