svsm::cpu::idt::common

Struct X86ExceptionContext

#[repr(C, packed(1))]
pub struct X86ExceptionContext {
    pub ssp: usize,
    pub regs: X86GeneralRegs,
    pub error_code: usize,
    pub frame: X86InterruptFrame,
}

Fields
ssp: usizeregs: X86GeneralRegserror_code: usizeframe: X86InterruptFrame
Implementations

impl X86ExceptionContext

pub unsafe fn set_rip(&mut self, new_rip: usize)

Safety

The caller must ensure to update the rest of the execution state as actual hardware would have done it (e.g. for MMIO emulation, CPUID, MSR, etc.).