svsm::cpu::ipi

Function send_multicast_ipi

pub fn send_multicast_ipi<M: IpiMessage + Sync>(
    target_set: IpiTarget<'_>,
    ipi_message: &M,
)

Sends an IPI message to multiple CPUs.

Safety

The IPI message must NOT contain any references to data unless that data is known to be in memory that is visible across CPUs/tasks. Otherwise, the recipient could attempt to access a pointer that is invalid in the target context, or - worse - points to completely incorrect data in the target context.

Arguments

target_set - The set of CPUs to which to send the IPI.

ipi_message - The message to send.