svsm::cpu::ipi
Trait IpiMessage

pub unsafe trait IpiMessage {
    // Required method
    fn invoke(&self);

    // Provided method
    fn copy_to_shared(&self, buffer: &mut [u8])
       where Self: Sized { ... }
}

Safety

This trait implements a method to copy IPI message contents into a shared buffer. If that serialization is performed incorrectly, then IPI message receipt will be unsound because the message may contain incorrect pointers and references that refer to invalid memory - or memory that belongs to another owner. All implementations of this trait must verify that the copy routine correctly copies all data and resolves all references within the copied data.

Required Methods

fn invoke(&self)

Invokes the IPI handler for the message.

Provided Methods

fn copy_to_shared(&self, buffer: &mut [u8])
where Self: Sized,

All IPI messages must be copied into a shared IPI buffer since stack locals are not visible across CPU/task contexts. This function must perform a deep copy of the contents of the source buffer into the shared destination buffer.

Arguments:

*src: A pointer to the input message. *buffer: A byte slice in shared memory which will be the target of the copy.