svsm::cpu
Module irq_state

Structs

IrqGuard

And IRQ guard which saves the current IRQ state and disabled interrupts upon creation. When the guard goes out of scope the previous IRQ state is restored.

IrqState

This structure keeps track of PerCpu IRQ states. It tracks the original IRQ state and how deep IRQ-disable calls have been nested. The use of atomics is necessary for interior mutability and to make state modifications safe wrt. to IRQs.

Constants

EFLAGS_IF

Interrupt flag in RFLAGS register

Functions

irqs_disabled

Query IRQ state on current CPU

irqs_enabled

Query IRQ state on current CPU

raw_irqs_disable

Unconditionally disable IRQs

raw_irqs_enable

Unconditionally enable IRQs