svsm::cpu::irq_state

Struct IrqState

pub struct IrqState {
    state: AtomicBool,
    counts: [AtomicI32; 16],
    phantom: PhantomData<*const ()>,
}

This structure keeps track of PerCpu IRQ states. It tracks the original IRQ state and how deep IRQ-disable calls have been nested. The use of atomics is necessary for interior mutability and to make state modifications safe wrt. to IRQs.

The original state needs to be stored to not accidentially enable IRQs in contexts which have IRQs disabled by other means, e.g. in an exception or NMI/HV context.

Fields
state: AtomicBool

IRQ state when count was 0

counts: [AtomicI32; 16]

Depth of IRQ-disabled nesting. Index 0 specifies the count of IRQ disables and the remaining indices specify the nesting count for eached raised TPR level.

phantom: PhantomData<*const ()>

Make the type !Send + !Sync

Implementations

impl IrqState

pub fn new() -> Self

Create a new instance of IrqState

pub fn push_nesting(&self, was_enabled: bool)

Increase IRQ-disable nesting level by 1. The method will not disable interrupts because it is assumed to be called when interrupts are already disabled.

The caller needs to make sure to match the number of push_nesting calls with the number of pop_nesting calls.

was_enabled - indicates whether interrupts were enabled at the time of nesting. This may not be the current state of interrupts because interrupts may have been disabled for architectural reasons prior to his function being called.

Returns

The previous nesting level.

pub fn disable(&self)

Increase IRQ-disable nesting level by 1. The method will disable IRQs.

The caller needs to make sure to match the number of disable calls with the number of enable calls.

pub fn pop_nesting(&self) -> i32

Decrease IRQ-disable nesting level by 1. The method will not restore the original IRQ state when the nesting level reaches 0.

The caller needs to make sure to match the number of pop_nesting calls with the number of push_nesting calls.

Returns

The new IRQ nesting level.

pub fn enable(&self)

Decrease IRQ-disable nesting level by 1. The method will restore the original IRQ state when the nesting level reaches 0.

The caller needs to make sure to match the number of disable calls with the number of enable calls.

pub fn count(&self) -> i32

Returns the current nesting count

Returns

Levels of IRQ-disable nesting currently active

pub fn set_restore_state(&self, enabled: bool)

Changes whether interrupts will be enabled when the nesting count drops to zero.

The caller must ensure that the current nesting count is non-zero, and must ensure that the specified value is appropriate for the current environment.

pub fn raise_tpr(&self, tpr_value: usize)

Increments TPR.

The caller must ensure that a raise_tpr() call is followed by a matching call to lower_tpr().

tpr_value - The new TPR value. Must be greater than or equal to the current TPR value.

pub fn lower_tpr(&self, tpr_value: usize)

Decrements TPR.

The caller must ensure that a lower call balances a preceding raise call to the indicated level.

tpr_value - The TPR from which the caller would like to lower. Must be less than or equal to the current TPR.