Struct svsm::cpu::percpu::PerCpu

pub struct PerCpu {
    cpu_unsafe: *const PerCpuUnsafe,
    apic_id: u32,
    pgtbl: SpinLock<PageTableRef>,
    tss: X86Tss,
    svsm_vmsa: Option<VmsaRef>,
    reset_ip: u64,
    vm_range: VMR,
    pub vrange_4k: VirtualRange,
    pub vrange_2m: VirtualRange,
    runqueue: RWLock<RunQueue>,
    request_waitqueue: WaitQueue,
}

Fields
cpu_unsafe: *const PerCpuUnsafeapic_id: u32pgtbl: SpinLock<PageTableRef>tss: X86Tsssvsm_vmsa: Option<VmsaRef>reset_ip: u64vm_range: VMR

PerCpu Virtual Memory Range

vrange_4k: VirtualRange

Address allocator for per-cpu 4k temporary mappings

vrange_2m: VirtualRange

Address allocator for per-cpu 2m temporary mappings

runqueue: RWLock<RunQueue>

Task list that has been assigned for scheduling on this CPU

request_waitqueue: WaitQueue

WaitQueue for request processing

Implementations

impl PerCpu

fn new(apic_id: u32, cpu_unsafe: *const PerCpuUnsafe) -> Self

pub fn alloc(apic_id: u32) -> Result<RefMut<'static, PerCpu>, SvsmError>

pub fn cpu_unsafe(&self) -> *const PerCpuUnsafe

fn shared(&self) -> &'static PerCpuShared

pub const fn get_apic_id(&self) -> u32

fn allocate_page_table(&mut self) -> Result<(), SvsmError>

pub fn set_pgtable(&mut self, pgtable: PageTableRef)

fn allocate_stack(&mut self, base: VirtAddr) -> Result<VirtAddr, SvsmError>

fn allocate_init_stack(&mut self) -> Result<(), SvsmError>

fn allocate_ist_stacks(&mut self) -> Result<(), SvsmError>

pub fn get_pgtable(&self) -> LockGuard<'_, PageTableRef>

pub fn setup_ghcb(&mut self) -> Result<(), SvsmError>

pub fn register_ghcb(&self) -> Result<(), SvsmError>

fn setup_tss(&mut self)

pub fn map_self_stage2(&mut self) -> Result<(), SvsmError>

pub fn map_self(&mut self) -> Result<(), SvsmError>

fn initialize_vm_ranges(&mut self) -> Result<(), SvsmError>

fn finish_page_table(&mut self)

pub fn dump_vm_ranges(&self)

pub fn setup(&mut self, platform: &dyn SvsmPlatform) -> Result<(), SvsmError>

pub fn setup_on_cpu( &mut self, platform: &dyn SvsmPlatform ) -> Result<(), SvsmError>

pub fn setup_idle_task( &mut self, entry: extern "C" fn() ) -> Result<(), SvsmError>

pub fn load_pgtable(&mut self)

pub fn load_tss(&mut self)

pub fn load(&mut self)

pub fn shutdown(&mut self) -> Result<(), SvsmError>

pub fn set_reset_ip(&mut self, reset_ip: u64)

pub fn alloc_svsm_vmsa(&mut self) -> Result<(), SvsmError>

pub fn get_svsm_vmsa(&mut self) -> &mut Option<VmsaRef>

pub fn prepare_svsm_vmsa(&mut self, start_rip: u64)

pub fn unmap_guest_vmsa(&self)

pub fn map_guest_vmsa(&self, paddr: PhysAddr) -> Result<(), SvsmError>

pub fn guest_vmsa_ref(&self) -> LockGuard<'_, GuestVmsaRef>

pub fn alloc_guest_vmsa(&mut self) -> Result<(), SvsmError>

pub fn unmap_caa(&self)

pub fn map_guest_caa(&self, paddr: PhysAddr) -> Result<(), SvsmError>

fn vmsa_tr_segment(&self) -> VMSASegment

pub fn virt_range_init(&mut self)

pub fn new_mapping( &mut self, mapping: Arc<Mapping> ) -> Result<VMRMapping<'_>, SvsmError>

Create a new virtual memory mapping in the PerCpu VMR

Arguments

mapping - The mapping to insert into the PerCpu VMR

Returns

On success, a new [‘VMRMapping’} that provides a virtual memory address for the mapping which remains valid until the [‘VRMapping’] is dropped.

On error, an [‘SvsmError’].

pub fn populate_page_table(&self, pt: &mut PageTableRef)

Add the PerCpu virtual range into the provided pagetable

Arguments

pt - The page table to populate the the PerCpu range into

pub fn handle_pf(&self, vaddr: VirtAddr, write: bool) -> Result<(), SvsmError>

pub fn schedule_init(&mut self) -> TaskPointer

pub fn schedule_prepare(&mut self) -> Option<(TaskPointer, TaskPointer)>

pub fn runqueue(&self) -> &RWLock<RunQueue>

pub fn current_task(&self) -> TaskPointer

pub fn set_tss_rsp0(&mut self, addr: VirtAddr)