svsm::cpu::x86
Module apic

Constants

APIC_ENABLE_MASK 🔒

APIC_SPIV_SW_ENABLE_MASK 🔒

APIC_SPIV_VECTOR_MASK 🔒

APIC_X2_ENABLE_MASK 🔒

MSR_APIC_BASE 🔒

MSR_X2APIC_EOI

End-of-Interrupt register MSR offset

MSR_X2APIC_ICR

Interrupt-Control-Register register MSR offset

MSR_X2APIC_ISR

Interrupt-Service-Register base MSR offset

MSR_X2APIC_SPIV

Spurious-Interrupt-Register MSR offset

Functions

apic_register_bit

Get the MSR offset relative to a bitmap base MSR and the mask for the MSR value to check for a specific vector bit being set in IRR, ISR, or TMR.

x2apic_enable

Enables the X2APIC by setting the AE and EXTD bits in the APIC base address register.

x2apic_eoi

Send an End-of-Interrupt notification to the X2APIC.

x2apic_icr_write

Write a command to the Interrupt Command Register.

x2apic_in_service

Check whether a give IRQ vector is currently being serviced by returning the value of its ISR bit from X2APIC.

x2apic_spiv_write

Set Spurious-Interrupt-Vector Register

x2apic_sw_enable

Enable the APIC-Software-Enable bit.