Trait svsm::crypto::aead::Aes256GcmTrait

pub trait Aes256GcmTrait {
    // Required methods
    fn encrypt(
        iv: &[u8; 12],
        key: &[u8; 32],
        aad: &[u8],
        inbuf: &[u8],
        outbuf: &mut [u8],
    ) -> Result<usize, SvsmReqError>;
    fn decrypt(
        iv: &[u8; 12],
        key: &[u8; 32],
        aad: &[u8],
        inbuf: &[u8],
        outbuf: &mut [u8],
    ) -> Result<usize, SvsmReqError>;
}

AES-256 GCM

Required Methods

fn encrypt( iv: &[u8; 12], key: &[u8; 32], aad: &[u8], inbuf: &[u8], outbuf: &mut [u8], ) -> Result<usize, SvsmReqError>

Encrypt the provided buffer using AES-256 GCM

Arguments

iv: Initialization vector

key: 256-bit key

aad: Additional authenticated data

inbuf: Cleartext buffer to be encrypted

outbuf: Buffer to store the encrypted data, it must be large enough to also hold the authenticated tag.

Returns

Success

usize: Number of bytes written to outbuf

Error

SvsmReqError

fn decrypt( iv: &[u8; 12], key: &[u8; 32], aad: &[u8], inbuf: &[u8], outbuf: &mut [u8], ) -> Result<usize, SvsmReqError>

Decrypt the provided buffer using AES-256 GCM

Returns

iv: Initialization vector

key: 256-bit key

aad: Additional authenticated data

inbuf: Cleartext buffer to be decrypted, followed by the authenticated tag

outbuf: Buffer to store the decrypted data

Returns

Success

usize: Number of bytes written to outbuf

Error

SvsmReqError

Object Safety

This trait is not object safe.