svsm::debug

Module symbols

Structs

ResolvedSym

A resolved symbol name and offset for any given address.

SymResolver 🔒

An address to symbol resolver, using ELF symbol information.

Statics

SYM_RESOLVER 🔒

Functions

init_symbols

Initializes symbol data so that it can be used when resolve_symbol() is called.

resolve_symbol

Attempts to locate the symbol that corresponds to the given address.