svsm

Module error

High level error typing for the public SVSM APIs.

This module contains the generic SvsmError type, which may be returned from any public API in this codebase to signal an error during SVSM operation. Each variant of the type may give more specific information about the source of the error.

As a general rule, functions private to a given module may directly return leaf error types, which are contained in SvsmError variants. Public functions should return an SvsmError containing a leaf error type, usually the one corresponding to that module. Each module should provide a way to convert a leaf error into a SvsmError via the From trait.

Enums

ApicError

Errors related to APIC handling. These may originate from multiple layers in the system.

AttestError

Errors related to Attestation handling. These may originate from multiple layers in the system. Added to protocol base for protocol error.

SvsmError

A generic error during SVSM operation.