svsm::fs
Module api

Macros

impl_fs_err 🔒

Used to define methods of FsError.

Structs

DirectoryEntry

Directory entries including their names.

Enums

DirEntry

Represents a directory entry which could either be a file or a subdirectory.

FsError

Represents the type of error occured while doing SVSM filesystem operations.

Constants

MAX_FILENAME_LENGTH 🔒

Maximum supported length for a single filename

Traits

Directory

Represents directory operations

File

Represents file operations

Type Aliases

FileName

FileNameArray