Trait svsm::fs::Directory

pub trait Directory: Debug + Send + Sync {
    // Required methods
    fn list(&self) -> Vec<FileName>;
    fn lookup_entry(&self, name: FileName) -> Result<DirEntry, SvsmError>;
    fn create_file(&self, name: FileName) -> Result<Arc<dyn File>, SvsmError>;
    fn create_directory(
        &self,
        name: FileName,
    ) -> Result<Arc<dyn Directory>, SvsmError>;
    fn unlink(&self, name: FileName) -> Result<(), SvsmError>;
}

Represents directory operations

Required Methods

fn list(&self) -> Vec<FileName>

Used to get the list of entries in the directory.

Returns

A Vec<FileName> containing all the entries in the directory.

fn lookup_entry(&self, name: FileName) -> Result<DirEntry, SvsmError>

Used to lookup for an entry in the directory.

Arguments

name: name of the entry to be looked up in the directory.

Returns

Result<DirEntry, SvsmError>: A Result containing the DirEntry corresponding to the entry being looked up in the directory if present, or an SvsmError if not present.

fn create_file(&self, name: FileName) -> Result<Arc<dyn File>, SvsmError>

Used to create a new file in the directory.

Arguments

name: name of the file to be created.

Returns

Result<DirEntry, SvsmError>: A Result containing the DirEntry of the new file created on success, or an SvsmError on failure

fn create_directory( &self, name: FileName, ) -> Result<Arc<dyn Directory>, SvsmError>

Used to create a subdirectory in the directory.

Arguments

name: name of the subdirectory to be created.

Returns

Result<DirEntry, SvsmError>: A Result containing the DirEntry of the subdirectory created on success, or an SvsmError on failure

fn unlink(&self, name: FileName) -> Result<(), SvsmError>

Used to remove an entry from the directory.

Arguments

name: name of the entry to be removed from the directory.

Returns

[Result<(), SvsmError>]: A Result containing the empty value on success, or an SvsmError on failure