Trait svsm::fs::File

pub trait File: Debug + Send + Sync {
    // Required methods
    fn read(&self, buf: &mut [u8], offset: usize) -> Result<usize, SvsmError>;
    fn write(&self, buf: &[u8], offset: usize) -> Result<usize, SvsmError>;
    fn truncate(&self, size: usize) -> Result<usize, SvsmError>;
    fn size(&self) -> usize;

    // Provided method
    fn mapping(&self, _offset: usize) -> Option<PageRef> { ... }
}

Represents file operations

Required Methods

fn read(&self, buf: &mut [u8], offset: usize) -> Result<usize, SvsmError>

Used to read contents of a file

Arguments

buf: buffer to read the file contents into.

offset: file offset to read from.

Returns

Result<usize, SvsmError>: A Result containing the number of bytes read if successful, or an SvsmError if there was a problem during the read operation.

fn write(&self, buf: &[u8], offset: usize) -> Result<usize, SvsmError>

Used to write contents to a file

Arguments

buf: buffer which holds the contents to be written to the file.

offset: file offset to write to.

Returns

Result<usize, SvsmError>: A Result containing the number of bytes written if successful, or an SvsmError if there was a problem during the write operation.

fn truncate(&self, size: usize) -> Result<usize, SvsmError>

Used to truncate the file to the specified size.

Arguments

size: specifies the size in bytes to which the file is to be truncated.

Returns

Result<usize, SvsmError>: A Result containing the size of the file after truncation if successful, or an SvsmError if there was a problem during the truncate operation.

fn size(&self) -> usize

Used to get the size of the file.

Returns

size of the file in bytes.

Provided Methods

fn mapping(&self, _offset: usize) -> Option<PageRef>

Get reference to backing pages of the file

Arguments

offset: offset to the requested page in bytes

Returns

Option<PageRef>: An Option with the requested page reference. None if the offset is not backed by a page.