svsm::greq

Module driver

Driver to send SNP_GUEST_REQUEST commands to the PSP. It can be any of the request or response command types defined in the SEV-SNP spec, regardless if it’s a regular or an extended command.

Structs

SnpGuestRequestDriver 🔒

SNP_GUEST_REQUEST driver

Enums

SnpGuestRequestClass 🔒

Class of the SNP_GUEST_REQUEST command: Regular or Extended

Constants

SNP_GUEST_REQ_ERR_BUSY 🔒

Hypervisor busy, try again

SNP_GUEST_REQ_INVALID_LEN 🔒

Buffer provided is too small

Statics

GREQ_DRIVER 🔒

Global SNP_GUEST_REQUEST driver instance

Functions

guest_request_driver_init

Initialize the global SnpGuestRequestDriver

send_extended_guest_request

Send the provided extended SNP_GUEST_REQUEST command to the PSP Further details can be found in the SnpGuestRequestDriver.send_request() documentation.

send_regular_guest_request

Send the provided regular SNP_GUEST_REQUEST command to the PSP. Further details can be found in the SnpGuestRequestDriver.send_request() documentation.