svsm::greq::driver

Struct SnpGuestRequestDriver

struct SnpGuestRequestDriver {
    request: SharedBox<SnpGuestRequestMsg>,
    response: SharedBox<SnpGuestRequestMsg>,
    staging: Box<SnpGuestRequestMsg>,
    ext_data: SharedBox<SnpGuestRequestExtData>,
    user_extdata_size: usize,
    vmpck0_seqno: u64,
}

SNP_GUEST_REQUEST driver

Fields
request: SharedBox<SnpGuestRequestMsg>

Shared page used for the SNP_GUEST_REQUEST request

response: SharedBox<SnpGuestRequestMsg>

Shared page used for the SNP_GUEST_REQUEST response

staging: Box<SnpGuestRequestMsg>

Encrypted page where we perform crypto operations

ext_data: SharedBox<SnpGuestRequestExtData>

Extended data buffer that will be provided to the hypervisor to store the SEV-SNP certificates

user_extdata_size: usize

Extended data size (certs size) provided by the user in super::services::get_extended_report. It will be provided to the hypervisor.

vmpck0_seqno: u64

Each SNP_GUEST_REQUEST message contains a sequence number per VMPCK. The sequence number is incremented with each message sent. Messages sent by the guest to the PSP and by the PSP to the guest must be delivered in order. If not, the PSP will reject subsequent messages by the guest when it detects that the sequence numbers are out of sync.

NOTE: If the vmpl field of a SNP_GUEST_REQUEST message is set to VMPL0, then it must contain the VMPL0 sequence number and be protected (encrypted) with the VMPCK0 key; additionally, if this message fails, the VMPCK0 key must be disabled. The same idea applies to the other VMPL levels.

The SVSM needs to support only VMPL0 SNP_GUEST_REQUEST commands because other layers in the software stack (e.g. OVMF and guest kernel) can send non-VMPL0 commands directly to PSP. Therefore, the SVSM needs to maintain the sequence number and the VMPCK only for VMPL0.

Implementations

impl SnpGuestRequestDriver

pub fn new() -> Result<Self, SvsmReqError>

Create a new SnpGuestRequestDriver

fn seqno_last_used(&self) -> u64

Get the last VMPCK0 sequence number accounted

fn seqno_add_two(&mut self)

Increase the VMPCK0 sequence number by two. In order to keep the sequence number in-sync with the PSP, this is called only when the SNP_GUEST_REQUEST response is received.

pub fn set_user_extdata_size(&mut self, n: usize) -> Result<(), SvsmReqError>

Set the user_extdata_size to n and clear the first n bytes from ext_data

fn send(&mut self, req_class: SnpGuestRequestClass) -> Result<(), SvsmReqError>

Call the GHCB layer to send the encrypted SNP_GUEST_REQUEST message to the PSP.

fn encrypt_request( &mut self, msg_type: SnpGuestRequestMsgType, msg_seqno: u64, buffer: &[u8], command_len: usize, ) -> Result<(), SvsmReqError>

fn decrypt_response( &mut self, msg_seqno: u64, msg_type: SnpGuestRequestMsgType, buffer: &mut [u8], ) -> Result<usize, SvsmReqError>

fn send_request( &mut self, req_class: SnpGuestRequestClass, msg_type: SnpGuestRequestMsgType, buffer: &mut [u8], command_len: usize, ) -> Result<usize, SvsmReqError>

Send the provided VMPL0 SNP_GUEST_REQUEST command to the PSP.

The command will be encrypted using AES-256 GCM.

Arguments

req_class: whether this is a regular or extended SNP_GUEST_REQUEST command

msg_type: type of the command stored in buffer, e.g. SNP_MSG_REPORT_REQ

buffer: buffer with the SNP_GUEST_REQUEST command to be sent. The same buffer will also be used to store the response.

command_len: Size (in bytes) of the command stored in buffer

Returns

Success:

usize: Size (in bytes) of the response stored in buffer

Error:

SvsmReqError

pub fn send_regular_guest_request( &mut self, msg_type: SnpGuestRequestMsgType, buffer: &mut [u8], command_len: usize, ) -> Result<usize, SvsmReqError>

Send the provided regular SNP_GUEST_REQUEST command to the PSP

pub fn send_extended_guest_request( &mut self, msg_type: SnpGuestRequestMsgType, buffer: &mut [u8], command_len: usize, certs: &mut [u8], ) -> Result<usize, SvsmReqError>

Send the provided extended SNP_GUEST_REQUEST command to the PSP