svsm::greq::msg

Struct SnpGuestRequestMsg

#[repr(C, align(4096))]
pub struct SnpGuestRequestMsg {
    hdr: SnpGuestRequestMsgHdr,
    pld: [u8; 4000],
}

SNP_GUEST_REQUEST message format

Fields
hdr: SnpGuestRequestMsgHdrpld: [u8; 4000]
Implementations

impl SnpGuestRequestMsg

pub fn encrypt_set( &mut self, msg_type: SnpGuestRequestMsgType, msg_seqno: u64, vmpck0: &[u8; 32], command: &[u8], ) -> Result<(), SvsmReqError>

Encrypt the provided SNP_GUEST_REQUEST command and store the result in the actual message payload

The command will be encrypted using AES-256 GCM and part of the message header will be used as additional authenticated data (AAD).

Arguments

msg_type: Type of the command stored in the command buffer.

msg_seqno: VMPL0 sequence number to be used in the message. The PSP will reject subsequent messages when it detects that the sequence numbers are out of sync. The sequence number is also used as initialization vector (IV) in encryption.

vmpck0: VMPCK0 key that will be used to encrypt the command.

command: command slice to be encrypted.

Returns

() on success and SvsmReqError on error.

Panic

The command length does not fit in a u16

The encrypted and the original command donโt have the same size

pub fn decrypt_get( &mut self, msg_type: SnpGuestRequestMsgType, msg_seqno: u64, vmpck0: &[u8; 32], outbuf: &mut [u8], ) -> Result<usize, SvsmReqError>

Decrypt the SNP_GUEST_REQUEST command stored in the message and store the decrypted command in the provided outbuf.

The command stored in the message payload is usually a response command received from the PSP. It will be decrypted using AES-256 GCM and part of the message header will be used as additional authenticated data (AAD).

Arguments

msg_type: Type of the command stored in the message payload

msg_seqno: VMPL0 sequence number that was used in the message.

vmpck0: VMPCK0 key, it will be used to decrypt the message

outbuf: buffer that will be used to store the decrypted message payload

Returns

Success

usize: Number of bytes written to outbuf

Error

SvsmReqError