svsm::greq::services

Function get_extended_report

pub fn get_extended_report(
    buffer: &mut [u8],
    certs: &mut [u8],
) -> Result<usize, SvsmReqError>

Request an extended VMPL0 attestation report to the PSP.

We say that it is extended because it requests a VMPL0 attestation report to the PSP (as in get_regular_report()) and also requests to the hypervisor the certificates required to verify the attestation report.

The VMPCK0 is disabled for subsequent calls if this function fails in a way that the VM state can be compromised.

Arguments

buffer: Buffer with the MSG_REPORT_REQ command that will be sent to the PSP. It must be large enough to hold the MSG_REPORT_RESP received from the PSP.

certs: Buffer to store the SEV-SNP certificates received from the hypervisor.

Return codes

Success

usize: Number of bytes written to buffer. It should match the MSG_REPORT_RESP size.

Error

SvsmReqError

SvsmReqError::FatalError(SvsmError::Ghcb(GhcbError::VmgexitError(certs_buffer_size, psp_rc))):

certs is not large enough to hold the certificates.

certs_buffer_size: number of bytes required.

psp_rc: PSP return code