svsm::greq::services

Function get_regular_report

pub fn get_regular_report(buffer: &mut [u8]) -> Result<usize, SvsmReqError>

Request a regular VMPL0 attestation report to the PSP.

Use the SNP_GUEST_REQUEST driver to send the provided MSG_REPORT_REQ command to the PSP. The VPML field of the command must be set to zero.

The VMPCK0 is disabled for subsequent calls if this function fails in a way that the VM state can be compromised.

Arguments

buffer: Buffer with the MSG_REPORT_REQ command that will be sent to the PSP. It must be large enough to hold the MSG_REPORT_RESP received from the PSP.

Returns

Success

usize: Number of bytes written to buffer. It should match the MSG_REPORT_RESP size.

Error

SvsmReqError