svsm::hyperv
Module hv

Structs

HvHypercallInput 🔒

HvHypercallOutput 🔒

HvInputEnableVpVtl 🔒

HvInputGetVpRegister 🔒

Dead code is allowed because the hypercall input logic doesn’t recognize that all fields are read when they are copied into the hypercall input page.

HvInputStartVirtualProcessor 🔒

HypercallPagesGuard

Enums

HvCallCode 🔒

Constants

HV_PARTITION_ID_SELF

HV_VP_INDEX_SELF

Statics

CURRENT_VTL 🔒

HYPERV_HYPERCALL_CODE_PAGE 🔒

Functions

enable_vp_vtl_hypercall 🔒

get_vp_register

hypercall 🔒 ⚠

Safety

hyperv_setup_hypercalls

hyperv_start_cpu

is_hyperv_hypervisor

start_vp_hypercall 🔒