svsm::igvm_params

Struct IgvmBox

pub struct IgvmBox<'a> {
    vaddr: VirtAddr,
    igvm_params: IgvmParams<'a>,
}

IgvmBox is a Box-type object that tracks the allocation lifetime of the IGVM parameters. This is implemented separately from PageBox because unlike normal heap allocations, the IGVM parameters are allocated as a sequence of single pages, and thus cannot be freed in a single operation.

Fields
vaddr: VirtAddrigvm_params: IgvmParams<'a>
Implementations

impl IgvmBox<'_>

pub unsafe fn new(vaddr: VirtAddr) -> Result<Self, SvsmError>

Safety

The caller is responsible for ensuring that the supplied virtual address corresponds to an IGVM parameter block.