Module svsm::insn_decode::decode

Structs

DecodedBytes ğŸ”’

DecodedInsnCtx

Represents the context of a decoded instruction, which is used to interpret the instruction. It holds the decoded instruction, its length and various components that are decoded from the instruction bytes.

DisBytes ğŸ”’

ImmBytes ğŸ”’

InsnBytes

Represents the raw bytes of an instruction and tracks the number of bytes being processed.

ModRM ğŸ”’

ModRmBytes ğŸ”’

MoffBytes ğŸ”’

OpCodeBytes

The instruction bytes specifically for OpCode decoding

PrefixBytes ğŸ”’

PrefixFlags ğŸ”’

RegCode ğŸ”’

RexPrefix ğŸ”’

Sib ğŸ”’

SibBytes ğŸ”’

Enums

CpuMode ğŸ”’

Mod ğŸ”’

PagingLevel ğŸ”’

RM ğŸ”’

Constants

MOD_DIRECT ğŸ”’

MOD_INDIRECT ğŸ”’

MOD_INDIRECT_DISP8 ğŸ”’

MOD_INDIRECT_DISP32 ğŸ”’

PREFIX_SIZE ğŸ”’

RM_DISP32 ğŸ”’

RM_SIB ğŸ”’

Traits

InsnMachineCtx

This trait provides the necessary context for an instruction decoder to decode instructions based on the current state of the machine that executed them. It abstracts the interfaces through which an instruction decoder can access specific registers and state that may influence the decoding from the machine (such as a CPU or VMM).

InsnMachineMem

Trait representing a machine memory for instruction decoding.

Functions

get_cpu_mode ğŸ”’

ioio_perm ğŸ”’

read_bytereg ğŸ”’

read_reg ğŸ”’

segment_base ğŸ”’

segment_limit ğŸ”’

write_bytereg ğŸ”’

write_reg ğŸ”’