svsm::insn_decode

Trait InsnMachineCtx

pub trait InsnMachineCtx: Debug {

    // Required methods
    fn read_efer(&self) -> u64;
    fn read_seg(&self, seg: SegRegister) -> u64;
    fn read_cr0(&self) -> u64;
    fn read_cr4(&self) -> u64;

    // Provided methods
    fn read_reg(&self, _reg: Register) -> usize { ... }
    fn read_flags(&self) -> usize { ... }
    fn write_reg(&mut self, _reg: Register, _val: usize) { ... }
    fn read_cpl(&self) -> usize { ... }
    fn map_linear_addr<T: Copy + 'static>(
        &self,
        _la: usize,
        _write: bool,
        _fetch: bool,
    ) -> Result<Box<dyn InsnMachineMem<Item = T>>, InsnError> { ... }
    fn ioio_perm(&self, _port: u16, _size: Bytes, _io_read: bool) -> bool { ... }
    fn ioio_in(&self, _port: u16, _size: Bytes) -> Result<u64, InsnError> { ... }
    fn ioio_out(
        &mut self,
        _port: u16,
        _size: Bytes,
        _data: u64,
    ) -> Result<(), InsnError> { ... }
    fn translate_linear_addr(
        &self,
        _la: usize,
        _write: bool,
        _fetch: bool,
    ) -> Result<(usize, bool), InsnError> { ... }
    unsafe fn handle_mmio_read(
        &self,
        _pa: usize,
        _shared: bool,
        _size: Bytes,
    ) -> Result<u64, InsnError> { ... }
    unsafe fn handle_mmio_write(
        &mut self,
        _pa: usize,
        _shared: bool,
        _size: Bytes,
        _data: u64,
    ) -> Result<(), InsnError> { ... }

}

This trait provides the necessary context for an instruction decoder to decode instructions based on the current state of the machine that executed them. It abstracts the interfaces through which an instruction decoder can access specific registers and state that may influence the decoding from the machine (such as a CPU or VMM).

Required Methods

fn read_efer(&self) -> u64

Read EFER register

fn read_seg(&self, seg: SegRegister) -> u64

Read a code segment register

fn read_cr0(&self) -> u64

Read CR0 register

fn read_cr4(&self) -> u64

Read CR4 register

Provided Methods

fn read_reg(&self, _reg: Register) -> usize

Read a register

fn read_flags(&self) -> usize

Read rflags register

fn write_reg(&mut self, _reg: Register, _val: usize)

Write a register

fn read_cpl(&self) -> usize

Read the current privilege level

fn map_linear_addr<T: Copy + 'static>( &self, _la: usize, _write: bool, _fetch: bool, ) -> Result<Box<dyn InsnMachineMem<Item = T>>, InsnError>

Map the given linear address region to a machine memory object which provides access to the memory of this linear address region.

Arguments

la - The linear address of the region to map.

write - Whether write access is allowed to the mapped region.

fetch - Whether fetch access is allowed to the mapped region.

Returns

A Result containing a boxed trait object representing the mapped memory, or an InsnError if mapping fails.

fn ioio_perm(&self, _port: u16, _size: Bytes, _io_read: bool) -> bool

Check IO permission bitmap.

Arguments

port - The I/O port to check.

size - The size of the I/O operation.

io_read - Whether the I/O operation is a read operation.

Returns

A Result containing true if the port is permitted otherwise false.

fn ioio_in(&self, _port: u16, _size: Bytes) -> Result<u64, InsnError>

Handle an I/O in operation.

Arguments

port - The I/O port to read from.

size - The size of the data to read.

Returns

A Result containing the read data if success or an InsnError if the operation fails.

fn ioio_out( &mut self, _port: u16, _size: Bytes, _data: u64, ) -> Result<(), InsnError>

Handle an I/O out operation.

Arguments

port - The I/O port to write to.

size - The size of the data to write.

data - The data to write to the I/O port.

Returns

A Result indicating success or an InsnError if the operation fails.

fn translate_linear_addr( &self, _la: usize, _write: bool, _fetch: bool, ) -> Result<(usize, bool), InsnError>

Translate the given linear address to a physical address.

Arguments

la - The linear address to translate.

write - Whether the translation is for a write operation.

fetch - Whether the translation is for a fetch operation.

Returns

A Result containing the translated physical address and a boolean indicating whether the physical address is shared or an InsnError if the translation fails.

unsafe fn handle_mmio_read( &self, _pa: usize, _shared: bool, _size: Bytes, ) -> Result<u64, InsnError>

Handle a memory-mapped I/O read operation.

Arguments

pa - The MMIO physical address to read from.

shared - Whether the MMIO address is shared.

size - The size of the data to read.

Returns

A Result containing the read data if success or an InsnError if the operation fails.

Safety

Caller must ensure that, when memory safety may be impacted, this operation does not cause Undefined Behavior.

unsafe fn handle_mmio_write( &mut self, _pa: usize, _shared: bool, _size: Bytes, _data: u64, ) -> Result<(), InsnError>

Handle a memory-mapped I/O write operation.

Arguments

pa - The MMIO physical address to write to.

shared - Whether the MMIO address is shared.

size - The size of the data to write.

data - The data to write to the MMIO.

Returns

A Result indicating success or an InsnError if the operation fails.

Safety

Caller must ensure that, when memory safety may be impacted, this operation does not cause Undefined Behavior.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.