Trait svsm::insn_decode::InsnMachineMem

pub trait InsnMachineMem {
    type Item;

    // Provided methods
    unsafe fn mem_read(&self) -> Result<Self::Item, InsnError> { ... }
    unsafe fn mem_write(&mut self, _data: Self::Item) -> Result<(), InsnError> { ... }
}

Trait representing a machine memory for instruction decoding.

Required Associated Types

type Item

Provided Methods

unsafe fn mem_read(&self) -> Result<Self::Item, InsnError>

Read data from the memory at the specified offset.

Safety

The caller must verify not to read data from arbitrary memory. The object implements this trait should guarantee the memory region is readable.

Returns

Returns the read data on success, or an InsnError if the read operation fails.

unsafe fn mem_write(&mut self, _data: Self::Item) -> Result<(), InsnError>

Write data to the memory at the specified offset.

Safety

The caller must verify not to write data to corrupt arbitrary memory. The object implements this trait should guarantee the memory region is writable.

Arguments

data - The data to write to the memory.

Returns

Returns Okon success, or an InsnError if the write operation fails.