svsm::locking::common

Trait IrqLocking

pub trait IrqLocking {
    // Required method
    fn acquire_lock() -> Self;
}

Abstracts TPR and interrupt state handling when taking and releasing locks. There are three implemenations:

IrqUnsafeLocking implements the methods as no-ops and does not change any IRQ or TPR state.

IrqGuardLocking actually disables and enables IRQs in the methods, ensuring that no interrupt can be taken while the lock is held.

TprGuardLocking raises and lowers TPR while the lock is held, ensuring that no higher priority interrupt can be taken while the lock is held. This will panic when attempting to acquire a lower priority lock from a higher priority interrupt context.

Required Methods

fn acquire_lock() -> Self

Associated helper function to modify TPR/interrupt state when a lock is acquired. This is used by lock implementations and will return an instance of the object.

Returns

New instance of implementing struct.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.