svsm::locking::rwlock

Struct RawRWLock

pub struct RawRWLock<T, I> {
    rwlock: AtomicU64,
    data: UnsafeCell<T>,
    phantom: PhantomData<fn(I)>,
}

A simple Read-Write Lock (RWLock) that allows multiple readers or one exclusive writer.

Fields
rwlock: AtomicU64

An atomic 64-bit integer used for synchronization

data: UnsafeCell<T>

An UnsafeCell for interior mutability

phantom: PhantomData<fn(I)>

Silence unused type warning

Implementations

impl<T: Send, I: IrqLocking> RawRWLock<T, I>

A reader-writer lock that allows multiple readers or a single writer to access the protected data. RWLock provides exclusive access for writers and shared access for readers, for efficient synchronization.

A lock can only be formed if the type it protects is Send, since the contents of the lock will be sent to different threads.

pub const fn new(data: T) -> Self

Creates a new RWLock instance with the provided initial data.

Parameters

data: The initial data to be protected by the RWLock.

Returns

A new RWLock instance with the specified initial data.

Example

use svsm::locking::RWLock;

#[derive(Debug)]
struct MyData {
    value: i32,
}

let data = MyData { value: 42 };
let rwlock = RWLock::new(data);

pub const fn as_ptr(&self) -> *mut T

Returns a pointer into the underlying data.

fn wait_for_writers(&self) -> u64

This function is used to wait until all writers have finished their operations and retrieve the current state of the RWLock.

Returns

A 64-bit value representing the current state of the RWLock, including the count of readers and writers.

fn wait_for_readers(&self) -> u64

This function is used to wait until all readers have finished their operations and retrieve the current state of the RWLock.

Returns

A 64-bit value representing the current state of the RWLock, including the count of readers and writers.

pub fn lock_write(&self) -> RawWriteLockGuard<'_, T, I>

This function ensures exclusive access for a single writer and waits for all readers to finish before granting access to the writer.

Returns

A WriteLockGuard that provides write access to the protected data.

pub fn try_lock_write(&self) -> Option<RawWriteLockGuard<'_, T, I>>

Attempts to acquire the lock for writing without blocking. If the access could not be granted at this time, None is returned.

pub fn write_noblock(&self) -> RawWriteLockGuard<'_, T, I>

Attempts to acquire the lock for writing in a single try, panicking if the operation would block.

This is useful for per-CPU structures that use locks to detect reentrancy, ensuring that a deadlock instead produces an immediate panic.

impl<T: Send + Sync, I: IrqLocking> RawRWLock<T, I>

A lock can only be acquired for read access if its inner type implements Sync as well as Send. This is because a read lock can be acquired simultaneously by multiple threads, and therefore the data must be shareable.

pub fn lock_read(&self) -> RawReadLockGuard<'_, T, I>

This function allows multiple readers to access the data concurrently.

Returns

A ReadLockGuard that provides read access to the protected data.

pub fn try_lock_read(&self) -> Option<RawReadLockGuard<'_, T, I>>

Attempts to acquire the lock for reading without blocking. If the access could not be granted at this time, None is returned.

pub fn read_noblock(&self) -> RawReadLockGuard<'_, T, I>

Attempts to acquire the lock for reading in a single try, panicking if the operation would block.

This is useful for per-CPU structures that use locks to detect reentrancy, ensuring that a deadlock instead produces an immediate panic.