Module svsm::locking::spinlock

Structs

RawLockGuard

A lock guard obtained from a SpinLock. This lock guard provides exclusive access to the data protected by a SpinLock, ensuring that the lock is released when it goes out of scope.

RawSpinLock

A simple ticket-spinlock implementation for protecting concurrent data access.

Type Aliases

LockGuard

LockGuardIrqSafe

SpinLock

SpinLockIrqSafe