Struct svsm::locking::spinlock::RawSpinLock

pub struct RawSpinLock<T, I = IrqUnsafeLocking> {
    current: AtomicU64,
    holder: AtomicU64,
    data: UnsafeCell<T>,
    phantom: PhantomData<fn(_: I)>,
}

A simple ticket-spinlock implementation for protecting concurrent data access.

Two variants are derived from this implementation:

SpinLock for general use. This implementation is not safe for use in IRQ handlers.

SpinLockIrqSafe for protecting data that is accessed in IRQ context.

Examples

use svsm::locking::SpinLock;

let data = 42;
let spin_lock = SpinLock::new(data);

// Acquire the lock and modify the protected data.
{
    let mut guard = spin_lock.lock();
    *guard += 1;
}; // Lock is automatically released when `guard` goes out of scope.

// Try to acquire the lock without blocking
if let Some(mut guard) = spin_lock.try_lock() {
    *guard += 2;
};

Fields
current: AtomicU64

This atomic counter is incremented each time a thread attempts to acquire the lock. It helps to determine the order in which threads acquire the lock.

holder: AtomicU64

This counter represents the thread that currently holds the lock and has access to the protected data.

data: UnsafeCell<T>

This UnsafeCell is used to provide interior mutability of the protected data. That is, it allows the data to be accessed/modified while enforcing the locking mechanism.

phantom: PhantomData<fn(_: I)>

Use generic type I in the struct without consuming space.

Implementations

impl<T, I: IrqLocking> RawSpinLock<T, I>

pub const fn new(data: T) -> Self

Creates a new SpinLock instance with the specified initial data.

Examples

use svsm::locking::SpinLock;

let data = 42;
let spin_lock = SpinLock::new(data);

pub fn lock(&self) -> RawLockGuard<'_, T, I>

Acquires the lock, providing access to the protected data.

Examples

use svsm::locking::SpinLock;

let spin_lock = SpinLock::new(42);

// Acquire the lock and modify the protected data.
{
    let mut guard = spin_lock.lock();
    *guard += 1;
}; // Lock is automatically released when `guard` goes out of scope.

pub fn try_lock(&self) -> Option<RawLockGuard<'_, T, I>>

This method tries to acquire the lock without blocking. If the lock is not available, it returns None. If the lock is successfully acquired, it returns a LockGuard that automatically releases the lock when it goes out of scope.