svsm::locking::spinlock

Type Alias SpinLock

pub type SpinLock<T> = RawSpinLock<T, IrqUnsafeLocking>;

Aliased Type

struct SpinLock<T> {
    current: AtomicU64,
    holder: AtomicU64,
    data: UnsafeCell<T>,
    phantom: PhantomData<fn(IrqUnsafeLocking)>,
}

Fields
current: AtomicU64

This atomic counter is incremented each time a thread attempts to acquire the lock. It helps to determine the order in which threads acquire the lock.

holder: AtomicU64

This counter represents the thread that currently holds the lock and has access to the protected data.

data: UnsafeCell<T>

This UnsafeCell is used to provide interior mutability of the protected data. That is, it allows the data to be accessed/modified while enforcing the locking mechanism.

phantom: PhantomData<fn(IrqUnsafeLocking)>

Use generic type I in the struct without consuming space.