svsm
Module mm

Re-exports

pub use guestmem::copy_from_user;

pub use guestmem::copy_to_user;

pub use guestmem::GuestPtr;

pub use memory::valid_phys_address;

pub use memory::writable_phys_addr;

pub use pagetable::PageTablePart;

pub use alloc::allocate_file_page;

pub use alloc::PageRef;

pub use global_memory::map_global_range;

pub use global_memory::map_global_range_2m_private;

pub use global_memory::map_global_range_2m_shared;

pub use global_memory::map_global_range_4k_private;

pub use global_memory::map_global_range_4k_shared;

pub use global_memory::GlobalRangeGuard;

pub use mappings::mmap_kernel;

pub use mappings::mmap_user;

pub use mappings::munmap_kernel;

pub use mappings::munmap_user;

pub use mappings::VMMappingGuard;

pub use address_space::*;

pub use ptguards::*;

Modules

address_space

alloc

global_memory

guestmem

mappings

memory

page_visibility

pagebox 🔒

pagetable

ptguards

validate

virtualrange

vm

Structs

PageBox

An abstraction, similar to a Box, for types that need to be allocated using page allocator directly.