svsm::mm::alloc

Struct MemoryRegion

struct MemoryRegion {
    start_phys: PhysAddr,
    start_virt: VirtAddr,
    page_count: usize,
    nr_pages: [usize; 6],
    next_page: [usize; 6],
    free_pages: [usize; 6],
}

Memory region with its physical/virtual addresses, page count, as well as other details.

Fields
start_phys: PhysAddrstart_virt: VirtAddrpage_count: usizenr_pages: [usize; 6]next_page: [usize; 6]free_pages: [usize; 6]
Implementations

impl MemoryRegion

const fn new() -> Self

Creates a new MemoryRegion with default values.

fn phys_to_virt(&self, paddr: PhysAddr) -> Option<VirtAddr>

Converts a physical address within this memory region to a virtual address.

fn virt_to_phys(&self, vaddr: VirtAddr) -> Option<PhysAddr>

Converts a virtual address to a physical address within the memory region.

unsafe fn page_info_mut_ptr(&mut self, pfn: usize) -> *mut PageStorageType

Gets a mutable pointer to the page information for a given page frame number.

Safety

The caller must provide a valid pfn, otherwise the returned pointer is undefined, as the compiler is allowed to optimize assuming there will be no arithmetic overflows.

unsafe fn page_info_ptr(&self, pfn: usize) -> *const PageStorageType

Gets a pointer to the page information for a given page frame number.

Safety

The caller must provide a valid pfn, otherwise the returned pointer is undefined, as the compiler is allowed to optimize assuming there will be no arithmetic overflows.

fn check_pfn(&self, pfn: usize)

Checks if a page frame number is valid.

Panics

Panics if the page frame number is invalid.

fn end_virt(&self) -> VirtAddr

Calculates the end virtual address of the memory region.

fn write_page_info(&mut self, pfn: usize, pi: PageInfo)

Writes page information for a given page frame number.

fn read_page_info(&self, pfn: usize) -> PageInfo

Reads page information for a given page frame number.

fn get_virt_offset(&self, vaddr: VirtAddr) -> Option<usize>

Gets the virtual offset of a virtual address within the memory region.

fn get_pfn(&self, vaddr: VirtAddr) -> Result<usize, AllocError>

Gets the page frame number for a given virtual address.

fn get_next_page(&mut self, order: usize) -> Result<usize, AllocError>

Gets the next available page frame number for a given order.

fn mark_compound_page(&mut self, pfn: usize, order: usize)

Marks a compound page and updates page information for neighboring pages.

fn init_compound_page(&mut self, pfn: usize, order: usize, next_pfn: usize)

Initializes a compound page with given page frame numbers and order.

fn split_page(&mut self, pfn: usize, order: usize) -> Result<(), AllocError>

Splits a page into two pages of the next lower order.

fn refill_page_list(&mut self, order: usize) -> Result<(), AllocError>

Refills the free page list for a given order.

fn allocate_pages_info( &mut self, order: usize, pg: PageInfo, ) -> Result<VirtAddr, AllocError>

Allocates pages with a specific order and page information.

fn allocate_pages(&mut self, order: usize) -> Result<VirtAddr, AllocError>

Allocates pages with a specific order.

fn allocate_page(&mut self) -> Result<VirtAddr, AllocError>

Allocates a single page.

fn allocate_zeroed_page(&mut self) -> Result<VirtAddr, AllocError>

Allocates a zeroed page.

fn allocate_slab_page(&mut self, item_size: u16) -> Result<VirtAddr, AllocError>

Allocates a slab page.

fn allocate_file_page(&mut self) -> Result<VirtAddr, AllocError>

Allocates a file page with initial reference count.

fn get_file_page(&mut self, vaddr: VirtAddr) -> Result<(), AllocError>

Gets a file page and increments its reference count.

fn put_file_page(&mut self, vaddr: VirtAddr) -> Result<(), AllocError>

Releases a file page and decrements its reference count.

fn compound_neighbor( &self, pfn: usize, order: usize, ) -> Result<usize, AllocError>

Finds the neighboring page frame number for a compound page.

fn merge_pages( &mut self, pfn1: usize, pfn2: usize, order: usize, ) -> Result<usize, AllocError>

Merges two pages of the same order into a new compound page.

fn next_free_pfn(&self, pfn: usize, order: usize) -> usize

Gets the next free page frame number from the free list.

fn allocate_pfn(&mut self, pfn: usize, order: usize) -> Result<(), AllocError>

Allocates a specific page frame number (pfn) within a given order. If the page frame number is not found or is already allocated, an error is returned. If the requested page frame number is the first in the list, it is marked as allocated, and the next page in the list becomes the new first page.

Panics

Panics if order is greater than MAX_ORDER.

fn free_page_raw(&mut self, pfn: usize, order: usize)

Frees a raw page by updating the free list and marking it as a free page.

Panics

Panics if order is greater than MAX_ORDER.

fn try_to_merge_page( &mut self, pfn: usize, order: usize, ) -> Result<usize, AllocError>

Attempts to merge a given page with its neighboring page. If successful, returns the new page frame number after merging. If unsuccessful, the page remains unmerged, and an error is returned.

fn free_page_order(&mut self, pfn: usize, order: usize)

Frees a page of a specific order. If merging is successful, it continues merging until merging is no longer possible. If merging fails, the page is marked as a free page.

fn free_page(&mut self, vaddr: VirtAddr)

Frees a page based on its virtual address, determining the page order and freeing accordingly.

fn memory_info(&self) -> MemInfo

Retrieves information about memory, including total and free pages in different orders.

fn init_memory(&mut self)

Initializes memory by marking certain pages as reserved and the rest as allocated. It then frees all pages and organizes them into their respective order buckets.