svsm::mm::alloc

Struct SvsmAllocator

pub struct SvsmAllocator {
    slab32: SpinLock<Slab<32>>,
    slab64: SpinLock<Slab<64>>,
    slab128: SpinLock<Slab<128>>,
    slab256: SpinLock<Slab<256>>,
    slab512: SpinLock<Slab<512>>,
    slab1024: SpinLock<Slab<1024>>,
    slab2048: SpinLock<Slab<2048>>,
}

Represents a simple virtual-to-physical memory allocator (SvsmAllocator) implementing the GlobalAlloc trait.

This allocator uses slab allocation for fixed-size objects and falls back to page allocation for larger objects.

Fields
slab32: SpinLock<Slab<32>>slab64: SpinLock<Slab<64>>slab128: SpinLock<Slab<128>>slab256: SpinLock<Slab<256>>slab512: SpinLock<Slab<512>>slab1024: SpinLock<Slab<1024>>slab2048: SpinLock<Slab<2048>>
Implementations

impl SvsmAllocator

pub const fn new() -> Self

Creates a new instance of SvsmAllocator with initialized slab allocators.

fn allocate(&self, size: usize) -> Option<Result<VirtAddr, AllocError>>

unsafe fn deallocate(&self, addr: VirtAddr, size: usize) -> Option<()>

Deallocates memory from a Slab.

Safety:

The addr parameter must point to a virtual address allocated via a Slab managed by / SvsmAllocator. The size parameter must match the value used during / allocation of the virtual address.