svsm::mm
Module global_memory

Structs

GlobalRangeGuard

GlobalRanges 🔒

Statics

GLOBAL_RANGES 🔒

Functions

init_global_ranges

Initialize global allocatable virtual address ranges.

map_global_range

Map physical addresses into the global shared address range.

map_global_range_2m_private

Create a private mapping using of physical addresses into the global shared address range using 2MiB pages.

map_global_range_2m_shared

Create a shared mapping using of physical addresses into the global shared address range using 2MiB pages.

map_global_range_4k_private

Create a private mapping using of physical addresses into the global shared address range using 4KiB pages.

map_global_range_4k_shared

Create a shared mapping using of physical addresses into the global shared address range using 4KiB pages.