svsm::mm::global_memory

Function map_global_range_2m_shared

pub fn map_global_range_2m_shared(
    pstart: PhysAddr,
    size: usize,
    flags: PTEntryFlags,
) -> Result<GlobalRangeGuard, SvsmError>

Create a shared mapping using of physical addresses into the global shared address range using 2MiB pages.

Arguments

pstart: Start physical to map, must be aligned to 2MiB.

size: Number of bytes to map. Will be aligned up to 2MiB.

flages: Page-table flags to use for mapping.

Returns

A Result with a GlobalRangeGuard on success or SvsmError on failure.