svsm::mm

Module guestmem

Structs

GuestPtr

UserAccessGuard 🔒

UserPtr

Functions

check_bounds_user 🔒

checked_guest_region 🔒

copy_bytes 🔒 ⚠

Copies size number of bytes from src to dst, catching any fault that might happen during the operation.

copy_from_guest⚠

Reads a slice of bytes from a physical address region outside of SVSM use.

copy_from_user

copy_slice_from_guest

Reads a slice of bytes from a physical address region outside of SVSM use.

copy_slice_to_guest

Writes a slice of bytes to a physical address region outside of SVSM use.

copy_to_user

do_movsb 🔒 ⚠

Copies src to dst.

read_bytes_from_guest

Reads a vector of bytes from a physical address region outside of SVSM use.

read_from_guest

Reads an instance of T from a physical address region outside of SVSM use.

read_u8⚠

Read one byte from a virtual address.

read_u16 🔒 ⚠

Read one word from a virtual address.

read_u32 🔒 ⚠

Read one dword from a virtual address.

read_u64 🔒 ⚠

Read one qword from a virtual address.

write_to_guest

Writes a value to a physical address region outside of SVSM use.

write_u8⚠

Writes one byte at a virtual address.