svsm::mm::guestmem

Function copy_bytes

unsafe fn copy_bytes(
    src: *const u8,
    dst: *mut u8,
    size: usize,
) -> Result<(), SvsmError>

Copies size number of bytes from src to dst, catching any fault that might happen during the operation.

Safety

The caller must make sure that writing to dst does not harm memory safety.