svsm::mm::guestmem

Struct GuestPtr

pub struct GuestPtr<T: Copy> {
    ptr: *mut T,
}

Fields
ptr: *mut T
Implementations

impl<T: Copy> GuestPtr<T>

pub fn new(v: VirtAddr) -> Self

pub const fn from_ptr(p: *mut T) -> Self

pub unsafe fn read(&self) -> Result<T, SvsmError>

Safety

The caller must verify not to read arbitrary memory, as this function doesn’t make any checks in that regard.

Returns

Returns an error if the specified address is not mapped.

pub unsafe fn write(&self, buf: T) -> Result<(), SvsmError>

Safety

The caller must verify not to corrupt arbitrary memory, as this function doesn’t make any checks in that regard.

Returns

Returns an error if the specified address is not mapped or is not mapped with the appropriate write permissions.

pub unsafe fn write_ref(&self, buf: &T) -> Result<(), SvsmError>

Safety

The caller must verify not to corrupt arbitrary memory, as this function doesn’t make any checks in that regard.

Returns

Returns an error if the specified address is not mapped or is not mapped with the appropriate write permissions.

pub const fn cast<N: Copy>(&self) -> GuestPtr<N>

pub fn offset(&self, count: isize) -> Self