svsm::mm

Module page_visibility

Structs

SharedBox

SharedBox is a safe wrapper around memory pages shared with the host.

Functions

make_page_private⚠

Makes a virtual page private by updating the page tables, modifying the page state, and revalidating the page.

make_page_shared⚠

Makes a virtual page shared by revoking its validation, updating the page state, and modifying the page tables accordingly.