svsm::mm::page_visibility
Function make_page_shared

unsafe fn make_page_shared(vaddr: VirtAddr) -> Result<(), SvsmError>

Makes a virtual page shared by revoking its validation, updating the page state, and modifying the page tables accordingly.

Arguments

vaddr - The virtual address of the page to be made shared.

Safety

Converting the memory at vaddr must be safe within Rust’s memory model. Notably any objects at vaddr must tolerate unsynchronized writes of any bit pattern.