svsm::mm
Module pagebox

Structs

PageBox

An abstraction, similar to a Box, for types that need to be allocated using page allocator directly.

Functions

check_size_requirements 🔒

Check the size requrements for a type to be allocated through PageBox.