svsm::mm::pagebox

Struct PageBox

#[repr(transparent)]
pub struct PageBox<T: ?Sized> {
    ptr: NonNull<T>,
    _phantom: PhantomData<T>,
}

An abstraction, similar to a Box, for types that need to be allocated using page allocator directly.

Constructing a PageBox is very similar to constructing a regular Box:

let p = PageBox::try_new([0u8; 4096])?;

The type guarantees that the allocated memory will have a minimum alignment of the page size, and that memory will be valid until it is dropped.

The type does not support zero sized types nor unsized types. On the other hand, it is able to check at compile time that the contained T can fit in a page allocation with the required alignment. For example, the following will not build because its size exceeds the maximum page order:

let p = PageBox::try_new([0u8; 0x80000])?;

Fields
ptr: NonNull<T>_phantom: PhantomData<T>
Implementations

impl<T> PageBox<T>

pub fn try_new(x: T) -> Result<Self, SvsmError>

Allocates enough pages to hold a T, initializing them with the given value.

pub fn try_new_zeroed() -> Result<PageBox<T>, SvsmError>
where T: FromZeros,

Allocates enough pages to hold a T, and zeroes them out.

const fn get_order() -> usize

Gets the page order required for an allocation to hold a T. It also checks that the size and alignment requirements of the type can be serviced.

pub fn try_new_uninit() -> Result<PageBox<MaybeUninit<T>>, SvsmError>

Allocates enough pages to hold a T, but does not initialize them.

impl<T: ?Sized> PageBox<T>

pub const unsafe fn from_raw(ptr: NonNull<T>) -> Self

Create a PageBox from a previous allocation of the same type.

Safety

The provided pointer must come from a previous use of PageBox (likely through leak(), and must not be aliased.

pub fn leak<'a>(b: Self) -> &'a mut T

Consumes and leaks the PageBox, returning a mutable reference. The contents will never be freed unless the mutable reference is converted back to a PageBox via from_raw().

pub fn vaddr(&self) -> VirtAddr

Returns the virtual address of this allocation.

impl<T> PageBox<MaybeUninit<T>>

pub unsafe fn assume_init(self) -> PageBox<T>

Transforms a PageBox<MaybeUninit<T>> into a PageBox<T>.

Safety

See the safety requirements for MaybeUninit::assume_init().

impl<T: Clone> PageBox<[T]>

pub fn try_new_slice(val: T, len: NonZeroUsize) -> Result<Self, SvsmError>

Allocates a dynamically-sized slice of len items of type T, and populates it with the given value. The slice cannot be resized.

pub fn try_new_uninit_slice( len: NonZeroUsize, ) -> Result<PageBox<[MaybeUninit<T>]>, SvsmError>

Allocates a dynamically-sized slice of len uninitialized items of type T. The slice cannot be resized.

impl<T> PageBox<[MaybeUninit<T>]>

pub unsafe fn assume_init_slice(self) -> PageBox<[T]>

Transforms a [PageBox<[MaybeUninit<T>]>] into a [PageBox<[T]>].

Safety

See the safety requirements for MaybeUninit::assume_init().