Module svsm::mm::pagetable

Structs

PTEntry

PTEntryFlags

PTPage

PageTable

PageTablePart

Sub-tree of a page table that can be populated at the top-level used for virtual memory management

PageTableRef

RawPageTablePart 🔒

Represents a sub-tree of a page-table which can be mapped at a top-level index

Enums

Mapping

Constants

ENTRY_COUNT 🔒

LAUNCH_VMSA_ADDR

Statics

FEATURE_MASK 🔒

INIT_PGTABLE 🔒

MAX_PHYS_ADDR 🔒

PRIVATE_PTE_MASK 🔒

SHARED_PTE_MASK 🔒

Functions

get_init_pgtable_locked

init_encrypt_mask 🔒

make_private_address 🔒

make_shared_address 🔒

max_phys_addr

Returns the exclusive end of the physical address space.

paging_init

paging_init_early

private_pte_mask 🔒

set_init_pgtable

shared_pte_mask 🔒

strip_confidentiality_bits 🔒

supported_flags 🔒