svsm::mm::pagetable

Struct PTEntry

#[repr(C)]
pub struct PTEntry(PhysAddr);

Represents a page table entry.

Tuple Fields
0: PhysAddr
Implementations

impl PTEntry

pub fn is_clear(&self) -> bool

Check if the page table entry is clear (null).

pub fn clear(&mut self)

Clear the page table entry.

pub fn present(&self) -> bool

Check if the page table entry is present.

pub fn huge(&self) -> bool

Check if the page table entry is huge.

pub fn writable(&self) -> bool

Check if the page table entry is writable.

pub fn nx(&self) -> bool

Check if the page table entry is NX (no-execute).

pub fn user(&self) -> bool

Check if the page table entry is user-accessible.

pub fn global(&self) -> bool

Check if the page table entry is global.

pub fn has_reserved_bits(&self, pm: PagingMode, level: usize) -> bool

Check if the page table entry has reserved bits set.

pub fn raw(&self) -> u64

Get the raw bits (u64) of the page table entry.

pub fn flags(&self) -> PTEntryFlags

Get the flags of the page table entry.

pub fn set(&mut self, addr: PhysAddr, flags: PTEntryFlags)

Set the page table entry with the specified address and flags.

pub fn page_frame(&self) -> PhysAddr

Get the address from the page table entry, including the shared bit.

pub fn address(&self) -> PhysAddr

Get the address from the page table entry, excluding the C/shared bit.

pub unsafe fn read_pte(vaddr: VirtAddr) -> Self

Read a page table entry from the specified virtual address.

Safety

Reads from an arbitrary virtual address, making this essentially a raw pointer read. The caller must be certain to calculate the correct address.