Struct svsm::mm::pagetable::PageTablePart

pub struct PageTablePart {
    raw: Option<Box<RawPageTablePart>>,
    idx: usize,
}

Sub-tree of a page table that can be populated at the top-level used for virtual memory management

Fields
raw: Option<Box<RawPageTablePart>>

The root of the page-table sub-tree

idx: usize

The top-level index this PageTablePart is populated at

Implementations

impl PageTablePart

pub fn new(start: VirtAddr) -> Self

Create a new PageTablePart and allocate a root page for the page-table sub-tree.

Arguments

start: Virtual start address this PageTablePart maps

Returns

A new instance of PageTablePart

pub fn alloc(&mut self)

fn get_or_init_mut(&mut self) -> &mut RawPageTablePart

fn get_mut(&mut self) -> Option<&mut RawPageTablePart>

fn get(&self) -> Option<&RawPageTablePart>

pub fn index(&self) -> usize

Request PageTable index to populate this instance to

Returns

Index of the top-level PageTable this sub-tree is populated to

pub fn address(&self) -> Option<PhysAddr>

Request physical base address of the page-table sub-tree. This is needed to populate the PageTablePart.

Returns

Physical base address of the page-table sub-tree

pub fn map_4k( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: PTEntryFlags, shared: bool, ) -> Result<(), SvsmError>

Map a 4KiB page in the page table sub-tree

Arguments

vaddr - Virtual address to create the mapping. Must be aligned to 4KiB.

paddr - Physical address to map. Must be aligned to 4KiB.

flags - PTEntryFlags used for the mapping

shared - Defines whether the page is mapped shared or private

Returns

OK(()) on Success, Err(SvsmError::Mem) on error.

This function can fail when there not enough memory to allocate pages for the mapping.

Panics

This method panics when either vaddr or paddr are not aligned to 4KiB.

pub fn unmap_4k(&mut self, vaddr: VirtAddr) -> Option<PTEntry>

Unmaps a 4KiB page from the page table sub-tree

Arguments

vaddr - The virtual address to unmap. Must be aligned to 4KiB.

Returns

Returns a copy of the PTEntry that mapped the virtual address, if any.

Panics

This method panics when vaddr is not aligned to 4KiB.

pub fn map_2m( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: PTEntryFlags, shared: bool, ) -> Result<(), SvsmError>

Map a 2MiB page in the page table sub-tree

Arguments

vaddr - Virtual address to create the mapping. Must be aligned to 2MiB.

paddr - Physical address to map. Must be aligned to 2MiB.

flags - PTEntryFlags used for the mapping

shared - Defines whether the page is mapped shared or private

Returns

OK(()) on Success, Err(SvsmError::Mem) on error.

This function can fail when there not enough memory to allocate pages for the mapping.

Panics

This method panics when either vaddr or paddr are not aligned to 2MiB.

pub fn unmap_2m(&mut self, vaddr: VirtAddr) -> Option<PTEntry>

Unmaps a 2MiB page from the page table sub-tree

Arguments

vaddr - The virtual address to unmap. Must be aligned to 2MiB.

Returns

Returns a copy of the PTEntry that mapped the virtual address, if any.

Panics

This method panics when vaddr is not aligned to 2MiB.