svsm::mm

Module transition_pgtable

Structs

TransitionPageTable

The TransitionPageTable structure represents a temporary set of page tables to be used by secondary processors as they are started. It comprises the paging root that is used for SVSM kernel initialization (which is not needed after the per-CPU page tables are created) and it maps the low 4 GB of memory with an identity map, as well as the global virtual address range.