Trait svsm::mm::vm::mapping::api::VirtualMapping

pub trait VirtualMapping: Debug {
    // Required methods
    fn mapping_size(&self) -> usize;
    fn map(&self, offset: usize) -> Option<PhysAddr>;
    fn pt_flags(&self, offset: usize) -> PTEntryFlags;

    // Provided methods
    fn has_data(&self) -> bool { ... }
    fn unmap(&self, _offset: usize) { ... }
    fn page_size(&self) -> PageSize { ... }
    fn shared(&self) -> bool { ... }
    fn handle_page_fault(
        &mut self,
        _vmr: &VMR,
        _offset: usize,
        _write: bool,
    ) -> Result<VMPageFaultResolution, SvsmError> { ... }
}

Required Methods

fn mapping_size(&self) -> usize

Request the size of the virtual memory mapping

Returns

Mapping size. Will always be a multiple of VirtualMapping::page_size()

fn map(&self, offset: usize) -> Option<PhysAddr>

Request physical address to map for a given offset

Arguments

offset - Offset into the virtual memory mapping

Returns

Physical address to map for the given offset, if any. None is also a valid return value and does not indicate an error.

fn pt_flags(&self, offset: usize) -> PTEntryFlags

Request the PTEntryFlags used for this virtual memory mapping.

Arguments

‘offset’ -> The offset in bytes into the VirtualMapping. The flags returned from this function relate to the page at the given offset

Returns

A combination of:

PTEntryFlags::WRITABLE

PTEntryFlags::NX,

PTEntryFlags::ACCESSED

PTEntryFlags::DIRTY

Provided Methods

fn has_data(&self) -> bool

Indicates whether the mapping has any associated data.

Returns

true' if there is associated physical data, or false’ if there is none.

fn unmap(&self, _offset: usize)

Inform the virtual memory mapping about an offset being unmapped. Implementing unmap() is optional.

Arguments

_offset

fn page_size(&self) -> PageSize

Request the page size used for mappings

Returns

Either PAGE_SIZE or PAGE_SIZE_2M

fn shared(&self) -> bool

Request whether the mapping is shared or private. Defaults to private unless overwritten by the specific type.

Returns

True - When mapping is shared

False - When mapping is private

fn handle_page_fault( &mut self, _vmr: &VMR, _offset: usize, _write: bool, ) -> Result<VMPageFaultResolution, SvsmError>

Handle a page fault that occurred on a virtual memory address within this mapping.

Arguments

‘vmr’ - Virtual memory range that contains the mapping. This VirtualMapping can use this to insert/remove regions as necessary to handle the page fault.

offset - Offset into the virtual mapping that was the subject of the page fault.

‘write’ - true if the fault was due to a write to the memory location, or ‘false’ if the fault was due to a read.