svsm::mm::vm::mapping::kernel_stack

Struct VMKernelStack

pub struct VMKernelStack {
    alloc: RawAllocMapping,
    guard_pages: usize,
    shadow: bool,
}

Mapping to be used as a kernel stack. This maps a stack including guard pages at the top and bottom.

Fields
alloc: RawAllocMapping

Allocation for stack pages

guard_pages: usize

Number of guard pages to reserve address space for

shadow: bool

True for shadow stacks

Implementations

impl VMKernelStack

pub fn top_of_stack(&self) -> usize

Returns the virtual address for the top of this kernel stack

Arguments

base - Virtual base address this stack is mapped at (including guard pages).

Returns

Offset from start of virtual mapping to top-of-stack.

pub fn bounds(&self, base: VirtAddr) -> MemoryRegion<VirtAddr>

Returns the stack bounds of this kernel stack

Arguments

base - Virtual base address this stack is mapped at (including guard pages).

Returns

A MemoryRegion object containing the bottom and top addresses for the stack

fn new_size(size: usize, shadow: bool) -> Result<Self, SvsmError>

Create a new VMKernelStack with a given size. This function will already allocate the backing pages for the stack.

Arguments

size - Size of the kernel stack, without guard pages

shadow - Whether this mapping is for a shadow stack

Returns

Initialized stack on success, Err(SvsmError::Mem) on error

pub fn new() -> Result<Self, SvsmError>

Create a new VMKernelStack with the default size. This function will already allocate the backing pages for the stack.

Returns

Initialized stack on success, Err(SvsmError::Mem) on error

pub fn new_shadow() -> Result<Self, SvsmError>

Create a new VMKernelStack with one page in size for use as a shadow stack.

Returns

Shadow stack on success, Err(SvsmError::Mem) on error

pub fn shadow_page(&self) -> PageRef

Get a reference to the shadow stack page. This is required because shadow stacks are mapped read-only and a separate mapping needs to be created to initialize it.

Returns

A PageRef reference to the shadow stack page

Panics

The function panics if called on a non-shadow-stack object.

pub fn new_mapping() -> Result<Mapping, SvsmError>

Create a new VMKernelStack with the default size, packed into a Mapping. This function will allocate the backing pages for the stack.

Returns

Initialized Mapping to stack on success, Err(SvsmError::Mem) on error

fn alloc_pages(&mut self) -> Result<(), SvsmError>