svsm::mm::vm

Struct RawAllocMapping

pub struct RawAllocMapping {
    pages: Vec<Option<PageRef>>,
    count: usize,
}

Contains base functionality for all VirtualMapping types which use self-allocated PageFile pages.

Fields
pages: Vec<Option<PageRef>>

A vec containing references to PageFile allocations

count: usize

Number of pages required in pages

Implementations

impl RawAllocMapping

pub fn new(size: usize) -> Self

Creates a new instance of RawAllocMapping

Arguments

size - Size of the mapping in bytes

Returns

New instance of RawAllocMapping. Still needs to call alloc_pages() on it before it can be used.

pub fn alloc_page(&mut self, offset: usize) -> Result<(), SvsmError>

Allocates a single backing page of type PageFile if the page has not already been allocated

Argument

‘offset’ - The offset in bytes from the start of the mapping

Returns

Ok(()) if the page has been allocated, Err(SvsmError::Mem) otherwise

pub fn alloc_pages(&mut self) -> Result<(), SvsmError>

Allocates a full set of backing pages of type PageFile

Returns

Ok(()) when all pages could be allocated, Err(SvsmError::Mem) otherwise

pub fn page(&self, index: usize) -> PageRef

Returns a reference to a page at a given index. The page must already been allocated.

Arguments

index - Page index to reference

Returns

A reference to the requested page.

Panics

This function panics if an invalid index is accessed or the page at the index has not been allocated.

pub fn mapping_size(&self) -> usize

Request size of the mapping in bytes

Returns

The size of the mapping in bytes as usize.

pub fn map(&self, offset: usize) -> Option<PhysAddr>

Request physical address to map for a given offset

Arguments

offset - Byte offset into the memory mapping

Returns

Physical address to map for the given offset.

pub fn unmap(&self, _offset: usize)

Unmap call-back - currently nothing to do in this function

Arguments

_offset - Byte offset into the mapping

pub fn present(&self, offset: usize) -> bool

Check if a page has been allocated

Arguments

‘offset’ - Byte offset into the mapping

Returns

‘true’ if the page containing the offset has been allocated otherwise ‘false’.