svsm::mm::vm

Struct VMR

pub struct VMR {
    start_pfn: usize,
    end_pfn: usize,
    tree: RWLock<RBTree<VMMAdapter>>,
    pgtbl_parts: RWLock<Vec<PageTablePart>>,
    pt_flags: PTEntryFlags,
    per_cpu: bool,
}

Virtual Memory Region

This struct manages the mappings in a region of the virtual address space. The region size is a multiple of 512GiB so that every region will fully allocate one or more top-level page-table entries on x86-64. For the same reason the start address must also be aligned to 512GB.

Fields
start_pfn: usize

Start address of this range as virtual PFN (VirtAddr >> PAGE_SHIFT). Virtual address must be aligned to VMR_GRANULE (512GB on x86-64).

end_pfn: usize

End address of this range as virtual PFN (VirtAddr >> PAGE_SHIFT) Virtual address must be aligned to VMR_GRANULE (512GB on x86-64).

tree: RWLock<RBTree<VMMAdapter>>

RBTree containing all [struct VMM] instances with valid mappings in the covered virtual address region. The [struct VMM]s are sorted by their start address and stored in an RBTree for faster lookup.

pgtbl_parts: RWLock<Vec<PageTablePart>>

[struct PageTableParts] needed to map this VMR into a page-table. There is one [struct PageTablePart] per VMR_GRANULE covered by the region.

pt_flags: PTEntryFlags

PTEntryFlags global to all mappings in this region. This is a combination of PTEntryFlags::GLOBAL and PTEntryFlags::USER.

per_cpu: bool

Indicates that this [struct VMR] is visible only on a single CPU and therefore TLB flushes do not require broadcast.

Implementations

impl VMR

pub fn new(start: VirtAddr, end: VirtAddr, flags: PTEntryFlags) -> Self

Creates a new [struct VMR]

Arguments

start - Virtual start address for the memory region. Must be aligned to VMR_GRANULE

end - Virtual end address (non-inclusive) for the memory region. Must be bigger than start and aligned to VMR_GRANULE.

flags - Global PTEntryFlags to use for this [struct VMR].

Returns

A new instance of [struct VMR].

pub fn set_per_cpu(&mut self, per_cpu: bool)

Marks a [struct VMR] as being associated with only a single CPU so that TLB flushes do not require broadcast.

fn alloc_page_tables(&self, lazy: bool) -> Result<(), SvsmError>

Allocated all PageTableParts needed to map this region

Returns

Ok(()) on success, Err(SvsmError::Mem) on allocation error

pub fn populate(&self, pgtbl: &mut PageTable)

Populate PageTableParts of the VMR into a page-table

Arguments

pgtbl - A PageTable pointing to the target page-table

pub fn populate_addr(&self, pgtbl: &mut PageTable, vaddr: VirtAddr)

unsafe fn initialize_common(&self, lazy: bool) -> Result<(), SvsmError>

Initialize this VMR by checking the start and end values and allocating the PageTableParts required for the mappings.

Safety

Callers must ensure that the bounds of the address range are appropriately aligned to prevent the possibility that adjacent address ranges may attempt to share top-level paging entries. If any overlap is attempted, page tables may be corrupted.

Arguments

lazy - When true, use lazy allocation of PageTablePart pages.

Returns

Ok(()) on success, Err(SvsmError::Mem) on allocation error

pub unsafe fn initialize(&self) -> Result<(), SvsmError>

Initialize this VMR by calling VMR::initialize_common with lazy = false

Safety

Callers must ensure that the bounds of the address range are appropriately aligned to prevent the possibility that adjacent address ranges may attempt to share top-level paging entries. If any overlap is attempted, page tables may be corrupted.

Returns

Ok(()) on success, Err(SvsmError::Mem) on allocation error

pub unsafe fn initialize_lazy(&self) -> Result<(), SvsmError>

Initialize this VMR by calling VMR::initialize_common with lazy = true

Safety

Callers must ensure that the bounds of the address range are appropriately aligned to prevent the possibility that adjacent address ranges may attempt to share top-level paging entries. If any overlap is attempted, page tables may be corrupted.

Returns

Ok(()) on success, Err(SvsmError::Mem) on allocation error

fn virt_range(&self) -> (VirtAddr, VirtAddr)

Returns the virtual start and end addresses for this region

Returns

Tuple containing start and end virtual address of the memory region

fn map_vmm(&self, vmm: &VMM) -> Result<(), SvsmError>

Map a VMM into the PageTableParts of this region

Arguments

vmm - Reference to a VMM instance to map into the page-table

Returns

Ok(()) on success, Err(SvsmError::Mem) on allocation error

fn unmap_vmm(&self, vmm: &VMM)

Unmap a VMM from the PageTableParts of this region

Arguments

vmm - Reference to a VMM instance to unmap from the page-table

fn do_insert( &self, mapping: Arc<Mapping>, start_pfn: usize, cursor: &mut CursorMut<'_, VMMAdapter>, ) -> Result<(), SvsmError>

pub fn insert_at( &self, vaddr: VirtAddr, mapping: Arc<Mapping>, ) -> Result<VirtAddr, SvsmError>

Inserts VMM at a specified virtual base address. This method checks that the VMM does not overlap with any other region.

Arguments

vaddr - Virtual base address to map the VMM at

mapping - Rc pointer to the VMM to insert

Returns

Base address where the VMM was inserted on success or SvsmError::Mem on error

pub fn insert_aligned( &self, hint: VirtAddr, mapping: Arc<Mapping>, align: usize, ) -> Result<VirtAddr, SvsmError>

Inserts VMM with the specified alignment. This method walks the RBTree to search for a suitable region.

Arguments

mapping - Rc pointer to the VMM to insert

align - Alignment to use for tha mapping

Returns

Base address where the VMM was inserted on success or SvsmError::Mem on error

pub fn insert_hint( &self, addr: VirtAddr, mapping: Arc<Mapping>, ) -> Result<VirtAddr, SvsmError>

Inserts VMM into the virtual memory region. This method takes the next power-of-two larger of the mapping size and uses that as the alignment for the mappings base address. The search for the base address starts at addr. With that it calls VMR::insert_aligned.

Arguments

addr - The virtual address at which the search for a mapping area starts

mapping - Arc pointer to the VMM to insert

Returns

Base address where the VMM was inserted on success or SvsmError::Mem on error

pub fn insert(&self, mapping: Arc<Mapping>) -> Result<VirtAddr, SvsmError>

Inserts VMM into the virtual memory region. It searches from the beginning of the VMR region for a suitable slot.

Arguments

mapping - Rc pointer to the VMM to insert

Returns

Base address where the VMM was inserted on success or SvsmError::Mem on error

pub fn remove(&self, base: VirtAddr) -> Result<Box<VMM>, SvsmError>

Removes the mapping from a given base address from the RBTree

Arguments

base - Virtual base address of the VMM to remove

Returns

The removed mapping on success, SvsmError::Mem on error

pub fn dump_ranges(&self)

Dump all VMM mappings in the RBTree. This function is included for debugging purposes. And should not be called in production code.

pub fn handle_page_fault( &self, vaddr: VirtAddr, _write: bool, ) -> Result<(), SvsmError>

Notify the range that a page fault has occurred. This should be called from the page fault handler. The mappings withing this virtual memory region are examined and if they overlap with the page fault address then VMR::handle_page_fault() is called to handle the page fault within that range.

Arguments

vaddr - Virtual memory address that was the subject of the page fault

‘write’ - ‘true’ if a write was attempted. ‘false’ if a read was attempted.

Returns

‘()’ if the page fault was successfully handled.

‘SvsmError::Mem’ if the page fault should propogate to the next handler.