svsm

Module persistence

Functionality related to persistent SVSM storage.

Structs

PersistenceBootstrapInfo

Persistence metadata info returned by persistence_discover().

SvsmNvBlkDev 🔒

Wrapper around BlockDriver implementors, itself implementing the cocoon-tpm-storage crate’s [NvBlkDev] block device abstraction.

SvsmNvBlkDevFlushQueuedWritesFuture 🔒

[NvBlkDev::FlushQueuedWritesFuture] implementation for SvsmNvBlkDev.

SvsmNvBlkDevReadFuture 🔒

[NvBlkDev::ReadFuture] implementation for SvsmNvBlkDev.

SvsmNvBlkDevResizeFuture 🔒

[NvBlkDev::ResizeFuture] implementation for SvsmNvBlkDev.

SvsmNvBlkDevTrimFuture 🔒

[NvBlkDev::TrimFuture] implementation for SvsmNvBlkDev.

SvsmNvBlkDevWriteFuture 🔒

[NvBlkDev::WriteFuture] implementation for SvsmNvBlkDev.

SvsmNvBlkDevWriteSyncFuture 🔒

[NvBlkDev::WriteSyncFuture] implementation for SvsmNvBlkDev.

Enums

InodeNamespace

The 64-bit CocoonFS inode number is split into two halves: the upper 32 bits are statically assigned to each service here, and the lower 32 bits are managed by the service itself. Namespace 0 is reserved: although CocoonFS currently only reserves the first 16 inodes for internal use, all inodes with the upper 32 bits set to zero are kept reserved for simplicity.

Statics

SVSM_COCOONFS_INSTANCE 🔒

Traits

Inode

Trait for subsystem-specific inode enums.

Functions

instantiate_cocoonfs_commit_transaction_fut 🔒

Instantiate a [CocoonFs::CommitTransactionFut].

instantiate_cocoonfs_open_fut 🔒

Instantiate a [cocoonfs::OpenFsFuture].

instantiate_cocoonfs_read_fs_metadata_fut 🔒

Instantiate a [cocoonfs::ReadFsMetadataFuture].

instantiate_cocoonfs_read_inode_fut 🔒

Instantiate a [CocoonFs::ReadInodeFut].

instantiate_cocoonfs_start_transaction_fut 🔒

Instantiate a [CocoonFs::StartTransactionFut].

instantiate_cocoonfs_write_inode_fut 🔒

Instantiate a [CocoonFs::WriteInodeFut].

nvfs_error_to_svsm_error 🔒

persistence_available

Test whether persistence functionality is available.

persistence_discover

Obtain persistence bootstrap info from block devices, if any.

persistence_init

Finalize the initialization of the persistence subsystem.

persistence_read_inode_sync

Synchronously read data from an inode on persistent storage.

persistence_write_inode_sync

Synchronously write data to an inode on persistent storage.

svsm_error_to_nvblkdev_io_error 🔒

Convert a SvsmError to a [NvBlkDevIoError].

Type Aliases

SvsmCocoonFsSyncRcPtrRefType 🔒

SvsmCocoonFsSyncRcPtrType 🔒

SvsmCocoonFsType 🔒