svsm

Module platform

Modules

capabilities

guest_cpu

native

snp

snp_fw 🔒

tdp

Structs

PageEncryptionMasks

SevFWMetaData

Enums

PageStateChangeOp

PageValidateOp

Stage2PlatformCell

SvsmPlatformCell

Statics

CAPS

SVSM_PLATFORM

SVSM_PLATFORM_CELL 🔒

SVSM_PLATFORM_TYPE 🔒

Traits

Stage2Platform

This defines a platform abstraction to permit stage2 to run on different underlying architectures. It includes only functionality that is used only in stage2.

SvsmPlatform

This defines a platform abstraction to permit the SVSM to run on different underlying architectures.

Functions

halt

init_capabilities

init_platform_type

terminate

Terminates the guest with a platform-specific mechanism