svsm::platform

Trait Stage2Platform

pub trait Stage2Platform {
    // Provided methods
    fn get_cpuid_page(
        &self,
        _launch_info: &Stage2LaunchInfo,
    ) -> Option<VirtAddr> { ... }
    fn get_secrets_page(
        &self,
        _launch_info: &Stage2LaunchInfo,
    ) -> Option<VirtAddr> { ... }
}

This defines a platform abstraction to permit stage2 to run on different underlying architectures. It includes only functionality that is used only in stage2.

Provided Methods

fn get_cpuid_page(&self, _launch_info: &Stage2LaunchInfo) -> Option<VirtAddr>

Obtains the virtual address of the CPUID page, if any.

fn get_secrets_page(&self, _launch_info: &Stage2LaunchInfo) -> Option<VirtAddr>

Obtains the virtual address of the secrets page, if any.