svsm::platform
Trait SvsmPlatform

pub trait SvsmPlatform {

    // Required methods
    fn env_setup(
        &mut self,
        debug_serial_port: u16,
        vtom: usize,
    ) -> Result<(), SvsmError>;
    fn env_setup_late(
        &mut self,
        debug_serial_port: u16,
    ) -> Result<(), SvsmError>;
    fn env_setup_svsm(&self) -> Result<(), SvsmError>;
    fn setup_percpu(&self, cpu: &PerCpu) -> Result<(), SvsmError>;
    fn setup_percpu_current(&self, cpu: &PerCpu) -> Result<(), SvsmError>;
    fn get_page_encryption_masks(&self) -> PageEncryptionMasks;
    fn capabilities(&self) -> Caps;
    fn cpuid(&self, eax: u32) -> Option<CpuidResult>;
    fn setup_guest_host_comm(&mut self, cpu: &PerCpu, is_bsp: bool);
    fn get_io_port(&self) -> &'static dyn IOPort;
    fn page_state_change(
        &self,
        region: MemoryRegion<PhysAddr>,
        size: PageSize,
        op: PageStateChangeOp,
    ) -> Result<(), SvsmError>;
    fn validate_physical_page_range(
        &self,
        region: MemoryRegion<PhysAddr>,
        op: PageValidateOp,
    ) -> Result<(), SvsmError>;
    fn validate_virtual_page_range(
        &self,
        region: MemoryRegion<VirtAddr>,
        op: PageValidateOp,
    ) -> Result<(), SvsmError>;
    fn configure_alternate_injection(
        &mut self,
        alt_inj_requested: bool,
    ) -> Result<(), SvsmError>;
    fn change_apic_registration_state(
        &self,
        incr: bool,
    ) -> Result<bool, SvsmError>;
    fn query_apic_registration_state(&self) -> bool;
    fn use_interrupts(&self) -> bool;
    fn post_irq(&self, icr: u64) -> Result<(), SvsmError>;
    fn eoi(&self);
    fn is_external_interrupt(&self, vector: usize) -> bool;
    fn start_cpu(&self, cpu: &PerCpu, start_rip: u64) -> Result<(), SvsmError>;

    // Provided methods
    fn halt()
       where Self: Sized { ... }
    fn prepare_fw(
        &self,
        _config: &SvsmConfig<'_>,
        _kernel_region: MemoryRegion<PhysAddr>,
    ) -> Result<(), SvsmError> { ... }
    fn launch_fw(&self, _config: &SvsmConfig<'_>) -> Result<(), SvsmError> { ... }
    fn determine_cet_support(&self) -> bool { ... }
    fn flush_tlb(&self, flush_scope: &TlbFlushScope) { ... }
    fn start_svsm_request_loop(&self) -> bool { ... }

}

This defines a platform abstraction to permit the SVSM to run on different underlying architectures.

Required Methods

fn env_setup( &mut self, debug_serial_port: u16, vtom: usize, ) -> Result<(), SvsmError>

Performs basic early initialization of the runtime environment.

fn env_setup_late(&mut self, debug_serial_port: u16) -> Result<(), SvsmError>

Performs initialization of the platform runtime environment after the core system environment has been initialized.

fn env_setup_svsm(&self) -> Result<(), SvsmError>

Performs initialiation of the environment specfic to the SVSM kernel (for services not used by stage2).

fn setup_percpu(&self, cpu: &PerCpu) -> Result<(), SvsmError>

Completes initialization of a per-CPU object during construction.

fn setup_percpu_current(&self, cpu: &PerCpu) -> Result<(), SvsmError>

Completes initialization of a per-CPU object on the target CPU.

fn get_page_encryption_masks(&self) -> PageEncryptionMasks

Determines the paging encryption masks for the current architecture.

fn capabilities(&self) -> Caps

Get the features and the capabilities of the platform.

fn cpuid(&self, eax: u32) -> Option<CpuidResult>

Obtain CPUID using platform-specific tables.

fn setup_guest_host_comm(&mut self, cpu: &PerCpu, is_bsp: bool)

Establishes state required for guest/host communication.

fn get_io_port(&self) -> &'static dyn IOPort

Obtains a reference to an I/O port implemetation appropriate to the platform.

fn page_state_change( &self, region: MemoryRegion<PhysAddr>, size: PageSize, op: PageStateChangeOp, ) -> Result<(), SvsmError>

Performs a page state change between private and shared states.

fn validate_physical_page_range( &self, region: MemoryRegion<PhysAddr>, op: PageValidateOp, ) -> Result<(), SvsmError>

Marks a physical range of pages as valid or invalid for use as private pages. Not usable in stage2.

fn validate_virtual_page_range( &self, region: MemoryRegion<VirtAddr>, op: PageValidateOp, ) -> Result<(), SvsmError>

Marks a virtual range of pages as valid or invalid for use as private pages. Provided primarily for use in stage2 where validation by physical address cannot be supported.

fn configure_alternate_injection( &mut self, alt_inj_requested: bool, ) -> Result<(), SvsmError>

Configures the use of alternate injection as requested.

fn change_apic_registration_state(&self, incr: bool) -> Result<bool, SvsmError>

Changes the state of APIC registration on this system, returning either the current registration state or an error.

fn query_apic_registration_state(&self) -> bool

Queries the state of APIC registration on this system.

fn use_interrupts(&self) -> bool

Determines whether the platform supports interrupts to the SVSM.

fn post_irq(&self, icr: u64) -> Result<(), SvsmError>

Signal an IRQ on one or more CPUs.

fn eoi(&self)

Perform an EOI of the current interrupt.

fn is_external_interrupt(&self, vector: usize) -> bool

Determines whether a given interrupt vector was invoked as an external interrupt.

fn start_cpu(&self, cpu: &PerCpu, start_rip: u64) -> Result<(), SvsmError>

Start an additional processor.

Provided Methods

fn halt()
where Self: Sized,

Halts the system as required by the platform.

fn prepare_fw( &self, _config: &SvsmConfig<'_>, _kernel_region: MemoryRegion<PhysAddr>, ) -> Result<(), SvsmError>

Performs the necessary preparations for launching guest boot firmware.

fn launch_fw(&self, _config: &SvsmConfig<'_>) -> Result<(), SvsmError>

Launches guest boot firmware.

fn determine_cet_support(&self) -> bool

Determine whether shadow stacks are supported.

fn flush_tlb(&self, flush_scope: &TlbFlushScope)

Performs a system-wide TLB flush.

fn start_svsm_request_loop(&self) -> bool

Indicates whether this platform should invoke the SVSM request loop.