svsm::protocols::attest

Struct AttestServicesOp

#[repr(C, packed(1))]
pub struct AttestServicesOp {
    report_gpa: u64,
    report_size: u32,
    reserved_1: [u8; 4],
    nonce_gpa: u64,
    nonce_size: u16,
    reserved_2: [u8; 6],
    manifest_gpa: u64,
    manifest_size: u32,
    reserved_3: [u8; 4],
    certificate_gpa: u64,
    certificate_size: u32,
    reserved_4: [u8; 4],
}

Fields
report_gpa: u64report_size: u32reserved_1: [u8; 4]nonce_gpa: u64nonce_size: u16reserved_2: [u8; 6]manifest_gpa: u64manifest_size: u32reserved_3: [u8; 4]certificate_gpa: u64certificate_size: u32reserved_4: [u8; 4]
Implementations

impl AttestServicesOp

pub fn try_from_as_ref(buffer: &[u8]) -> Result<&Self, SvsmReqError>

Take a slice and return a reference for Self

pub fn is_reserved_clear(&self) -> bool

Checks if reserved fields are all set to zero

pub fn get_nonce(&self) -> Result<Vec<u8>, SvsmReqError>

Returns the nonce

pub fn get_manifest_region( &self, ) -> Result<MemoryRegion<PhysAddr>, SvsmReqError>

Returns the manifest buffer gpa and size Checks if gpa is page aligned and valid. Manifest buffer size can be greater than 4k, so it can cross page boundary.

pub fn get_report_region(&self) -> Result<MemoryRegion<PhysAddr>, SvsmReqError>

Returns the report buffer gpa and size Checks if gpa is page aligned and valid. Report buffer size can be greater than 4k, so it can cross page boundary.