svsm::protocols

Module uefivars

UEFI MM protocol implementation

Process MM communication requests from edk2 firmware.

Usually this request serialization format is used by edk2 for communication between normal mode and management mode (MM for short).

This SVSM protocol is a thin wrapper to allow edk2 firmware send those requests to the SVSM instead. The actual request processing is implemented by the virtfw_varstore crate. The latter implements the MM protocols needed to provide an UEFI variable store.

Structs

UefiMmManifest 🔒

UefiMmManifestFlags

Constants

SVSM_UEFI_MM_REQUEST 🔒

UEFI_MM_BUFFER_LIMIT 🔒

UEFI_MM_PROTOCOL_VERSION_MAX

UEFI_MM_PROTOCOL_VERSION_MIN

Statics

STORE 🔒

Functions

check_buffer 🔒

uefi_mm_get_manifest

uefi_mm_protocol_init

uefi_mm_protocol_request

uefi_mm_request 🔒