svsm::sev::ghcb

Struct GHCB

#[repr(C)]
pub struct GHCB {

    reserved_1: [AtomicU8; 203],
    cpl: AtomicU8,
    reserved_2: [AtomicU8; 116],
    xss: AtomicU64,
    reserved_3: [AtomicU8; 24],
    dr7: AtomicU64,
    reserved_4: [AtomicU8; 144],
    rax: AtomicU64,
    reserved_5: [AtomicU8; 256],
    reserved_6: AtomicU64,
    rcx: AtomicU64,
    rdx: AtomicU64,
    rbx: AtomicU64,
    reserved_7: AtomicU64,
    rbp: AtomicU64,
    rsi: AtomicU64,
    rdi: AtomicU64,
    r8: AtomicU64,
    r9: AtomicU64,
    r10: AtomicU64,
    r11: AtomicU64,
    r12: AtomicU64,
    r13: AtomicU64,
    r14: AtomicU64,
    r15: AtomicU64,
    reserved_8: [AtomicU8; 16],
    sw_exit_code: AtomicU64,
    sw_exit_info_1: AtomicU64,
    sw_exit_info_2: AtomicU64,
    sw_scratch: AtomicU64,
    reserved_9: [AtomicU8; 56],
    xcr0: AtomicU64,
    valid_bitmap: [AtomicU64; 2],
    x87_state_gpa: AtomicU64,
    reserved_10: [AtomicU8; 1016],
    buffer: [AtomicU8; 2032],
    reserved_11: [AtomicU8; 10],
    version: AtomicU16,
    usage: AtomicU32,

}

Fields
reserved_1: [AtomicU8; 203]cpl: AtomicU8reserved_2: [AtomicU8; 116]xss: AtomicU64reserved_3: [AtomicU8; 24]dr7: AtomicU64reserved_4: [AtomicU8; 144]rax: AtomicU64reserved_5: [AtomicU8; 256]reserved_6: AtomicU64rcx: AtomicU64rdx: AtomicU64rbx: AtomicU64reserved_7: AtomicU64rbp: AtomicU64rsi: AtomicU64rdi: AtomicU64r8: AtomicU64r9: AtomicU64r10: AtomicU64r11: AtomicU64r12: AtomicU64r13: AtomicU64r14: AtomicU64r15: AtomicU64reserved_8: [AtomicU8; 16]sw_exit_code: AtomicU64sw_exit_info_1: AtomicU64sw_exit_info_2: AtomicU64sw_scratch: AtomicU64reserved_9: [AtomicU8; 56]xcr0: AtomicU64valid_bitmap: [AtomicU64; 2]x87_state_gpa: AtomicU64reserved_10: [AtomicU8; 1016]buffer: [AtomicU8; 2032]reserved_11: [AtomicU8; 10]version: AtomicU16usage: AtomicU32
Implementations

impl GHCB

fn get_cpl_valid(&self) -> Result<u8, GhcbError>

fn set_cpl_valid(&self, val: u8)

fn get_xss_valid(&self) -> Result<u64, GhcbError>

fn set_xss_valid(&self, val: u64)

fn get_rax_valid(&self) -> Result<u64, GhcbError>

fn set_rax_valid(&self, val: u64)

fn get_rcx_valid(&self) -> Result<u64, GhcbError>

fn set_rcx_valid(&self, val: u64)

fn get_rdx_valid(&self) -> Result<u64, GhcbError>

fn set_rdx_valid(&self, val: u64)

fn get_rbx_valid(&self) -> Result<u64, GhcbError>

fn set_rbx_valid(&self, val: u64)

fn get_r8_valid(&self) -> Result<u64, GhcbError>

fn set_r8_valid(&self, val: u64)

fn get_exit_code_valid(&self) -> Result<u64, GhcbError>

fn set_exit_code_valid(&self, val: u64)

fn get_exit_info_1_valid(&self) -> Result<u64, GhcbError>

fn set_exit_info_1_valid(&self, val: u64)

fn get_exit_info_2_valid(&self) -> Result<u64, GhcbError>

fn set_exit_info_2_valid(&self, val: u64)

fn get_sw_scratch_valid(&self) -> Result<u64, GhcbError>

fn set_sw_scratch_valid(&self, val: u64)

fn get_sw_xcr0_valid(&self) -> Result<u64, GhcbError>

fn set_sw_xcr0_valid(&self, val: u64)

fn get_sw_x87_state_gpa_valid(&self) -> Result<u64, GhcbError>

fn set_sw_x87_state_gpa_valid(&self, val: u64)

fn get_version_valid(&self) -> Result<u16, GhcbError>

fn set_version_valid(&self, val: u16)

fn get_usage_valid(&self) -> Result<u32, GhcbError>

fn set_usage_valid(&self, val: u32)

pub fn vmmcall(&self, regs: &mut X86GeneralRegs) -> Result<(), SvsmError>

pub fn cpuid(&self, eax: u32, ecx: u32) -> Result<CpuidResult, SvsmError>

pub fn cpuid_regs(&self, regs: &mut X86GeneralRegs) -> Result<(), SvsmError>

pub fn rdtscp_regs(&self, regs: &mut X86GeneralRegs) -> Result<(), SvsmError>

pub fn rdtsc_regs(&self, regs: &mut X86GeneralRegs) -> Result<(), SvsmError>

pub fn wrmsr(&self, msr_index: u32, value: u64) -> Result<(), SvsmError>

pub fn wrmsr_regs(&self, regs: &X86GeneralRegs) -> Result<(), SvsmError>

pub fn wrmsr_raw(&self, rcx: u64, rax: u64, rdx: u64) -> Result<(), SvsmError>

pub fn rdmsr_raw(&self, msr_index: u32) -> Result<(u32, u32), SvsmError>

Perform an MSR read via a GHCB call

Arguments

msr_index - MSR to read

Returns

On a successful call, returns Result with a tuple (u32, u32). The first value is the RAX and the second the RDX register. On failure a Result with SvsmError is returned.

pub fn rdmsr(&self, msr_index: u32) -> Result<u64, SvsmError>

Perform an MSR read via a GHCB call an return a u64 value.

Arguments

msr_index - MSR to read

Returns

On a successful call, returns Result with a u64 value. On failure a Result with SvsmError is returned.

pub fn rdmsr_regs(&self, regs: &mut X86GeneralRegs) -> Result<(), SvsmError>

pub fn register(&self) -> Result<(), SvsmError>

pub fn clear(&self)

fn set_valid(&self, offset: usize)

fn is_valid(&self, offset: usize) -> bool

fn vmgexit( &self, exit_code: GHCBExitCode, exit_info_1: u64, exit_info_2: u64, ) -> Result<(), GhcbError>

pub fn ioio_in(&self, port: u16, size: GHCBIOSize) -> Result<u64, SvsmError>

pub fn ioio_out( &self, port: u16, size: GHCBIOSize, value: u64, ) -> Result<(), SvsmError>

fn write_buffer<T>(&self, data: &T, offset: usize) -> Result<(), GhcbError>
where T: IntoBytes + Immutable,

fn write_buffer_slice( &self, data: &[u8], offset: usize, ) -> Result<(), GhcbError>

fn read_buffer_slice( &self, data: &mut [MaybeUninit<u8>], offset: usize, ) -> Result<(), GhcbError>

fn pepare_buffer(&self)

pub unsafe fn mmio_write( &self, pa: PhysAddr, value: &[u8], ) -> Result<(), SvsmError>

Perfrom a write to a memory-mapped IO area

Safety

Caller must ensure that pa points to a properly aligned memory location and the memory accessed is part of a valid MMIO range.

pub unsafe fn mmio_read( &self, pa: PhysAddr, value: &mut [MaybeUninit<u8>], ) -> Result<(), SvsmError>

Perfrom a read from a memory-mapped IO area

Safety

Caller must ensure that pa points to a properly aligned memory location and the memory accessed is part of a valid MMIO range.

pub fn psc_entry( &self, paddr: PhysAddr, op_mask: u64, current_page: u64, size: PageSize, ) -> u64

pub fn page_state_change( &self, region: MemoryRegion<PhysAddr>, size: PageSize, op: PageStateChangeOp, ) -> Result<(), SvsmError>

pub fn ap_create( &self, vmsa_gpa: PhysAddr, apic_id: u64, vmpl: u64, sev_features: u64, ) -> Result<(), SvsmError>

pub fn register_guest_vmsa( &self, vmsa_gpa: PhysAddr, apic_id: u64, vmpl: u64, sev_features: u64, ) -> Result<(), SvsmError>

pub fn register_hv_doorbell(&self, paddr: PhysAddr) -> Result<(), SvsmError>

pub fn guest_request( &self, req_page: VirtAddr, resp_page: VirtAddr, ) -> Result<(), SvsmError>

pub fn guest_ext_request( &self, req_page: VirtAddr, resp_page: VirtAddr, data_pages: VirtAddr, data_size: u64, ) -> Result<(), SvsmError>

pub fn hv_ipi(&self, icr: u64) -> Result<(), SvsmError>

pub fn configure_interrupt_injection( &self, vector: usize, ) -> Result<(), SvsmError>

pub fn specific_eoi(&self, vector: u8, vmpl: u8) -> Result<(), SvsmError>

pub fn disable_alternate_injection( &self, tpr: u8, in_intr_shadow: bool, interrupts_enabled: bool, ) -> Result<(), SvsmError>