svsm::sev

Module utils

Structs

RMPFlags

Enums

PvalidateOp

The desired state of the page passed to PVALIDATE.

SevSnpError

Functions

get_dr7

pvalidate⚠

Safety

pvalidate_range⚠

Safety

pvalidate_range_4k 🔒 ⚠

Safety

raw_vmgexit

VMGEXIT operations generally need to be performed with interrupts disabled to ensure that an interrupt cannot cause the GHCB MSR to change prior to exiting to the host. It is the caller’s responsibility to ensure that interrupt handling is configured correctly for the attemtped operation.

raw_vmmcall⚠

Executes the vmmcall instruction.

rmp_adjust⚠

Safety

rmp_clear_guest_vmsa⚠

Safety

rmp_grant_guest_access⚠

Safety

rmp_revoke_guest_access

rmp_set_guest_vmsa⚠

Safety

set_dr7

Sets the dr7 register to the given value