svsm::syscall
Module obj

Structs

ObjHandle

ObjHandle is a unique identifier for an object in the current process. An ObjHandle can be converted to a u32 id which can be used by the user mode to access this object. The passed id from the user mode by syscalls can be converted to an ObjHandle, which is used to access the object in the COCONUT-SVSM kernel.

Enums

ObjError

Traits

Obj

An object represents the type of resource like file, VM, vCPU in the COCONUT-SVSM kernel which can be accessible by the user mode. The Obj trait is defined for such type of resource, which can be used to define the common functionalities of the objects. With the trait bounds of Send and Sync, the objects implementing Obj trait could be sent to another thread and shared between threads safely.

Functions

obj_add

Add an object to the current process and assigns it an ObjHandle.

obj_close

Closes an object identified by its ObjHandle.

obj_get

Retrieves an object by its ObjHandle.