svsm::syscall::obj

Function obj_add

pub fn obj_add(obj: Arc<dyn Obj>) -> Result<ObjHandle, SvsmError>

Add an object to the current process and assigns it an ObjHandle.

Arguments

obj - An Arc<dyn Obj> representing the object to be added.

Returns

Result<ObjHandle, SvsmError> - Returns the object handle of the added object if successful, or an SvsmError on failure.

Errors

This function will return an error if adding the object to the current task fails.