svsm::syscall
Trait Obj

pub trait Obj:
    Send
    + Sync
    + Debug {
    // Provided method
    fn as_fs(&self) -> Option<&FsObj> { ... }
}

An object represents the type of resource like file, VM, vCPU in the COCONUT-SVSM kernel which can be accessible by the user mode. The Obj trait is defined for such type of resource, which can be used to define the common functionalities of the objects. With the trait bounds of Send and Sync, the objects implementing Obj trait could be sent to another thread and shared between threads safely.

Provided Methods

fn as_fs(&self) -> Option<&FsObj>