Module svsm::task

Modules

exec 🔒

schedule 🔒

Round-Robin scheduler implementation for COCONUT-SVSM

tasks 🔒

waiting 🔒

Structs

RunQueue

A RunQueue implementation that uses an RBTree to efficiently sort the priority of tasks within the queue.

Task

TaskContext

TaskListAdapter

TaskRunListAdapter

WaitQueue

Enums

TaskError

TaskState

Constants

INITIAL_TASK_ID

TASK_FLAG_SHARE_PT

Statics

TASKLIST

Functions

create_kernel_task

create_user_task

current_task

current_task_terminated⚠

Terminates the current task.

exec_user

is_current_task

Check to see if the task scheduled on the current processor has the given id

is_task_fault

schedule

Perform a task switch and hand the CPU over to the next task on the run-list. In case the current task is terminated, it will be destroyed after the switch to the next task.

schedule_init

Initializes the RunQueue on the current CPU. It will switch to the idle task and initialize the current_task field of the RunQueue. After this function has ran it is safe to call schedule() on the current CPU.

schedule_task

terminate

Type Aliases

TaskPointer