svsm::task

Function start_kernel_thread

pub fn start_kernel_thread(
    start_info: KernelThreadStartInfo,
) -> Result<TaskPointer, SvsmError>

Creates, initializes and starts a new kernel thread of the currently running kernel task. Note that the thread has already started to run before this function returns.

Arguments

entry - The function to run as the new task’s main function

start_parameter - Parameter of type usize to pass to the new threads main function.

Returns

A new instance of TaskPointer on success, SvsmError on failure.