svsm::task::schedule

Struct RunQueue

pub struct RunQueue {
    run_list: LinkedList<TaskRunListAdapter>,
    current_task: Option<TaskPointer>,
    idle_task: Option<TaskPointer>,
    terminated_task: Option<TaskPointer>,
    wake_from_idle: Option<TaskPointer>,
    set_affinity: Option<(TaskPointer, usize)>,
}

A RunQueue implementation that uses an RBTree to efficiently sort the priority of tasks within the queue.

Fields
run_list: LinkedList<TaskRunListAdapter>

Linked list with runable tasks

current_task: Option<TaskPointer>

Pointer to currently running task

idle_task: Option<TaskPointer>

Idle task - runs when there is no other runnable task

terminated_task: Option<TaskPointer>

Temporary storage for tasks which are about to be terminated

wake_from_idle: Option<TaskPointer>

Pointer to a task that should be woken when returning from idle

set_affinity: Option<(TaskPointer, usize)>

Pointer to a task that is requesting an affinity change to another processor, along with the CPU index describing the new affinity..

Implementations

impl RunQueue

pub fn new() -> Self

Create a new runqueue for an id. The id would normally be set to the APIC ID of the CPU that owns the runqueue and is used to determine the affinity of tasks.

fn get_next_task(&mut self) -> TaskPointer

Find the next task to run, which is either the task at the front of the run_list or the idle task, if the run_list is empty.

Returns

Pointer to next task to run

Panics

Panics if there are no tasks to run and no idle task has been allocated via set_idle_task().

fn handle_task(&mut self, task: TaskPointer)

Update state before a task is scheduled out. Non-idle tasks in RUNNING state will be put at the end of the run_list. Terminated tasks will be stored in the terminated_task field of the RunQueue and be destroyed after the task-switch.

pub fn schedule_init(&mut self) -> TaskPointer

Initialized the scheduler for this (RunQueue)RunQueue. This method is called on the very first scheduling event when there is no current task yet.

Returns

TaskPointer to the first task to run

pub fn schedule_prepare(&mut self) -> Option<(TaskPointer, TaskPointer)>

Prepares a task switch. The function checks if a task switch needs to be done and return pointers to the current and next task. It will also call handle_task() on the current task in case a task-switch is requested.

Returns

None when no task-switch is needed. Some with current and next task in case a task-switch is required.

Panics

Panics if there is no current task.

pub fn current_task_id(&self) -> u32

pub fn set_idle_task(&mut self, task: TaskPointer)

Sets the idle task for this RunQueue. This function sets a OnceCell at the end and can thus be only called once.

Returns

Ok(()) on success, SvsmError on failure

Panics

Panics if the idle task was already set.

pub fn current_task(&self) -> TaskPointer

Gets a pointer to the current task

Panics

Panics if there is no current task.

pub fn wake_from_idle(&mut self) -> Option<TaskPointer>

Wakes a task from idle if required.

Returns

An Option<TaskPointer> indicating which task should be woken, if any.