svsm::task

Module tasks

Structs

CreateTaskArguments 🔒

KernelThreadStartInfo

Task

TaskContext

TaskIDAllocator 🔒

TaskListAdapter

TaskRunListAdapter

TaskSchedState 🔒

Enums

TaskError

TaskState

ThreadStartInfo 🔒

Constants

INITIAL_TASK_ID

TASK_FLAG_SHARE_PT

Statics

TASK_ID_ALLOCATOR 🔒

Traits

KernelThreadStartParameter

This trait is used to describe a type that can be used as the start parameter for a kernel thread. The thread start path requires that all start paremeters be able to fit into a single register, which constrains them to usize. Other types can be used as long as there is an appropriate mechanism to covert those types to and from the usize that must be passed via the parameter register. This trait defines such a conversion mechanism.

Functions

is_task_fault

run_kernel_task 🔒 ⚠

Safety

setup_new_task_common 🔒 ⚠

setup_user_task 🔒 ⚠

Runs the first time a new task is scheduled, in the context of the new task. Any first-time initialization and setup work for a new task that needs to happen in its context must be done here.

task_attach_console 🔒

task_exit 🔒

Type Aliases

TaskPointer