svsm::task::tasks

Struct Task

pub struct Task {

    pub rsp: u64,
    pub ssp: VirtAddr,
    pub xsa: PageBox<[u8]>,
    pub stack_bounds: MemoryRegion<VirtAddr>,
    pub shadow_stack_base: VirtAddr,
    pub page_table: SpinLock<PageBox<PageTable>>,
    _ktask_region: TaskVirtualRegionGuard,
    vm_kernel_range: VMR,
    vm_user_range: Option<VMR>,
    sched_state: RWLock<TaskSchedState>,
    name: String,
    id: u32,
    rootdir: Arc<dyn Directory>,
    list_link: LinkedListAtomicLink,
    runlist_link: LinkedListAtomicLink,
    objs: Arc<RWLock<BTreeMap<ObjHandle, Arc<dyn Obj>>>>,

}

Fields
rsp: u64ssp: VirtAddrxsa: PageBox<[u8]>

XSave area

stack_bounds: MemoryRegion<VirtAddr>shadow_stack_base: VirtAddrpage_table: SpinLock<PageBox<PageTable>>

Page table that is loaded when the task is scheduled

_ktask_region: TaskVirtualRegionGuard

Virtual address region that has been allocated for this task. This is not referenced but must be stored so that it is dropped when the Task is dropped.

vm_kernel_range: VMR

Task virtual memory range for use at CPL 0

vm_user_range: Option<VMR>

Task virtual memory range for use at CPL 3 - None for kernel tasks

sched_state: RWLock<TaskSchedState>

State relevant for scheduler

name: String

User-visible name of task

id: u32

ID of the task

rootdir: Arc<dyn Directory>

Root directory for this task

list_link: LinkedListAtomicLink

Link to global task list

runlist_link: LinkedListAtomicLink

Link to scheduler run queue

objs: Arc<RWLock<BTreeMap<ObjHandle, Arc<dyn Obj>>>>

Objects shared among threads within the same process

Implementations

impl Task

fn create_common( cpu: &PerCpu, args: CreateTaskArguments, ) -> Result<TaskPointer, SvsmError>

pub fn create( cpu: &PerCpu, entry: extern "C" fn(usize), start_parameter: usize, name: String, ) -> Result<TaskPointer, SvsmError>

pub fn create_user( cpu: &PerCpu, user_entry: usize, root: Arc<dyn Directory>, name: String, ) -> Result<TaskPointer, SvsmError>

pub fn stack_bounds(&self) -> MemoryRegion<VirtAddr>

pub fn get_task_name(&self) -> &String

pub fn get_task_id(&self) -> u32

pub fn rootdir(&self) -> Arc<dyn Directory>

pub fn set_task_running(&self)

pub fn set_task_terminated(&self)

pub fn set_task_blocked(&self)

pub fn is_running(&self) -> bool

pub fn is_terminated(&self) -> bool

pub fn set_idle_task(&self)

pub fn is_idle_task(&self) -> bool

pub fn update_cpu(&self, new_cpu_index: usize) -> usize

pub fn handle_pf(&self, vaddr: VirtAddr, write: bool) -> Result<(), SvsmError>

pub fn fault(&self, vaddr: VirtAddr, write: bool) -> Result<(), SvsmError>

fn allocate_stack_common() -> Result<(Arc<Mapping>, MemoryRegion<VirtAddr>), SvsmError>

fn allocate_ktask_stack( cpu: &PerCpu, entry: usize, xsa_addr: usize, start_parameter: usize, ) -> Result<(Arc<Mapping>, MemoryRegion<VirtAddr>, usize), SvsmError>

fn allocate_utask_stack( cpu: &PerCpu, user_entry: usize, xsa_addr: usize, ) -> Result<(Arc<Mapping>, MemoryRegion<VirtAddr>, usize), SvsmError>

fn allocate_xsave_area() -> PageBox<[u8]>

pub fn mmap_common( vmr: &VMR, addr: VirtAddr, file: Option<&FileHandle>, offset: usize, size: usize, flags: VMFileMappingFlags, ) -> Result<VirtAddr, SvsmError>

pub fn mmap_kernel( &self, addr: VirtAddr, file: Option<&FileHandle>, offset: usize, size: usize, flags: VMFileMappingFlags, ) -> Result<VirtAddr, SvsmError>

pub fn mmap_kernel_guard<'a>( &'a self, addr: VirtAddr, file: Option<&FileHandle>, offset: usize, size: usize, flags: VMFileMappingFlags, ) -> Result<VMMappingGuard<'a>, SvsmError>

pub fn mmap_user( &self, addr: VirtAddr, file: Option<&FileHandle>, offset: usize, size: usize, flags: VMFileMappingFlags, ) -> Result<VirtAddr, SvsmError>

pub fn munmap_kernel(&self, addr: VirtAddr) -> Result<(), SvsmError>

pub fn munmap_user(&self, addr: VirtAddr) -> Result<(), SvsmError>

pub fn add_obj(&self, obj: Arc<dyn Obj>) -> Result<ObjHandle, SvsmError>

Adds an object to the current task.

Arguments

obj - The object to be added.

Returns

Result<ObjHandle, SvsmError> - Returns the object handle for the object to be added if successful, or an SvsmError on failure.

Errors

This function will return an error if allocating the object handle fails.

pub fn add_obj_at( &self, obj: Arc<dyn Obj>, handle: ObjHandle, ) -> Result<ObjHandle, SvsmError>

Adds an object to the current task and maps it to a given object-id.

Arguments

obj - The object to be added.

handle - Object handle to reference the object.

Returns

Result<ObjHandle, SvsmError> - Returns the object handle for the object to be added if successful, or an SvsmError on failure.

Errors

This function will return an error if allocating the object handle fails or the object id is already in use.

pub fn remove_obj(&self, id: ObjHandle) -> Result<Arc<dyn Obj>, SvsmError>

Removes an object from the current task.

Arguments

id - The ObjHandle for the object to be removed.

Returns

Result<Arc<dyn Obj>>, SvsmError> - Returns the removed Arc<dyn Obj> on success, or an SvsmError on failure.

Errors

This function will return an error if the object handle id does not exist in the current task.

pub fn get_obj(&self, id: ObjHandle) -> Result<Arc<dyn Obj>, SvsmError>

Retrieves an object from the current task.

Arguments

id - The ObjHandle for the object to be retrieved.

Returns

Result<Arc<dyn Obj>>, SvsmError> - Returns the Arc<dyn Obj> on success, or an SvsmError on failure.

Errors

This function will return an error if the object handle id does not exist in the current task.