svsm::task::tasks

Trait KernelThreadStartParameter

pub unsafe trait KernelThreadStartParameter {
    // Required methods
    fn to_usize(parameter: Self) -> usize;
    unsafe fn from_usize(parameter: usize) -> Self;
}

This trait is used to describe a type that can be used as the start parameter for a kernel thread. The thread start path requires that all start paremeters be able to fit into a single register, which constrains them to usize. Other types can be used as long as there is an appropriate mechanism to covert those types to and from the usize that must be passed via the parameter register. This trait defines such a conversion mechanism.

Safety

Implementations of this thread must ensure that the to_usize and from_usize methods correctly preserve the safety of the parameter as it transitions from its native type to and from a temporary usize representation.

Required Methods

fn to_usize(parameter: Self) -> usize

unsafe fn from_usize(parameter: usize) -> Self

Safety

This function must only be called with a usize parameter that was previously returned from a call to to_usize().

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.