svsm::utils::memory_region

Struct MemoryRegion

pub struct MemoryRegion<A> {
    start: A,
    end: A,
}

An abstraction over a memory region, expressed in terms of physical (PhysAddr) or virtual (VirtAddr) addresses.

Fields
start: Aend: A
Implementations

impl<A> MemoryRegion<A>
where A: Address,

pub fn new(start: A, len: usize) -> Self

Create a new memory region starting at address start, spanning len bytes.

pub fn checked_new(start: A, len: usize) -> Option<Self>

Create a new memory region with overflow checks.

let start = VirtAddr::from(u64::MAX);
let region = MemoryRegion::checked_new(start, PAGE_SIZE);
assert!(region.is_none());

pub fn from_addresses(start: A, end: A) -> Self

Create a memory region from two raw addresses.

pub const fn start(&self) -> A

The base address of the memory region, originally set in MemoryRegion::new().

pub fn len(&self) -> usize

The length of the memory region in bytes, originally set in MemoryRegion::new().

pub fn is_empty(&self) -> bool

Returns whether the region spans any actual memory.

let r = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), 0);
assert!(r.is_empty());

pub const fn end(&self) -> A

The end address of the memory region.

let base = VirtAddr::from(0xffffff0000u64);
let region = MemoryRegion::new(base, PAGE_SIZE);
assert_eq!(region.end(), VirtAddr::from(0xffffff1000u64));

pub fn overlap(&self, other: &Self) -> bool

Checks whether two regions overlap. This does not include contiguous regions, use MemoryRegion::contiguous() for that purpose.

let r1 = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
let r2 = MemoryRegion::new(VirtAddr::from(0xffffff2000u64), PAGE_SIZE);
assert!(!r1.overlap(&r2));

let r1 = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE * 2);
let r2 = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE);
assert!(r1.overlap(&r2));

// Contiguous regions do not overlap
let r1 = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
let r2 = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE);
assert!(!r1.overlap(&r2));

pub fn contiguous(&self, other: &Self) -> bool

Checks whether two regions are contiguous or overlapping. This is a less strict check than MemoryRegion::overlap().

let r1 = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
let r2 = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE);
assert!(r1.contiguous(&r2));

let r1 = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
let r2 = MemoryRegion::new(VirtAddr::from(0xffffff2000u64), PAGE_SIZE);
assert!(!r1.contiguous(&r2));

pub fn merge(&self, other: &Self) -> Self

Merge two regions. It does not check whether the two regions are contiguous in the first place, so the resulting region will cover any non-overlapping memory between both.

let r1 = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
let r2 = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE);
let r3 = r1.merge(&r2);
assert_eq!(r3.start(), r1.start());
assert_eq!(r3.len(), r1.len() + r2.len());
assert_eq!(r3.end(), r2.end());

pub fn iter_pages(&self, size: PageSize) -> impl Iterator<Item = A>

Iterate over the addresses covering the memory region in jumps of the specified page size. Note that if the base address of the region is not page aligned, returned addresses will not be aligned either.

let region = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE * 2);
let mut iter = region.iter_pages(PageSize::Regular);
assert_eq!(iter.next(), Some(VirtAddr::from(0xffffff0000u64)));
assert_eq!(iter.next(), Some(VirtAddr::from(0xffffff1000u64)));
assert_eq!(iter.next(), None);

pub fn contains(&self, addr: A) -> bool

Check whether an address is within this region.

let region = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
assert!(region.contains(VirtAddr::from(0xffffff0000u64)));
assert!(region.contains(VirtAddr::from(0xffffff0fffu64)));
assert!(!region.contains(VirtAddr::from(0xffffff1000u64)));

pub fn contains_inclusive(&self, addr: A) -> bool

Check whether an address is within this region, treating end as part of the region.

let region = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE);
assert!(region.contains_inclusive(VirtAddr::from(0xffffff0000u64)));
assert!(region.contains_inclusive(VirtAddr::from(0xffffff0fffu64)));
assert!(region.contains_inclusive(VirtAddr::from(0xffffff1000u64)));
assert!(!region.contains_inclusive(VirtAddr::from(0xffffff1001u64)));

pub fn contains_region(&self, other: &Self) -> bool

Check whether this region fully contains a different region.

let big = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE * 2);
let small = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE);
let overlapping = MemoryRegion::new(VirtAddr::from(0xffffff0000u64), PAGE_SIZE * 2);
assert!(big.contains_region(&small));
assert!(!small.contains_region(&big));
assert!(!overlapping.contains_region(&big));
assert!(!big.contains_region(&overlapping));

pub fn expand(&self, len: usize) -> Self

Returns a new memory region with the specified added length at the end.

let region = MemoryRegion::new(VirtAddr::from(0xffffff1000u64), PAGE_SIZE);
let bigger = region.expand(PAGE_SIZE);
assert_eq!(bigger.len(), PAGE_SIZE * 2);