svsm::utils::unsafe_copy

Function unsafe_copy_bytes

pub unsafe fn unsafe_copy_bytes(src: usize, dst: usize, size: usize)

Safety

The caller is required to ensure that the source/destination pointers and size will not result in overwriting memory unexpectedly.