Module svsm::vtpm

This crate defines the Virtual TPM interfaces and shows what TPM backends are supported

Modules

mstpm

TPM 2.0 Reference Implementation by Microsoft This crate implements the virtual TPM interfaces for the TPM 2.0 Reference Implementation (by Microsoft)

Statics

VTPM 🔒

Traits

MsTpmSimulatorInterface

This implements one handler for each TpmPlatformCommand supported by the VTPM Protocol. These handlers are based on the TPM Simulator interface (by Microsoft), but with a few changes to make it more Rust idiomatic.

VtpmInterface

Basic TPM driver services

VtpmProtocolInterface

Basic services required to perform the VTPM Protocol

Functions

vtpm_get_locked

vtpm_init

Initialize the TPM by calling the init() implementation of the VtpmInterface