svsm::vtpm

Module tcgtpm

TPM 2.0 Reference Implementation This crate implements the virtual TPM interfaces for the TPM 2.0 Reference Implementation (by Microsoft)

Modules

ek_templates

tss 🔒

wrapper 🔒

Functions required to build the TPM 2.0 Reference Implementation libraries Implement functions required to build the TPM 2.0 Reference Implementation libraries. All these functionalities are owned by the SVSM Rust code, so we just need to create wrappers for them.

Structs

TcgTpm

Constants

TPM_BUFFER_MAX_SIZE

TPM_CMDS_SUPPORTED 🔒