svsm::vtpm::tcgtpm
Module wrapper

Functions required to build the TPM 2.0 Reference Implementation libraries Implement functions required to build the TPM 2.0 Reference Implementation libraries. All these functionalities are owned by the SVSM Rust code, so we just need to create wrappers for them.

Functions

abort

calloc

free⚠

malloc

realloc⚠

serial_out⚠