Trait svsm::vtpm::MsTpmSimulatorInterface

pub trait MsTpmSimulatorInterface: VtpmProtocolInterface {
    // Required methods
    fn send_tpm_command(
        &self,
        buffer: &mut [u8],
        length: &mut usize,
        locality: u8,
    ) -> Result<(), SvsmReqError>;
    fn signal_poweron(&mut self, only_reset: bool) -> Result<(), SvsmReqError>;
    fn signal_nvon(&self) -> Result<(), SvsmReqError>;
}

This implements one handler for each TpmPlatformCommand supported by the VTPM Protocol. These handlers are based on the TPM Simulator interface (by Microsoft), but with a few changes to make it more Rust idiomatic.

ms-tpm-20-ref/TPMCmd/Simulator/include/prototypes/Simulator_fp.h

Required Methods

fn send_tpm_command( &self, buffer: &mut [u8], length: &mut usize, locality: u8, ) -> Result<(), SvsmReqError>

Send a command for the TPM to run in a given locality

Arguments

buffer: Buffer with the command to be sent to the TPM. It has to be large enough to hold the response received from the TPM.

length: The length of the command stored in buffer. It will be updated with the size of the TPM response received from the TPM.

locality: TPM locality the TPM command will be executed

fn signal_poweron(&mut self, only_reset: bool) -> Result<(), SvsmReqError>

Power-on the TPM, which also triggers a reset

Arguments

*`only_reset``: If enabled, it will only reset the vTPM; however, the vtPM has to be powered on previously. Otherwise, it will fail.

fn signal_nvon(&self) -> Result<(), SvsmReqError>

In a system where the NV memory used by the TPM is not within the TPM, the NV may not always be available. This function indicates that NV is available.