svsm::acpi

Module tables

Structs

ACPICPUInfo

Information about an ACPI CPU

ACPITable

ACPI table, both header and contents

ACPITableBuffer 🔒

ACPI Table Buffer A buffer containing ACPI tables. Responsible for loading the tables from a firmware configuration

ACPITableHeader 🔒

Higher level representation of the raw ACPI table header

ACPITableMeta 🔒

ACPI Table Metadata Metadata associated with an ACPI, information about signature and offset

RSDPDesc 🔒

ACPI Root System Description Pointer (RSDP) used by ACPI programming interface

RawACPITableHeader 🔒

Raw header of an ACPI table. It corresponds to the beginning portion of ACPI tables, before any specific table data

RawMADTEntryHeader 🔒

Header of an entry within MADT

RawMADTEntryLocalApic 🔒

Entry for a local APIC within MADT

RawMADTEntryLocalX2Apic 🔒

Entry for a local X2APIC within MADT

Constants

MADT_HEADER_SIZE 🔒

MAX_ACPI_TABLES_SIZE 🔒

Functions

load_acpi_cpu_info

Loads ACPI CPU information by parsing the ACPI tables. It processes the Multiple APIC Description Table (MADT) to extract information about each CPU’s APIC ID and enabled status.

load_fw_cpu_info

Loads ACPI CPU information by parsing the ACPI tables provided by the firmware It processes the Multiple APIC Description Table (MADT) to extract information about each CPU’s APIC ID and enabled status.