svsm::cpu

Module ipi

Structs

IpiBoard 🔒

IpiHelperMut

IpiHelperShared

IpiState

Enums

IpiRequest

IpiTarget

This module implements inter-processor interrupt support, including the ability to send and receive messages across CPUs. Two types of IPI messages are supported: multicast and unicast. Sending a multicast IPI will cause a message to be delivered to one or more CPUs as a shared reference. Sending a unicast IPI will cause a message to be delivered to a single target as a mutable reference, permitting the receiving processor to modify the contents of the message such that the sender of the message can observe the response. In all cases, the request to send a message will not complete until all receiving CPUs have completed handling the request.

Statics

IPI_AVAILABLE_CPU_COUNT 🔒

The count of CPUs that have not yet requested blocking of IPI usage. This is initially set to 1 to count the BSP, and each AP that starts will increment the count.

Traits

IpiHelper

IpiMessage

Safety

IpiMessageMut

Safety

Functions

flag_ipi 🔒 ⚠

Safety

handle_ipi_interrupt

ipi_available

Indicates whether use of IPIs is currently available.

ipi_start_cpu

Count the startup of another AP for IPI blocking purposes.

receive_single_ipi 🔒 ⚠

Safety

send_ipi 🔒

send_ipi_irq 🔒

send_multicast_ipi

Sends an IPI message to multiple CPUs.

send_single_ipi_irq 🔒

send_unicast_ipi

Sends an IPI message to a single CPU. Because only a single CPU can receive the message, the message object can be mutable.

wait_for_ipi_block

Request IPI blocking on the current CPU and wait until all other CPUs have done the same.