svsm::cpu::ipi

Trait IpiMessage

pub unsafe trait IpiMessage {
    // Required method
    fn invoke(&self);

    // Provided methods
    fn copy_to_shared(&self, buffer: &mut [u8])
       where Self: Sized { ... }
    fn finalize(&self, _shared_buffer: &Self) { ... }
}

Safety

This trait implements a method to copy IPI message contents into a shared buffer. If that serialization is performed incorrectly, then IPI message receipt will be unsound because the message may contain incorrect pointers and references that refer to invalid memory - or memory that belongs to another owner. All implementations of this trait must verify that the copy routine correctly copies all data and resolves all references within the copied data.

Required Methods

fn invoke(&self)

Invokes the IPI handler for the message.

Provided Methods

fn copy_to_shared(&self, buffer: &mut [u8])
where Self: Sized,

All IPI messages must be copied into a shared IPI buffer since stack locals are not visible across CPU/task contexts. This function must perform a deep copy of the contents of the source buffer into the shared destination buffer.

Arguments:

*src: A pointer to the input message. *buffer: A byte slice in shared memory which will be the target of the copy.

fn finalize(&self, _shared_buffer: &Self)

If an IPI message has any atomic members, then they may be modified by IPI execution. Therefore, it may be necessary to transfer the modified atomic members back to the original message so the sender can observe the final atomic value. IPI message implementations only need to implement this method if they require such finalization.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.