svsm::cpu

Module percpu

Structs

GuestVmsaRef

IstStacks 🔒

PerCpu

CPU-local data.

PerCpuAreas

PerCpuShared

PerCpuVmsas

VmsaRegistryEntry

Statics

PERCPU_AREAS

PERCPU_VMSAS

Functions

cpu_idle_loop

current_ghcb

Gets the GHCB for this CPU.

current_task

irq_nesting_count

Get IRQ-disable nesting count on the current CPU

irqs_disable

Disables IRQs on the current CPU. Keeps track of the nesting level and the original IRQ state.

irqs_enable

Reduces IRQ-disable nesting level on the current CPU and restores the original IRQ state when the level reaches 0.

lower_tpr

Lowers TPR from the current level to the new level required by the current nesting state.

raise_tpr

Raises TPR on the current CPU. Keeps track of the nesting level.

this_cpu

this_cpu_shared

try_this_cpu