Struct svsm::cpu::percpu::PerCpu

pub struct PerCpu {

    shared: PerCpuShared,
    irq_state: IrqState,
    pgtbl: RefCell<Option<&'static mut PageTable>>,
    tss: Cell<X86Tss>,
    svsm_vmsa: OnceCell<VmsaPage>,
    reset_ip: Cell<u64>,
    vm_range: VMR,
    pub vrange_4k: RefCell<VirtualRange>,
    pub vrange_2m: RefCell<VirtualRange>,
    runqueue: RWLockIrqSafe<RunQueue>,
    request_waitqueue: RefCell<WaitQueue>,
    apic: RefCell<Option<LocalApic>>,
    ghcb: OnceCell<GhcbPage>,
    hv_doorbell: Cell<Option<&'static HVDoorbell>>,
    init_stack: Cell<Option<VirtAddr>>,
    ist: IstStacks,
    current_stack: Cell<MemoryRegion<VirtAddr>>,

}

CPU-local data.

This type is not Sync, as its contents will only be accessed from the local CPU, much like thread-local data in an std environment. The only part of the struct that may be accessed from a different CPU is the shared field, a reference to which will be stored in PERCPU_AREAS.

Fields
shared: PerCpuShared

Per-CPU storage that might be accessed from other CPUs.

irq_state: IrqState

PerCpu IRQ state tracking

pgtbl: RefCell<Option<&'static mut PageTable>>tss: Cell<X86Tss>svsm_vmsa: OnceCell<VmsaPage>reset_ip: Cell<u64>vm_range: VMR

PerCpu Virtual Memory Range

vrange_4k: RefCell<VirtualRange>

Address allocator for per-cpu 4k temporary mappings

vrange_2m: RefCell<VirtualRange>

Address allocator for per-cpu 2m temporary mappings

runqueue: RWLockIrqSafe<RunQueue>

Task list that has been assigned for scheduling on this CPU

request_waitqueue: RefCell<WaitQueue>

WaitQueue for request processing

apic: RefCell<Option<LocalApic>>

Local APIC state for APIC emulation if enabled

ghcb: OnceCell<GhcbPage>

GHCB page for this CPU.

hv_doorbell: Cell<Option<&'static HVDoorbell>>

#HV doorbell page for this CPU.

init_stack: Cell<Option<VirtAddr>>ist: IstStackscurrent_stack: Cell<MemoryRegion<VirtAddr>>

Stack boundaries of the currently running task.

Implementations

impl PerCpu

fn new(apic_id: u32) -> Self

Creates a new default PerCpu struct.

pub fn alloc(apic_id: u32) -> Result<&'static Self, SvsmError>

Creates a new default PerCpu struct, allocates it via the page allocator and adds it to the global per-cpu area list.

pub fn shared(&self) -> &PerCpuShared

pub unsafe fn irqs_disable(&self)

Disables IRQs on the current CPU. Keeps track of the nesting level and the original IRQ state.

Safety

Caller needs to make sure to match every disable() call with an enable() call.

pub unsafe fn irqs_enable(&self)

Reduces IRQ-disable nesting level on the current CPU and restores the original IRQ state when the level reaches 0.

Safety

Caller needs to make sure to match every disable() call with an enable() call.

pub fn irq_nesting_count(&self) -> isize

Get IRQ-disable nesting count on the current CPU

Returns

Current nesting depth of irq_disable() calls.

pub fn setup_ghcb(&self) -> Result<(), SvsmError>

Sets up the CPU-local GHCB page.

fn ghcb(&self) -> Option<&GhcbPage>

pub fn hv_doorbell(&self) -> Option<&'static HVDoorbell>

pub fn hv_doorbell_addr(&self) -> *const &'static HVDoorbell

Gets a pointer to the location of the HV doorbell pointer in the PerCpu structure. Pointers and references have the same layout, so the return type is equivalent to *const *const HVDoorbell.

pub fn get_top_of_stack(&self) -> VirtAddr

pub fn get_top_of_df_stack(&self) -> VirtAddr

pub fn get_current_stack(&self) -> MemoryRegion<VirtAddr>

pub fn get_apic_id(&self) -> u32

pub fn init_page_table( &self, pgtable: PageBox<PageTable>, ) -> Result<(), SvsmError>

pub fn set_pgtable(&self, pgtable: &'static mut PageTable)

fn allocate_stack(&self, base: VirtAddr) -> Result<VirtAddr, SvsmError>

fn allocate_init_stack(&self) -> Result<(), SvsmError>

fn allocate_ist_stacks(&self) -> Result<(), SvsmError>

pub fn get_pgtable(&self) -> RefMut<'_, PageTable>

pub fn register_ghcb(&self) -> Result<(), SvsmError>

Registers an already set up GHCB page for this CPU.

Panics

Panics if the GHCB for this CPU has not been set up via PerCpu::setup_ghcb().

fn setup_hv_doorbell(&self) -> Result<(), SvsmError>

pub fn configure_hv_doorbell(&self) -> Result<(), SvsmError>

Configures the HV doorbell page if restricted injection is enabled.

Panics

Panics if this function is called more than once for a given CPU and restricted injection is enabled.

fn setup_tss(&self)

pub fn map_self_stage2(&self) -> Result<(), SvsmError>

pub fn map_self(&self) -> Result<(), SvsmError>

fn initialize_vm_ranges(&self) -> Result<(), SvsmError>

fn finish_page_table(&self)

pub fn dump_vm_ranges(&self)

pub fn setup( &self, platform: &dyn SvsmPlatform, pgtable: PageBox<PageTable>, ) -> Result<(), SvsmError>

pub fn setup_on_cpu(&self, platform: &dyn SvsmPlatform) -> Result<(), SvsmError>

pub fn setup_idle_task(&self, entry: extern "C" fn()) -> Result<(), SvsmError>

pub fn load_pgtable(&self)

pub fn load_tss(&self)

pub fn load(&self)

pub fn set_reset_ip(&self, reset_ip: u64)

pub fn alloc_svsm_vmsa( &self, vtom: u64, start_rip: u64, ) -> Result<(PhysAddr, u64), SvsmError>

Allocates and initializes a new VMSA for this CPU. Returns its physical address and SEV features. Returns an error if allocation fails of this CPU’s VMSA was already initialized.

pub fn unmap_guest_vmsa(&self)

pub fn map_guest_vmsa(&self, paddr: PhysAddr) -> Result<(), SvsmError>

pub fn guest_vmsa_ref(&self) -> LockGuard<'_, GuestVmsaRef>

pub fn alloc_guest_vmsa(&self) -> Result<(), SvsmError>

fn apic(&self) -> Option<Ref<'_, LocalApic>>

Returns a shared reference to the local APIC, or None if APIC emulation is not enabled.

fn apic_mut(&self) -> Option<RefMut<'_, LocalApic>>

Returns a mutable reference to the local APIC, or None if APIC emulation is not enabled.

pub fn unmap_caa(&self)

pub fn map_guest_caa(&self, paddr: PhysAddr) -> Result<(), SvsmError>

pub fn disable_apic_emulation(&self)

pub fn clear_pending_interrupts(&self)

pub fn update_apic_emulation(&self, vmsa: &mut VMSA, caa_addr: Option<VirtAddr>)

pub fn use_apic_emulation(&self) -> bool

pub fn read_apic_register(&self, register: u64) -> Result<u64, SvsmError>

pub fn write_apic_register( &self, register: u64, value: u64, ) -> Result<(), SvsmError>

pub fn configure_apic_vector( &self, vector: u8, allowed: bool, ) -> Result<(), SvsmError>

fn vmsa_tr_segment(&self) -> VMSASegment

fn virt_range_init(&self)

pub fn new_mapping( &self, mapping: Arc<Mapping>, ) -> Result<VMRMapping<'_>, SvsmError>

Create a new virtual memory mapping in the PerCpu VMR

Arguments

mapping - The mapping to insert into the PerCpu VMR

Returns

On success, a new [‘VMRMapping’} that provides a virtual memory address for the mapping which remains valid until the [‘VRMapping’] is dropped.

On error, an [‘SvsmError’].

pub fn populate_page_table(&self, pt: &mut PageTable)

Add the PerCpu virtual range into the provided pagetable

Arguments

pt - The page table to populate the the PerCpu range into

pub fn handle_pf(&self, vaddr: VirtAddr, write: bool) -> Result<(), SvsmError>

pub fn schedule_init(&self) -> TaskPointer

pub fn schedule_prepare(&self) -> Option<(TaskPointer, TaskPointer)>

pub fn runqueue(&self) -> &RWLockIrqSafe<RunQueue>

pub fn current_task(&self) -> TaskPointer

pub fn set_tss_rsp0(&self, addr: VirtAddr)