svsm::cpu::percpu

Struct PerCpu

pub struct PerCpu {

    shared: &'static PerCpuShared,
    apic: X86Apic,
    irq_state: IrqState,
    pgtbl: RefCell<Option<&'static mut PageTable>>,
    tss: X86Tss,
    isst: Cell<Isst>,
    svsm_vmsa: ImmutAfterInitCell<VmsaPage>,
    reset_ip: AtomicU64,
    vm_range: VMR,
    pub vrange_4k: RefCell<VirtualRange>,
    pub vrange_2m: RefCell<VirtualRange>,
    runqueue: RWLockIrqSafe<RunQueue>,
    guest_apic: RefCell<Option<LocalApic>>,
    ghcb: ImmutAfterInitCell<GhcbPage>,
    hypercall_pages: RefCell<Option<(HypercallPage, HypercallPage)>>,
    hv_doorbell: ImmutAfterInitCell<SharedBox<HVDoorbell>>,
    init_shadow_stack: ImmutAfterInitCell<VirtAddr>,
    context_switch_stack: ImmutAfterInitCell<VirtAddr>,
    ist: IstStacks,
    current_stack: Cell<MemoryRegion<VirtAddr>>,

}

CPU-local data.

This type is not Sync, as its contents will only be accessed from the local CPU, much like thread-local data in an std environment. The only part of the struct that may be accessed from a different CPU is the shared field, a reference to which will be stored in PERCPU_AREAS.

Fields
shared: &'static PerCpuShared

Reference to the PerCpuShared that is valid in the global, shared address space.

apic: X86Apic

APIC access object

irq_state: IrqState

PerCpu IRQ state tracking

pgtbl: RefCell<Option<&'static mut PageTable>>tss: X86Tssisst: Cell<Isst>svsm_vmsa: ImmutAfterInitCell<VmsaPage>reset_ip: AtomicU64vm_range: VMR

PerCpu Virtual Memory Range

vrange_4k: RefCell<VirtualRange>

Address allocator for per-cpu 4k temporary mappings

vrange_2m: RefCell<VirtualRange>

Address allocator for per-cpu 2m temporary mappings

runqueue: RWLockIrqSafe<RunQueue>

Task list that has been assigned for scheduling on this CPU

guest_apic: RefCell<Option<LocalApic>>

Local APIC state for APIC emulation if enabled

ghcb: ImmutAfterInitCell<GhcbPage>

GHCB page for this CPU.

hypercall_pages: RefCell<Option<(HypercallPage, HypercallPage)>>

Hypercall input/output pages for this CPU if running under Hyper-V.

hv_doorbell: ImmutAfterInitCell<SharedBox<HVDoorbell>>

#HV doorbell page for this CPU.

init_shadow_stack: ImmutAfterInitCell<VirtAddr>context_switch_stack: ImmutAfterInitCell<VirtAddr>ist: IstStackscurrent_stack: Cell<MemoryRegion<VirtAddr>>

Stack boundaries of the currently running task.

Implementations

impl PerCpu

fn new(shared: &'static PerCpuShared) -> Self

Creates a new default PerCpu struct.

pub fn alloc(shared: &'static PerCpuShared) -> Result<&'static Self, SvsmError>

Creates a new default PerCpu struct, allocates it via the page allocator and adds it to the global per-cpu area list.

pub fn shared(&self) -> &PerCpuShared

pub fn initialize_apic(&self, accessor: &'static dyn ApicAccess)

pub fn get_apic(&self) -> &X86Apic

Get a reference to the X86Apic object for this cpu.

Returns

Reference to the X86Apic object of the local CPU.

pub fn irqs_disable(&self)

Disables IRQs on the current CPU. Keeps track of the nesting level and the original IRQ state.

Caller needs to make sure to match every disable() call with an enable() call.

pub fn irqs_enable(&self)

Reduces IRQ-disable nesting level on the current CPU and restores the original IRQ state when the level reaches 0.

Caller needs to make sure to match every disable() call with an enable() call.

pub fn irqs_push_nesting(&self, was_enabled: bool)

Increments IRQ-disable nesting level on the current CPU without disabling interrupts. This is used by exception and interrupt dispatch routines that have already disabled interrupts.

Caller needs to make sure to match every push_nesting() call with a pop_nesting() call.

pub fn irqs_pop_nesting(&self)

Reduces IRQ-disable nesting level on the current CPU without restoring the original IRQ state original IRQ state. This is used by exception and interrupt dispatch routines that will restore interrupt state naturally.

Caller needs to make sure to match every disable() call with a pop_state() call.

pub fn irq_nesting_count(&self) -> i32

Get IRQ-disable nesting count on the current CPU

Returns

Current nesting depth of irq_disable() calls.

pub fn raise_tpr(&self, tpr_value: usize)

Raises TPR on the current CPU. Keeps track of the nesting level.

The caller must ensure that every raise_tpr() call is followed by a matching call to lower_tpr().

pub fn lower_tpr(&self, tpr_value: usize)

Lowers TPR from the current level to the new level required by the current nesting state.

The caller must ensure that a lower_tpr() call balances a preceding raise_tpr() call to the indicated level.

tpr_value - The TPR from which the caller would like to lower. Must be less than or equal to the current TPR.

pub fn setup_ghcb(&self) -> Result<(), SvsmError>

Sets up the CPU-local GHCB page.

fn ghcb(&self) -> Option<&GhcbPage>

pub fn allocate_hypercall_pages(&self) -> Result<(), SvsmError>

Allocates hypercall input/output pages for this CPU.

pub fn get_hypercall_pages(&self) -> HypercallPagesGuard<'_>

pub fn hv_doorbell(&self) -> Option<&HVDoorbell>

pub fn process_hv_events_if_required(&self)

pub fn hv_doorbell_addr(&self) -> *const *const HVDoorbell

Gets a pointer to the location of the HV doorbell pointer in the PerCpu structure.

pub fn get_top_of_shadow_stack(&self) -> Option<VirtAddr>

pub fn get_top_of_context_switch_stack(&self) -> Option<VirtAddr>

pub fn get_top_of_df_stack(&self) -> Option<VirtAddr>

pub fn get_top_of_df_shadow_stack(&self) -> Option<VirtAddr>

pub fn get_current_stack(&self) -> MemoryRegion<VirtAddr>

pub fn set_current_stack(&self, stack: MemoryRegion<VirtAddr>)

pub fn get_cpu_index(&self) -> usize

pub fn get_apic_id(&self) -> u32

pub fn init_page_table( &self, pgtable: PageBox<PageTable>, ) -> Result<(), SvsmError>

pub fn set_pgtable(&self, pgtable: &'static mut PageTable)

fn allocate_stack(&self, base: VirtAddr) -> Result<VirtAddr, SvsmError>

fn allocate_shadow_stack( &self, base: VirtAddr, init: ShadowStackInit, ) -> Result<VirtAddr, SvsmError>

fn allocate_init_shadow_stack(&self) -> Result<(), SvsmError>

fn allocate_context_switch_stack(&self) -> Result<(), SvsmError>

fn allocate_context_switch_shadow_stack(&self) -> Result<(), SvsmError>

fn allocate_ist_stacks(&self) -> Result<(), SvsmError>

fn allocate_isst_shadow_stacks(&self) -> Result<(), SvsmError>

pub fn get_pgtable(&self) -> RefMut<'_, PageTable>

pub fn register_ghcb(&self) -> Result<(), SvsmError>

Registers an already set up GHCB page for this CPU.

Panics

Panics if the GHCB for this CPU has not been set up via PerCpu::setup_ghcb().

pub fn setup_hv_doorbell(&self) -> Result<(), SvsmError>

fn setup_tss(&self)

fn setup_isst(&self)

pub fn map_self_stage2(&self) -> Result<(), SvsmError>

pub fn map_self(&self) -> Result<(), SvsmError>

fn initialize_vm_ranges(&self) -> Result<(), SvsmError>

fn finish_page_table(&self)

pub fn dump_vm_ranges(&self)

pub fn setup( &self, platform: &dyn SvsmPlatform, pgtable: PageBox<PageTable>, ) -> Result<(), SvsmError>

pub fn setup_on_cpu(&self, platform: &dyn SvsmPlatform) -> Result<(), SvsmError>

fn setup_idle_task_internal( &self, start_info: KernelThreadStartInfo, ) -> Result<(), SvsmError>

pub fn setup_idle_task(&self) -> Result<(), SvsmError>

pub fn setup_bsp_idle_task( &self, start_info: KernelThreadStartInfo, ) -> Result<(), SvsmError>

pub fn load_gdt_tss(&'static self, init_gdt: bool)

pub fn load_isst(&self)

pub fn load(&'static self)

pub fn set_reset_ip(&self, reset_ip: u64)

pub fn get_initial_context(&self, start_rip: u64) -> HvInitialVpContext

Fill in the initial context structure for the SVSM.

pub fn alloc_svsm_vmsa( &self, vtom: u64, start_rip: u64, ) -> Result<(PhysAddr, u64), SvsmError>

Allocates and initializes a new VMSA for this CPU. Returns its physical address and SEV features. Returns an error if allocation fails of this CPU’s VMSA was already initialized.

fn unmap_guest_vmsa(&self)

fn map_guest_vmsa(&self, paddr: PhysAddr) -> Result<(), SvsmError>

pub fn guest_vmsa_ref(&self) -> LockGuard<'_, GuestVmsaRef>

pub fn alloc_guest_vmsa(&self) -> Result<(), SvsmError>

fn guest_apic(&self) -> Option<Ref<'_, LocalApic>>

Returns a shared reference to the local APIC, or None if APIC emulation is not enabled.

fn guest_apic_mut(&self) -> Option<RefMut<'_, LocalApic>>

Returns a mutable reference to the local APIC, or None if APIC emulation is not enabled.

fn unmap_caa(&self)

fn map_guest_caa(&self, paddr: PhysAddr) -> Result<(), SvsmError>

pub fn update_guest_mappings(&self) -> Result<(), SvsmError>

pub fn disable_apic_emulation(&self)

pub fn clear_pending_interrupts(&self)

pub fn update_apic_emulation( &self, vmsa: &mut VMSA, caa: Option<NonNull<SvsmCaa>>, )

pub fn use_apic_emulation(&self) -> bool

pub fn read_apic_register(&self, register: u64) -> Result<u64, SvsmError>

pub fn write_apic_register( &self, register: u64, value: u64, ) -> Result<(), SvsmError>

pub fn configure_apic_vector( &self, vector: u8, allowed: bool, ) -> Result<(), SvsmError>

fn svsm_tr_segment(&self) -> HvSegmentRegister

fn virt_range_init(&self)

pub fn new_mapping( &self, mapping: Arc<Mapping>, ) -> Result<VMRMapping<'_>, SvsmError>

Create a new virtual memory mapping in the PerCpu VMR

Arguments

mapping - The mapping to insert into the PerCpu VMR

Returns

On success, a new [‘VMRMapping’} that provides a virtual memory address for the mapping which remains valid until the [‘VRMapping’] is dropped.

On error, an [‘SvsmError’].

pub fn populate_page_table(&self, pt: &mut PageTable)

Add the PerCpu virtual range into the provided pagetable

Arguments

pt - The page table to populate the the PerCpu range into

pub fn handle_pf(&self, vaddr: VirtAddr, write: bool) -> Result<(), SvsmError>

pub fn schedule_init(&self) -> TaskPointer

pub fn disable_interrupt_use(&self)

pub fn schedule_prepare(&self) -> Option<(TaskPointer, TaskPointer)>

pub fn runqueue(&self) -> &RWLockIrqSafe<RunQueue>

pub fn current_task(&self) -> TaskPointer

pub unsafe fn set_tss_rsp0(&self, addr: VirtAddr)

Safety

No checks are performed on the stack address. The caller must ensure that the address is valid for stack usage.