svsm::cpu::x86

Module apic

Structs

X86Apic

Constants

APIC_ENABLE_MASK 🔒

APIC_OFFSET_EOI

End-of-Interrupt register MSR offset

APIC_OFFSET_ICR

Interrupt-Control-Register register MSR offset

APIC_OFFSET_ID

Local APIC ID register MSR offset

APIC_OFFSET_ISR

Interrupt-Service-Register base MSR offset

APIC_OFFSET_SELF_IPI

SELF-IPI register MSR offset (x2APIC only)

APIC_OFFSET_SPIV

Spurious-Interrupt-Register MSR offset

APIC_SPIV_SW_ENABLE_MASK 🔒

APIC_SPIV_VECTOR_MASK 🔒

APIC_X2_ENABLE_MASK 🔒

MSR_APIC_BASE

APIC Base MSR

Traits

ApicAccess

Functions

apic_enable

Enables the X86 local APIC in X2APIC mode by writing to MSR_APIC_BASE.

apic_eoi

Send an EOI message

apic_in_service

Check whether a given IRQ vector is currently being serviced by returning the value of its ISR bit from X2APIC.

apic_initialize

Initialize the APIC by setting an accessor object. This function does not enable the APIC.

apic_post_irq

Sends an IPI specified by the X86 ICR value.

apic_register_bit 🔒

Get the MSR offset relative to a bitmap base MSR and the mask for the MSR value to check for a specific vector bit being set in IRR, ISR, or TMR.

apic_sw_enable

Enables software IRQs in the X86 local APIC by setting the SPIC.SW_ENABLE bit.