Module svsm::fs

Modules

api 🔒

filesystem 🔒

init 🔒

ramfs 🔒

Structs

DirectoryEntry

Directory entries including their names.

FileHandle

Represents a handle used for file operations in a thread-safe manner.

Enums

DirEntry

Represents a directory entry which could either be a file or a subdirectory.

FsError

Represents the type of error occured while doing SVSM filesystem operations.

Traits

Directory

Represents directory operations

File

Represents file operations

Functions

create

Used to create a file with the given path.

create_all

Used to create a file and the missing subdirectories in the given path.

initialize_fs

Used to initialize the filesystem with an empty root directory.

list_dir

Used to list the contents of a directory.

mkdir

Used to create a directory with the given path.

open

Used to open a file to get the file handle for further file operations.

populate_ram_fs

Used to create a SVSM RAM filesystem from a filesystem archive.

read

Used to read from a file handle.

seek

Used to set the file offset

unlink

Used to delete a file or a directory.

write

Used to write into file handle.

Type Aliases

FileName