svsm

Module fs

Modules

api 🔒

buffer 🔒

console 🔒

filesystem 🔒

init 🔒

obj 🔒

ramfs 🔒

Structs

ConsoleFile

DirectoryEntry

Directory entries including their names.

FileHandle

Represents a handle used for file operations in a thread-safe manner.

FsObj

SliceMutRefBuffer

Struct to add a Buffer interface to a mutable &[u8] slice

SliceRefBuffer

Struct to add a Buffer interface to a non-mutable &[u8] slice

UserBuffer

Enums

DirEntry

Represents a directory entry which could either be a file or a subdirectory.

FsError

Represents the type of error occurred while doing SVSM filesystem operations.

Traits

Buffer

Directory

Represents directory operations

File

Represents file operations

Functions

create

Create a file with the given path from the file-system root.

create_all

Used to create a file and the missing subdirectories in the given path.

create_root

Create a file with the given path from a given root directory.

find_dir

Looks for a directory entry using a relative path.

initialize_fs

Used to initialize the filesystem with an empty root directory.

list_dir

Used to list the contents of a directory.

mkdir

Create a directory with the given path relative to the file-system root.

mkdir_root

Create a directory with the given path relative to a given root directory.

open

Open a file to get the file handle for further file operations.

open_read

Open a file to get the file handle for reading.

open_root

Open a file to get the file handle for further file operations. Starts file-name search from a given root directory.

open_rw

Open a file to get the file handle for reading and writing.

open_write

Open a file to get the file handle for writing.

opendir

Used to open a directory object.

populate_ram_fs

Used to create a SVSM RAM filesystem from a filesystem archive.

position

Return position of file read/write pointer.

read

Used to read from a file handle.

rmdir

Removes a directory relative to file-system root.

rmdir_root

Removes a directory

seek_abs

Set the file offset to an absolute position.

seek_end

Set the file offset relative to end-of-file

seek_rel

Set the file offset relative to the current offset

stdout_open

truncate

Truncate file at a given offset.

unlink

Unlink a file from its parent directory.

unlink_root

Unlink a file from its parent directory.

write

Used to write into file handle.

Type Aliases

FileName