Module svsm::fs::filesystem

Structs

FileHandle

Represents a handle used for file operations in a thread-safe manner.

RawFileHandle 🔒

Represents a raw file handle.

SvsmFs 🔒

Represents SVSM filesystem

Statics

FS_ROOT 🔒

Functions

create

Used to create a file with the given path.

create_all

Used to create a file and the missing subdirectories in the given path.

initialize_fs

Used to initialize the filesystem with an empty root directory.

list_dir

Used to list the contents of a directory.

mkdir

Used to create a directory with the given path.

open

Used to open a file to get the file handle for further file operations.

read

Used to read from a file handle.

seek

Used to set the file offset

split_path 🔒

Used to get an iterator over all the directory and file names contained in a path. This function performs error checking.

split_path_allow_empty 🔒

Used to get an iterator over all the directory and file names contained in a path. Directory name or file name in the path can be an empty value.

unlink

Used to delete a file or a directory.

walk_path 🔒

Used to perform a walk over the items in a path while checking each item is a directory.

walk_path_create 🔒

Used to perform a walk over the items in a path while checking each existing item is a directory, while creating a directory for each non-existing item.

write

Used to write into file handle.