svsm::fs

Module filesystem

Structs

FileHandle

Represents a handle used for file operations in a thread-safe manner.

RawFileHandle 🔒

Represents a raw file handle.

SvsmFs 🔒

Represents SVSM filesystem

Statics

FS_ROOT 🔒

Functions

create

Create a file with the given path from the file-system root.

create_all

Used to create a file and the missing subdirectories in the given path.

create_root

Create a file with the given path from a given root directory.

find_dir

Looks for a directory entry using a relative path.

initialize_fs

Used to initialize the filesystem with an empty root directory.

list_dir

Used to list the contents of a directory.

mkdir

Create a directory with the given path relative to the file-system root.

mkdir_root

Create a directory with the given path relative to a given root directory.

open

Open a file to get the file handle for further file operations.

open_read

Open a file to get the file handle for reading.

open_root

Open a file to get the file handle for further file operations. Starts file-name search from a given root directory.

open_rw

Open a file to get the file handle for reading and writing.

open_write

Open a file to get the file handle for writing.

opendir

Used to open a directory object.

position

Return position of file read/write pointer.

read

Used to read from a file handle.

rmdir

Removes a directory relative to file-system root.

rmdir_root

Removes a directory

seek_abs

Set the file offset to an absolute position.

seek_end

Set the file offset relative to end-of-file

seek_rel

Set the file offset relative to the current offset

split_path 🔒

Used to get an iterator over all the directory and file names contained in a path. This function performs error checking.

split_path_allow_empty 🔒

Used to get an iterator over all the directory and file names contained in a path. Directory name or file name in the path can be an empty value.

truncate

Truncate file at a given offset.

unlink

Unlink a file from its parent directory.

unlink_root

Unlink a file from its parent directory.

walk_path 🔒

Used to perform a walk over the items in a path while checking each item is a directory, starting at the given directory.

walk_path_create 🔒

Used to perform a walk over the items in a path while checking each existing item is a directory, while creating a directory for each non-existing item.

walk_path_from_root 🔒

Used to perform a walk over the items in a path while checking each item is a directory, starting at the root directory.

write

Used to write into file handle.