svsm::greq

Module msg

Message that carries an encrypted SNP_GUEST_REQUEST command in the payload

Structs

SnpGuestRequestMsg

SNP_GUEST_REQUEST message format

SnpGuestRequestMsgHdr

SNP_GUEST_REQUEST message header format (AMD SEV-SNP spec. table 98)

Enums

SnpGuestRequestAead

AEAD Algorithm Encodings (AMD SEV-SNP spec. table 99)

SnpGuestRequestMsgType

Message Type Encodings (AMD SEV-SNP spec. table 100)

Constants

HDR_VERSION 🔒

Version of the message header

MSG_HDR_SIZE 🔒

Message header size

MSG_PAYLOAD_SIZE 🔒

Message payload size

MSG_VERSION 🔒

Version of the message payload

SNP_GUEST_REQ_MAX_DATA_SIZE

Maximum buffer size that the hypervisor takes to store the SEV-SNP certificates

Functions

build_iv 🔒

Build the initialization vector for AES-256 GCM

Type Aliases

SnpGuestRequestExtData

Data page(s) the hypervisor will use to store certificate data in an extended SNP_GUEST_REQUEST