svsm::hyperv

Module hv

Structs

HvHypercallInput

HvHypercallOutput

HvInputEnableVpVtl 🔒

HvInputGetVpRegister 🔒

Dead code is allowed because the hypercall input logic doesn’t recognize that all fields are read when they are copied into the hypercall input page.

HvInputStartVirtualProcessor 🔒

HypercallInput 🔒

A wrapper type to reinterpret an input hypercall page as a header H, optionally followed by a number of instances of T.

HypercallOutput 🔒

A wrapper type to reinterpret an output hypercall page as a sequence of a number of instances of a type T.

HypercallPage

An raw, owned page in shared memory used for Hyper-V hypercalls.

HypercallPagesGuard

A guard that holds an exclusive borrow of the Hyper-V hypercall pages. This type is typically constructed from PerCpu::get_hypercall_pages.

Enums

HvCallCode 🔒

Constants

HV_PARTITION_ID_SELF

HV_STATUS_OPERATION_FAILED

HV_STATUS_SUCCESS

HV_STATUS_TIMEOUT

HV_VP_INDEX_SELF

Statics

CURRENT_VTL 🔒

HYPERV_HYPERCALL_CODE_PAGE 🔒

IS_HYPERV

Functions

enable_vp_vtl_hypercall 🔒

execute_host_hypercall

execute_hypercall⚠

Safety

get_vp_register

hyperv_setup

hyperv_setup_hypercalls 🔒

hyperv_start_cpu

is_hyperv_hypervisor 🔒

setup_hypercall_page

start_vp_hypercall 🔒