svsm

Module insn_decode

Modules

decode 🔒

insn 🔒

opcode 🔒

Structs

DecodedInsnCtx

Represents the context of a decoded instruction, which is used to interpret the instruction. It holds the decoded instruction, its length and various components that are decoded from the instruction bytes.

Instruction

A view of an x86 instruction.

Enums

DecodedInsn

Immediate

An immediate value in an instruction

InsnError

An error that can occur during instruction decoding.

Operand

An operand in an instruction, which might be a register or an immediate.

Register

A register in an instruction

SegRegister

A Segment register in instruction

Constants

MAX_INSN_SIZE

Traits

InsnMachineCtx

This trait provides the necessary context for an instruction decoder to decode instructions based on the current state of the machine that executed them. It abstracts the interfaces through which an instruction decoder can access specific registers and state that may influence the decoding from the machine (such as a CPU or VMM).

InsnMachineMem

Trait representing a machine memory for instruction decoding.