svsm::locking

Module rwlock

Modules

tests 🔒

Structs

RawRWLock

A simple Read-Write Lock (RWLock) that allows multiple readers or one exclusive writer.

RawReadLockGuard

A guard that provides read access to the data protected by RWLock

RawWriteLockGuard

A guard that provides exclusive write access to the data protected by RWLock

Functions

compose_val 🔒

Composes a 64-bit value by combining the number of readers (low 32 bits) and writers (high 32 bits). This function is used to create a 64-bit synchronization value that represents the current state of the RWLock, including the count of readers and writers.

split_val 🔒

Splits a 64-bit value into two parts: readers (low 32 bits) and writers (high 32 bits).

Type Aliases

RWLock

RWLockAnyTpr

RWLockIrqSafe

RWLockTpr

ReadLockGuard

ReadLockGuardAnyTpr

ReadLockGuardIrqSafe

WriteLockGuard

WriteLockGuardAnyTpr

WriteLockGuardIrqSafe