Module svsm::mm

Re-exports

pub use guestmem::GuestPtr;

pub use memory::valid_phys_address;

pub use memory::writable_phys_addr;

pub use pagetable::PageTablePart;

pub use alloc::allocate_file_page;

pub use alloc::PageRef;

pub use mappings::mmap_kernel;

pub use mappings::mmap_user;

pub use mappings::munmap_kernel;

pub use mappings::munmap_user;

pub use mappings::VMMappingGuard;

pub use address_space::*;

pub use ptguards::*;

Modules

address_space

alloc

guestmem

mappings

memory

page_visibility

pagebox 🔒

pagetable

ptguards

validate

virtualrange

vm

Structs

PageBox

An abstraction, similar to a Box, for types that need to be allocated using page allocator directly.