Module svsm::mm::alloc

Structs

AllocatedInfo 🔒

Struct representing information about an allocated memory page.

CompoundInfo 🔒

Struct representing information about a compound memory page.

FileInfo 🔒

Struct representing information about a file memory page.

FreeInfo 🔒

Struct representing information about a free memory page.

MemInfo

Represents info about allocated and free pages in different orders.

MemoryRegion 🔒

Memory region with its physical/virtual addresses, page count, as well as other details.

PageRef

Represents a reference to a memory page, holding both virtual and physical addresses.

PageStorageType 🔒

Storage type of a memory page, including encoding and decoding methods

ReservedInfo 🔒

Struct representing information about a reserved memory page.

Slab 🔒

Represents a slab allocator for fixed-size objects.

SlabCommon 🔒

Represents common information shared among multiple slab pages.

SlabPage 🔒

Represents a slab memory page, used for efficient allocation of fixed-size objects.

SlabPageInfo 🔒

Struct representing information about a slab memory page.

SlabPageSlab 🔒

Represents a slab page for the SlabPageSlab allocator.

SvsmAllocator

Represents a simple virtual-to-physical memory allocator (SvsmAllocator) implementing the GlobalAlloc trait.

Enums

AllocError

Represents possible errors that can occur during memory allocation.

PageInfo 🔒

Enum representing different types of page information.

PageType 🔒

Enum representing the type of a memory page.

Constants

MAX_ORDER

Maximum order of page allocations (up to 128kb)

MIN_ALIGN

SLAB_PAGE_SIZE 🔒

Statics

ALLOCATOR 🔒

ROOT_MEM 🔒

Static spinlock-protected instance of MemoryRegion representing the root memory region.

SLAB_PAGE_SLAB 🔒

Static spinlock-protected instance of SlabPageSlab representing the slab page allocator.

Functions

allocate_file_page

Allocate a file page.

allocate_page

Allocates a single memory page from the root memory region.

allocate_pages

Allocates multiple memory pages with a specified order from the root memory region.

allocate_slab_page

Allocate a slab page.

allocate_zeroed_page

Allocate a zeroed page.

free_page

Free the page at the given virtual address.

get_file_page 🔒

get_order

Calculates the order of a given size for page allocation.

layout_from_ptr

layout_from_size

memory_info

Retrieve information about the root memory

print_memory_info

Prints memory information based on the provided MemInfo structure.

put_file_page 🔒

root_mem_init

Initializes the root memory region with the specified physical start address, virtual start address, and page count.