svsm::mm

Module memory

Constants

ISA_RANGE_END 🔒

The ending address of the ISA range.

ISA_RANGE_START 🔒

The starting address of the ISA range.

Statics

MEMORY_MAP 🔒

Global memory map containing various memory regions.

Functions

init_memory_map

Initializes the global memory map based on the provided configuration and kernel launch information.

valid_phys_address

Returns true if the provided physical address paddr is valid, i.e. it is within the configured memory regions, otherwise returns false.

valid_phys_region

Returns true if the provided physical region region is valid, i.e., it is within a configured memory region, otherwise returns false. Note this does NOT permit a region to span multiple MEMORY_MAP entries since they are assumed to be coalesced.

writable_phys_addr

Returns true if the provided physical address paddr is writable, otherwise returns false.

write_guest_memory_map