Module svsm::mm::pagetable

Structs

MemAccessMode

Flags to represent how memory is accessed, e.g. write data to the memory or fetch code from the memory.

PTEntry

Represents a page table entry.

PTEntryFlags

PTPage

A pagetable page with multiple entries.

PTWalkAttr

Attributes to determin Whether a memory access (write/fetch) is permitted by a translation which includes the paging-mode modifiers in CR0, CR4 and EFER; EFLAGS.AC; and the supervisor/user mode access.

PageTable

Page table structure containing a root page with multiple entries.

PageTablePart

Sub-tree of a page table that can be populated at the top-level used for virtual memory management

RawPageTablePart 🔒

Represents a sub-tree of a page-table which can be mapped at a top-level index

Enums

Mapping

Mapping levels of page table entries.

PagingMode

Represents paging mode.

Constants

ENTRY_COUNT 🔒

Number of entries in a page table (4KB/8B).

LAUNCH_VMSA_ADDR

Physical address for the Launch VMSA (Virtual Machine Saving Area).

Statics

FEATURE_MASK 🔒

Feature mask for page table entry flags.

MAX_PHYS_ADDR 🔒

Maximum physical address supported by the system.

PHYS_ADDR_SIZE 🔒

Maximum physical address bits supported by the system.

PRIVATE_PTE_MASK 🔒

Mask for private page table entry.

SHARED_PTE_MASK 🔒

Mask for shared page table entry.

Functions

init_encrypt_mask 🔒

Initializes the encrypt mask.

make_private_address 🔒

Set address as private via mask.

make_shared_address 🔒

Set address as shared via mask.

max_phys_addr

Returns the exclusive end of the physical address space.

paging_init

Initializes paging settings.

paging_init_early

Re-initializes early paging settings.

private_pte_mask 🔒

Returns the private encrypt mask value.

shared_pte_mask 🔒

Returns the shared encrypt mask value.

strip_confidentiality_bits 🔒

supported_flags 🔒

Returns the supported flags considering the feature mask.