svsm::mm::pagetable

Struct PageTable

#[repr(C)]
pub struct PageTable {
    root: PTPage,
}

Page table structure containing a root page with multiple entries.

Fields
root: PTPage
Implementations

impl PageTable

pub unsafe fn load(&self)

Load the current page table into the CR3 register.

Safety

The caller must ensure to take other actions to make sure a memory safe execution state is warranted (e.g. changing the stack and register state)

pub fn cr3_value(&self) -> PhysAddr

Get the CR3 register value for the current page table.

pub fn allocate_new() -> Result<PageBox<Self>, SvsmError>

Allocate a new page table root.

Errors

Returns SvsmError if the page cannot be allocated.

pub fn clone_shared(&self) -> Result<PageBox<PageTable>, SvsmError>

Clone the shared part of the page table; excluding the private parts.

Errors

Returns SvsmError if the page cannot be allocated.

pub fn copy_entry(&mut self, other: &Self, entry: usize)

Copy an entry entry from another PageTable.

pub fn index<const L: usize>(vaddr: VirtAddr) -> usize

Computes the index within a page table at the given level for a virtual address vaddr.

Parameters

vaddr: The virtual address to compute the index for.

Returns

The index within the page table.

fn walk_addr_lvl0(page: &mut PTPage, vaddr: VirtAddr) -> Mapping<'_>

Walks a page table at level 0 to find a mapping.

Parameters

page: A mutable reference to the root page table.

vaddr: The virtual address to find a mapping for.

Returns

A Mapping representing the found mapping.

fn walk_addr_lvl1(page: &mut PTPage, vaddr: VirtAddr) -> Mapping<'_>

Walks a page table at level 1 to find a mapping.

Parameters

page: A mutable reference to the root page table.

vaddr: The virtual address to find a mapping for.

Returns

A Mapping representing the found mapping.

fn walk_addr_lvl2(page: &mut PTPage, vaddr: VirtAddr) -> Mapping<'_>

Walks a page table at level 2 to find a mapping.

Parameters

page: A mutable reference to the root page table.

vaddr: The virtual address to find a mapping for.

Returns

A Mapping representing the found mapping.

fn walk_addr_lvl3(page: &mut PTPage, vaddr: VirtAddr) -> Mapping<'_>

Walks the page table to find a mapping for a given virtual address.

Parameters

page: A mutable reference to the root page table.

vaddr: The virtual address to find a mapping for.

Returns

A Mapping representing the found mapping.

fn walk_addr(&mut self, vaddr: VirtAddr) -> Mapping<'_>

Walk the virtual address and return the corresponding mapping.

Parameters

vaddr: The virtual address to find a mapping for.

Returns

A Mapping representing the found mapping.

fn get_pte_address(vaddr: VirtAddr) -> VirtAddr

Calculate the virtual address of a PTE in the self-map, which maps a specified virtual address.

Parameters

`vaddr’: The virtual address whose PTE should be located.

Returns

The virtual address of the PTE.

pub fn virt_to_frame(vaddr: VirtAddr) -> Option<PageFrame>

Perform a virtual to physical translation using the self-map.

Parameters

`vaddr’: The virtual address to translate.

Returns

Some(PageFrame) if the virtual address is valid. None if the virtual address is not valid.

fn alloc_pte_lvl3( entry: &mut PTEntry, vaddr: VirtAddr, size: PageSize, ) -> Mapping<'_>

fn alloc_pte_lvl2( entry: &mut PTEntry, vaddr: VirtAddr, size: PageSize, ) -> Mapping<'_>

fn alloc_pte_lvl1( entry: &mut PTEntry, vaddr: VirtAddr, size: PageSize, ) -> Mapping<'_>

fn alloc_pte_4k(&mut self, vaddr: VirtAddr) -> Mapping<'_>

Allocates a 4KB page table entry for a given virtual address.

Parameters

vaddr: The virtual address for which to allocate the PTE.

Returns

A Mapping representing the allocated or existing PTE for the address.

fn alloc_pte_2m(&mut self, vaddr: VirtAddr) -> Mapping<'_>

Allocates a 2MB page table entry for a given virtual address.

Parameters

vaddr: The virtual address for which to allocate the PTE.

Returns

A Mapping representing the allocated or existing PTE for the address.

fn do_split_4k(entry: &mut PTEntry) -> Result<(), SvsmError>

Splits a 2MB page into 4KB pages.

Parameters

entry: The 2M page table entry to split.

Returns

A result indicating success or an error SvsmError in failure.

fn split_4k(mapping: Mapping<'_>) -> Result<(), SvsmError>

Splits a page into 4KB pages if it is part of a larger mapping.

Parameters

mapping: The mapping to split.

Returns

A result indicating success or an error SvsmError.

fn make_pte_shared(entry: &mut PTEntry)

fn make_pte_private(entry: &mut PTEntry)

pub fn set_shared_4k(&mut self, vaddr: VirtAddr) -> Result<(), SvsmError>

Sets the shared state for a 4KB page.

Parameters

vaddr: The virtual address of the page.

Returns

A result indicating success or an error SvsmError if the operation fails.

pub fn set_encrypted_4k(&mut self, vaddr: VirtAddr) -> Result<(), SvsmError>

Sets the encryption state for a 4KB page.

Parameters

vaddr: The virtual address of the page.

Returns

A result indicating success or an error SvsmError.

pub fn check_mapping(&mut self, vaddr: VirtAddr) -> Option<PhysAddr>

Gets the physical address for a mapped vaddr or None if no such mapping exists.

pub fn map_2m( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: PTEntryFlags, ) -> Result<(), SvsmError>

Maps a 2MB page.

Parameters

vaddr: The virtual address to map.

paddr: The physical address to map to.

flags: The flags to apply to the mapping.

Returns

A result indicating success or failure (SvsmError).

Panics

Panics if either vaddr or paddr is not aligned to a 2MB boundary.

pub fn unmap_2m(&mut self, vaddr: VirtAddr)

Unmaps a 2MB page.

Parameters

vaddr: The virtual address of the mapping to unmap.

Panics

Panics if vaddr is not aligned to a 2MB boundary.

pub fn map_4k( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: PTEntryFlags, ) -> Result<(), SvsmError>

Maps a 4KB page.

Parameters

vaddr: The virtual address to map.

paddr: The physical address to map to.

flags: The flags to apply to the mapping.

Returns

A result indicating success or failure (SvsmError).

pub fn unmap_4k(&mut self, vaddr: VirtAddr)

Unmaps a 4KB page.

Parameters

vaddr: The virtual address of the mapping to unmap.

pub fn phys_addr(&mut self, vaddr: VirtAddr) -> Result<PhysAddr, SvsmError>

Retrieves the physical address of a mapping.

Parameters

vaddr: The virtual address to query.

Returns

The physical address of the mapping if present; otherwise, an error (SvsmError).

pub fn map_region_4k( &mut self, vregion: MemoryRegion<VirtAddr>, phys: PhysAddr, flags: PTEntryFlags, shared: bool, ) -> Result<(), SvsmError>

Maps a region of memory using 4KB pages.

Parameters

vregion: The virtual memory region to map.

phys: The starting physical address to map to.

flags: The flags to apply to the mapping.

shared: Indicates whether the mapping is shared.

Returns

A result indicating success or failure (SvsmError).

pub fn unmap_region_4k(&mut self, vregion: MemoryRegion<VirtAddr>)

Unmaps a region of memory using 4KB pages.

Parameters

vregion: The virtual memory region to unmap.

pub fn map_region_2m( &mut self, vregion: MemoryRegion<VirtAddr>, phys: PhysAddr, flags: PTEntryFlags, shared: bool, ) -> Result<(), SvsmError>

Maps a region of memory using 2MB pages.

Parameters

vregion: The virtual memory region to map.

phys: The starting physical address to map to.

flags: The flags to apply to the mapping.

shared: Indicates whether the mapping is shared.

Returns

A result indicating success or failure (SvsmError).

pub fn unmap_region_2m(&mut self, vregion: MemoryRegion<VirtAddr>)

Unmaps a region vregion of 2MB pages. The region must be 2MB-aligned and correspond to a set of huge mappings.

pub fn map_region( &mut self, region: MemoryRegion<VirtAddr>, phys: PhysAddr, flags: PTEntryFlags, ) -> Result<(), SvsmError>

Maps a memory region to physical memory with specified flags.

Parameters

region: The virtual memory region to map.

phys: The starting physical address to map to.

flags: The flags to apply to the page table entries.

Returns

A result indicating success (Ok) or failure (Err).

pub fn unmap_region(&mut self, vregion: MemoryRegion<VirtAddr>)

Unmaps the virtual memory region vregion.

pub fn populate_pgtbl_part(&mut self, part: &PageTablePart)

Populates this paghe table with the contents of the given subtree in part.

pub unsafe fn make_region_ro_4k( &mut self, region: MemoryRegion<VirtAddr>, ) -> Result<(), SvsmError>

Makes the memory region pages read-only. This method is meant for global pages only.

Safety

The caller should verify that region can be made read-only, i.e. that no write can happen or that a #PF raised by any tentative write is expected. The caller must also ensure that the region start and size are 4k aligned.