svsm::mm::pagetable

Struct RawPageTablePart

struct RawPageTablePart {
    page: PTPage,
}

Represents a sub-tree of a page-table which can be mapped at a top-level index

Fields
page: PTPage
Implementations

impl RawPageTablePart

fn free_lvl1(page: &PTPage)

Frees a level 1 page table.

fn free_lvl2(page: &PTPage)

Frees a level 2 page table, including all level 1 tables beneath it.

fn free(&self)

Frees the resources associated with this page table part.

fn address(&self) -> PhysAddr

Returns the physical address of this page table part.

fn walk_addr(&mut self, vaddr: VirtAddr) -> Mapping<'_>

Walks the page table at level 3 to find the mapping for a given virtual address.

Parameters

vaddr: The virtual address to find the mapping for.

Returns

The Mapping for the given virtual address.

fn alloc_pte_4k(&mut self, vaddr: VirtAddr) -> Mapping<'_>

Allocates a 4KB page table entry for a given virtual address.

Parameters

vaddr: The virtual address for which to allocate the PTE.

Returns

The Mapping representing the allocated or existing PTE for the address.

Panics

Panics if a level 3 mapping is attempted in a RawPageTablePart.

fn alloc_pte_2m(&mut self, vaddr: VirtAddr) -> Mapping<'_>

Allocates a 2MB page table entry for a given virtual address.

Parameters

vaddr: The virtual address for which to allocate the PTE.

Returns

The Mapping representing the allocated or existing PTE for the address.

fn map_4k( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: PTEntryFlags, shared: bool, ) -> Result<(), SvsmError>

Maps a 4KB page.

Parameters

vaddr: The virtual address to map.

paddr: The physical address to map to.

flags: The flags to apply to the mapping.

shared: Indicates whether the mapping is shared.

Returns

A result indicating success (Ok) or failure (Err).

fn unmap_4k(&mut self, vaddr: VirtAddr) -> Option<PTEntry>

Unmaps a 4KB page.

Parameters

vaddr: The virtual address of the mapping to unmap.

Returns

An optional PTEntry representing the unmapped page table entry.

fn map_2m( &mut self, vaddr: VirtAddr, paddr: PhysAddr, flags: PTEntryFlags, shared: bool, ) -> Result<(), SvsmError>

Maps a 2MB page.

Parameters

vaddr: The virtual address to map.

paddr: The physical address to map to.

flags: The flags to apply to the mapping.

shared: Indicates whether the mapping is shared

Returns

A result indicating success (Ok) or failure (Err).

Panics

Panics if vaddr or paddr are not 2MB-aligned

fn unmap_2m(&mut self, vaddr: VirtAddr) -> Option<PTEntry>

Unmaps a 2MB page.

Parameters

vaddr: The virtual address of the mapping to unmap.

Returns

An optional PTEntry representing the unmapped page table entry.

Panics

Panics if vaddr is not memory aligned.