Module svsm::mm::vm

Modules

mapping 🔒

range 🔒

Structs

Mapping

RawAllocMapping

Contains base functionality for all VirtualMapping types which use self-allocated PageFile pages.

VMFileMapping

Map view of a ramfs file into virtual memory

VMFileMappingFlags

VMKernelStack

Mapping to be used as a kernel stack. This maps a stack including guard pages at the top and bottom.

VMM

A single mapping of virtual memory in a virtual memory range

VMMAdapter

VMPhysMem

Map physically contiguous memory

VMR

Virtual Memory Region

VMRMapping

VMReserved

Reserve a region of address space so that no other mapping will be established there. The map function for this type will always return None.

VMalloc

Virtual mapping backed by allocated pages. This can be used for memory allocation if there is no need for the memory to be physically contiguous.

Constants

VMR_GRANULE

Granularity of ranges mapped by [struct VMR]. The mapped region of a [struct VMR] is always a multiple of this constant. One VMR_GRANULE covers one top-level page-table entry on x86-64 with 4-level paging.

Traits

VirtualMapping