svsm::platform

Trait SvsmPlatform

pub trait SvsmPlatform: Sync {

    // Required methods
    fn idle_halt(&self, _irq_guard: &IrqGuard);
    fn env_setup(
        &mut self,
        debug_serial_port: u16,
        vtom: usize,
    ) -> Result<(), SvsmError>;
    fn env_setup_late(
        &mut self,
        debug_serial_port: u16,
    ) -> Result<(), SvsmError>;
    fn env_setup_svsm(&self) -> Result<(), SvsmError>;
    fn setup_percpu(&self, cpu: &PerCpu) -> Result<(), SvsmError>;
    fn setup_percpu_current(&self, cpu: &PerCpu) -> Result<(), SvsmError>;
    fn get_page_encryption_masks(&self) -> PageEncryptionMasks;
    fn capabilities(&self) -> Caps;
    unsafe fn hypercall(
        &self,
        input_control: HvHypercallInput,
        hypercall_pages: &HypercallPagesGuard<'_>,
    ) -> HvHypercallOutput;
    fn cpuid(&self, eax: u32, ecx: u32) -> Option<CpuidResult>;
    unsafe fn write_host_msr(&self, msr: u32, value: u64);
    fn get_io_port(&self) -> &'static dyn IOPort;
    fn page_state_change(
        &self,
        region: MemoryRegion<PhysAddr>,
        size: PageSize,
        op: PageStateChangeOp,
    ) -> Result<(), SvsmError>;
    unsafe fn validate_physical_page_range(
        &self,
        region: MemoryRegion<PhysAddr>,
        op: PageValidateOp,
    ) -> Result<(), SvsmError>;
    unsafe fn validate_virtual_page_range(
        &self,
        region: MemoryRegion<VirtAddr>,
        op: PageValidateOp,
    ) -> Result<(), SvsmError>;
    fn configure_alternate_injection(
        &mut self,
        alt_inj_requested: bool,
    ) -> Result<(), SvsmError>;
    fn change_apic_registration_state(
        &self,
        incr: bool,
    ) -> Result<bool, SvsmError>;
    fn query_apic_registration_state(&self) -> bool;
    fn use_interrupts(&self) -> bool;
    fn is_external_interrupt(&self, vector: usize) -> bool;
    fn start_cpu(&self, cpu: &PerCpu, start_rip: u64) -> Result<(), SvsmError>;
    unsafe fn mmio_write(
        &self,
        vaddr: VirtAddr,
        data: &[u8],
    ) -> Result<(), SvsmError>;
    unsafe fn mmio_read(
        &self,
        vaddr: VirtAddr,
        data: &mut [MaybeUninit<u8>],
    ) -> Result<(), SvsmError>;
    fn terminate() -> !
       where Self: Sized;

    // Provided methods
    fn halt()
       where Self: Sized { ... }
    fn prepare_fw(
        &self,
        _config: &SvsmConfig<'_>,
        _kernel_region: MemoryRegion<PhysAddr>,
    ) -> Result<(), SvsmError> { ... }
    fn launch_fw(&self, _config: &SvsmConfig<'_>) -> Result<(), SvsmError> { ... }
    fn determine_cet_support(&self) -> bool { ... }
    fn setup_hyperv_hypercalls(&self) -> Result<(), SvsmError> { ... }
    fn setup_guest_host_comm(&mut self, _cpu: &PerCpu, _is_bsp: bool) { ... }
    unsafe fn validate_low_memory(&self, addr: u64) -> Result<(), SvsmError> { ... }
    fn flush_tlb(&self, flush_scope: &TlbFlushScope) { ... }
    fn start_svsm_request_loop(&self) -> bool { ... }

}

This defines a platform abstraction to permit the SVSM to run on different underlying architectures.

Required Methods

fn idle_halt(&self, _irq_guard: &IrqGuard)

Halts the system with interrupts enabled as required by the platform.

Arguments

_irq_guard: an IRQ guard structure. This is not actually used, but serves as proof that interrupts have been correctly disabled so that interrupt state can be correctly manipulated as required by the idle halt action.

fn env_setup( &mut self, debug_serial_port: u16, vtom: usize, ) -> Result<(), SvsmError>

Performs basic early initialization of the runtime environment.

fn env_setup_late(&mut self, debug_serial_port: u16) -> Result<(), SvsmError>

Performs initialization of the platform runtime environment after the core system environment has been initialized.

fn env_setup_svsm(&self) -> Result<(), SvsmError>

Performs initialiation of the environment specfic to the SVSM kernel (for services not used by stage2).

fn setup_percpu(&self, cpu: &PerCpu) -> Result<(), SvsmError>

Completes initialization of a per-CPU object during construction.

fn setup_percpu_current(&self, cpu: &PerCpu) -> Result<(), SvsmError>

Completes initialization of a per-CPU object on the target CPU.

fn get_page_encryption_masks(&self) -> PageEncryptionMasks

Determines the paging encryption masks for the current architecture.

fn capabilities(&self) -> Caps

Get the features and the capabilities of the platform.

unsafe fn hypercall( &self, input_control: HvHypercallInput, hypercall_pages: &HypercallPagesGuard<'_>, ) -> HvHypercallOutput

Perform a hypercall.

Safety

Hypercalls may have side-effects that affect the integrity of the system, and the caller must take responsibility for ensuring that the hypercall operation is safe.

fn cpuid(&self, eax: u32, ecx: u32) -> Option<CpuidResult>

Obtain CPUID using platform-specific tables.

unsafe fn write_host_msr(&self, msr: u32, value: u64)

Write a host-owned MSR.

Safety

The caller must ensure that the requested MSR modification does mot affect memory safety.

fn get_io_port(&self) -> &'static dyn IOPort

Obtains a reference to an I/O port implemetation appropriate to the platform.

fn page_state_change( &self, region: MemoryRegion<PhysAddr>, size: PageSize, op: PageStateChangeOp, ) -> Result<(), SvsmError>

Performs a page state change between private and shared states.

unsafe fn validate_physical_page_range( &self, region: MemoryRegion<PhysAddr>, op: PageValidateOp, ) -> Result<(), SvsmError>

Marks a physical range of pages as valid or invalid for use as private pages. Not usable in stage2.

Safety

See the safety considerations for SvsmPlatform::validate_virtual_page_range.

unsafe fn validate_virtual_page_range( &self, region: MemoryRegion<VirtAddr>, op: PageValidateOp, ) -> Result<(), SvsmError>

Marks a virtual range of pages as valid or invalid for use as private pages. Provided primarily for use in stage2 where validation by physical address cannot be supported.

Safety

The caller is required to ensure the safety of the validation operation on this memory range.

fn configure_alternate_injection( &mut self, alt_inj_requested: bool, ) -> Result<(), SvsmError>

Configures the use of alternate injection as requested.

fn change_apic_registration_state(&self, incr: bool) -> Result<bool, SvsmError>

Changes the state of APIC registration on this system, returning either the current registration state or an error.

fn query_apic_registration_state(&self) -> bool

Queries the state of APIC registration on this system.

fn use_interrupts(&self) -> bool

Determines whether the platform supports interrupts to the SVSM.

fn is_external_interrupt(&self, vector: usize) -> bool

Determines whether a given interrupt vector was invoked as an external interrupt.

fn start_cpu(&self, cpu: &PerCpu, start_rip: u64) -> Result<(), SvsmError>

Start an additional processor.

unsafe fn mmio_write( &self, vaddr: VirtAddr, data: &[u8], ) -> Result<(), SvsmError>

Perfrom a write to a memory-mapped IO area

Safety

Caller must ensure that vaddr points to a properly aligned memory location and the memory accessed is part of a valid MMIO range.

unsafe fn mmio_read( &self, vaddr: VirtAddr, data: &mut [MaybeUninit<u8>], ) -> Result<(), SvsmError>

Perfrom a read from a memory-mapped IO area

Safety

Caller must ensure that vaddr points to a properly aligned memory location and the memory accessed is part of a valid MMIO range.

fn terminate() -> !
where Self: Sized,

Terminates the guest.

Provided Methods

fn halt()
where Self: Sized,

Halts the system required by the platform. Interrupt state is under control of the caller.

fn prepare_fw( &self, _config: &SvsmConfig<'_>, _kernel_region: MemoryRegion<PhysAddr>, ) -> Result<(), SvsmError>

Performs the necessary preparations for launching guest boot firmware.

fn launch_fw(&self, _config: &SvsmConfig<'_>) -> Result<(), SvsmError>

Launches guest boot firmware.

fn determine_cet_support(&self) -> bool

Determine whether shadow stacks are supported.

fn setup_hyperv_hypercalls(&self) -> Result<(), SvsmError>

Enable platform-specific Hyper-V hypercall operations.

fn setup_guest_host_comm(&mut self, _cpu: &PerCpu, _is_bsp: bool)

Establishes state required for guest/host communication.

unsafe fn validate_low_memory(&self, addr: u64) -> Result<(), SvsmError>

Validates low memory below the specified physical address, with the exception of addresses reserved for use by the platform object which may pre-validated. Intended only for use during early boot.

Safety

The caller is required to ensure that it is safe to validate low memory.

fn flush_tlb(&self, flush_scope: &TlbFlushScope)

Performs a system-wide TLB flush.

fn start_svsm_request_loop(&self) -> bool

Indicates whether this platform should invoke the SVSM request loop.