Trait svsm::platform::SvsmPlatform

pub trait SvsmPlatform {
    // Required methods
    fn env_setup(&mut self);
    fn env_setup_late(&mut self);
    fn setup_percpu(&self, cpu: &mut PerCpu) -> Result<(), SvsmError>;
    fn setup_percpu_current(&self, cpu: &mut PerCpu) -> Result<(), SvsmError>;
    fn get_page_encryption_masks(&self, vtom: usize) -> PageEncryptionMasks;
    fn setup_guest_host_comm(&mut self, cpu: &mut PerCpu, is_bsp: bool);
    fn get_console_io_port(&self) -> &'static dyn IOPort;
    fn page_state_change(
        &self,
        start: PhysAddr,
        end: PhysAddr,
        size: PageSize,
        make_private: bool
    ) -> Result<(), SvsmError>;
    fn pvalidate_range(
        &self,
        region: MemoryRegion<VirtAddr>,
        valid: bool
    ) -> Result<(), SvsmError>;
}

This defines a platform abstraction to permit the SVSM to run on different underlying architectures.

Required Methods

fn env_setup(&mut self)

Performs basic early initialization of the runtime environment.

fn env_setup_late(&mut self)

Performs initialization of the platform runtime environment after console logging has been initialized.

fn setup_percpu(&self, cpu: &mut PerCpu) -> Result<(), SvsmError>

Completes initialization of a per-CPU object during construction.

fn setup_percpu_current(&self, cpu: &mut PerCpu) -> Result<(), SvsmError>

Completes initialization of a per-CPU object on the target CPU.

fn get_page_encryption_masks(&self, vtom: usize) -> PageEncryptionMasks

Determines the paging encryption masks for the current architecture.

fn setup_guest_host_comm(&mut self, cpu: &mut PerCpu, is_bsp: bool)

Establishes state required for guest/host communication.

fn get_console_io_port(&self) -> &'static dyn IOPort

Obtains a console I/O port reference.

fn page_state_change( &self, start: PhysAddr, end: PhysAddr, size: PageSize, make_private: bool ) -> Result<(), SvsmError>

Performs a page state change between private and shared states.

fn pvalidate_range( &self, region: MemoryRegion<VirtAddr>, valid: bool ) -> Result<(), SvsmError>

Marks a page as valid or invalid as a private page.