svsm::task

Module schedule

Round-Robin scheduler implementation for COCONUT-SVSM

This module implements a round-robin scheduler for cooperative multi-tasking. It works by assigning a single owner for each struct Task. The owner depends on the state of the task:

RUNNING A task in running state is owned by the RunQueue and either stored in the run_list (when the task is not actively running) or in current_task when it is scheduled on the CPU.

BLOCKED A task in this state is waiting for an event to become runnable again. It is owned by a wait object when in this state.

TERMINATED The task is about to be destroyed and owned by the RunQueue.

The scheduler is cooperative. A task runs until it voluntarily calls the schedule() function.

Only when a task is in RUNNING or TERMINATED state it is assigned to a specific CPU. Tasks in the BLOCKED state have no CPU assigned and will run on the CPU where their event is triggered that makes them RUNNING again.

Structs

RunQueue

A RunQueue implementation that uses an RBTree to efficiently sort the priority of tasks within the queue.

SetAffinityMessage 🔒

TaskList

Global task list This contains every task regardless of affinity or run state.

Constants

CONTEXT_SWITCH_RESTORE_TOKEN 🔒

The location of a cpu-local shadow stack restore token that’s mapped into every set of page tables for use during context switches.

Statics

TASKLIST

Functions

after_task_switch

create_user_task

Creates and initializes the kernel state of a new user task. The task is not added to the TASKLIST or run-queue yet.

current_task

current_task_terminated

Terminates the current task.

enqueue_task 🔒

finish_user_task

Finished user-space task creation by putting the task on the global TASKLIST and adding it to the run-queue.

go_idle

is_current_task

Check to see if the task scheduled on the current processor has the given id

preemption_checks 🔒

schedule

Perform a task switch and hand the CPU over to the next task on the run-list. In case the current task is terminated, it will be destroyed after the switch to the next task.

schedule_init⚠

Initializes the RunQueue on the current CPU. It will switch to the idle task and initialize the current_task field of the RunQueue. After this function has ran it is safe to call schedule() on the current CPU.

schedule_task

scheduler_idle

Enters an idle state if there is no task that can run.

set_affinity

start_kernel_task

Creates, initializes and starts a new kernel task. Note that the task has already started to run before this function returns.

start_kernel_thread

Creates, initializes and starts a new kernel thread of the currently running kernel task. Note that the thread has already started to run before this function returns.

switch_to 🔒 ⚠

task_pointer 🔒

terminate