Module svsm::utils::immut_after_init

Structs

ImmutAfterInitCell

A memory location which is effectively immutable after initalization code has run.

ImmutAfterInitRef

A reference to a memory location which is effectively immutable after initalization code has run. A ImmutAfterInitRef can either get initialized statically at link time or once from initialization code, basically following the protocol of a ImmutAfterInitCell itself:

Enums

ImmutAfterInitError

Statics

MULTI_THREADED 🔒

Functions

immut_after_init_set_multithreaded

Type Aliases

ImmutAfterInitResult