svsm::vtpm

Trait TcgTpmSimulatorInterface

pub trait TcgTpmSimulatorInterface: VtpmProtocolInterface {
    // Required methods
    fn send_tpm_command(
        &self,
        command: &[u8],
        locality: u8,
    ) -> Result<Vec<u8>, SvsmReqError>;
    fn signal_poweron(&mut self, only_reset: bool) -> Result<(), SvsmReqError>;
    fn signal_nvon(&self) -> Result<(), SvsmReqError>;
}

This implements one handler for each TpmPlatformCommand supported by the VTPM Protocol. These handlers are based on the TPM Simulator interface provided by the TPM 2.0 Reference Implementation, but with a few changes to make it more Rust idiomatic.

tpm-20-ref/TPMCmd/Simulator/include/prototypes/Simulator_fp.h

Required Methods

fn send_tpm_command( &self, command: &[u8], locality: u8, ) -> Result<Vec<u8>, SvsmReqError>

Send a command for the TPM to run in a given locality

Arguments

command: Buffer with the command to be sent to the TPM.

locality: TPM locality the TPM command will be executed

Returns

A Result containing the response received from the TPM on success, or an error.

fn signal_poweron(&mut self, only_reset: bool) -> Result<(), SvsmReqError>

Power-on the TPM, which also triggers a reset

Arguments

*`only_reset``: If enabled, it will only reset the vTPM; however, the vtPM has to be powered on previously. Otherwise, it will fail.

fn signal_nvon(&self) -> Result<(), SvsmReqError>

In a system where the NV memory used by the TPM is not within the TPM, the NV may not always be available. This function indicates that NV is available.